==> scripts/scRNAseq/1_integrated_seurat_generation.Rmd <==
---
title: "1_integrated_seurat_generation"
output: rmarkdown::github_document
date: '2022-07-08'
---

```{r setup, include=FALSE, echo=FALSE}
knitr::opts_chunk$set(echo = TRUE, root.dir = "C:/Users/Strix/GitRepos/NOD_spt_pd1")
knitr::opts_knit$set(echo = TRUE, root.dir = "C:/Users/Strix/GitRepos/NOD_spt_pd1")
```

```{r libraries, message = FALSE, warning = FALSE}
require(dplyr, quietly = T) #simple manipulation of data.frames
require(Seurat, quietly = T) #scRNA-seq analysis
require(reshape, quietly = T) #melting and casting data.frames
require(ggplot2, quietly = T) #plotting
require(gridExtra, quietly = T) #display multiple plots
require(ggrepel, quietly = T) #labels for plots
require(tibble, quietly = T) #rownames_to_column function
require(useful, quietly = T) #corner function to peek at data.frames
require(msigdbr, quietly = T) #accessing gene signatures from mSigDB
set.seed(1) #set a seed for reproducibility
```

## Initialize Seurat object

Matrices, genes, barcodes, filtered TCR contig annotations, and consensus TCR annotations for each sample are within the data/ directory named as **TISSUE_TREATMENT**. Initiate the Seurat object by reading in each individual output from cellranger multi using a loop and subsequent merge. Loop through each sample in samples 1) storing Gene Expression (GEX) and feature-barcoding (FB) data; 2) creating list of seurat objects with GEX and FB assay data under the list name of each sample; 3) sample meta data (tissue: pancreas [pan], pancreatic lymph node [pLN], blood [bld] and antibody treatment [isotype: IgG or anti-PD-1: PD1] or spontaneously diabetic mice [Spt]).
```{r initialize Seurat object}
conditions <- list.files('data/') #all samples are within individual folders
conditions

data <- list() #list containing GEX and FB data
seurat_objects <- list() #list containing all the Seurat objects

for (condition in conditions) {
  data[[condition]] <- Read10X(data.dir = paste0('data/',condition), strip.suffix = T)  #read 10X data
  #create Seurat object with gex data
  seurat_objects[[condition]] <- CreateSeuratObject(counts = data[[condition]][['Gene Expression']], 
                                                    project = condition, min.cells = 50, min.features = 200) 
  
  #add FB assay data to object with columns for only the cells (colnames) in original GEX data
  seurat_objects[[condition]][['FB']] <- CreateAssayObject(counts = 
                                        data[[condition]][['Antibody Capture']][,colnames(seurat_objects[[condition]])])
  
  ##add tissue and treatment meta data to Seurat object
  seurat_objects[[condition]]@meta.data$tissue <- strsplit(condition,"_")[[1]][1]
  seurat_objects[[condition]]@meta.data$treatment <- paste(strsplit(condition,"_")[[1]][2])
  seurat_objects[[condition]]@meta.data$pathology <- ifelse(seurat_objects[[condition]]@meta.data$treatment == "IgG", "non-diabetic", "diabetic")
}

# Merge all Seurat objects generating unique cell IDs
seurat_object <- merge(x = seurat_objects[[conditions[1]]], 
                       y = seurat_objects[2:length(seurat_objects)], 
                       add.cell.ids = conditions)

#rename and re-order some levels
new_names <- c("bld" = "blood", "pLN" = "pLN", "pan" = "pancreas")
seurat_object$tissue <- factor(paste(new_names[seurat_object$tissue]))
levels(seurat_object$tissue) <- c("blood","pLN","pancreas")
seurat_object$treatment <- factor(seurat_object$treatment, levels = c("IgG","Spt","PD1"))

seurat_object
```

## Data integration by tissue

Find anchors and integrate the data across the different tissues. Each tissue was run in a separate lane, but each treatment group was hashed within each lane, as differences between treatment groups are subtle, but are the most interesting. Differences between tissue groups are less interesting, so they were split into different lanes, which results in some batch effects. To reduce these batch effects, the Seurat is integrated by tissue. This step takes ~1 hour on a machine with 16gb of memory and 4 cores.
```{r Seurat integration by tissue}
split_seurat <- SplitObject(seurat_object, split.by = "tissue") # Split the seurat object for finding integration anchors

# Normalize the split data by tissue and define variable features for each
split_seurat <- lapply(X = split_seurat, FUN = function(x) {
  x <- NormalizeData(x)
  x <- FindVariableFeatures(x, selection.method = "vst", nfeatures = 2000)
})

seurat_anchors <- FindIntegrationAnchors(object.list = split_seurat, dims = 1:20) # Define anchors and integrate data (warning: takes a long time)
saveRDS(seurat_anchors, file='objects/combined_seurat_anchors.rds') # Save anchors just in case

seurat_integrated <- IntegrateData(anchorset = seurat_anchors, dims = 1:20) # Integrate Seurat data using anchors
DefaultAssay(seurat_integrated) <- "integrated"

saveRDS(seurat_integrated, file='objects/combined_seurat_integrated.rds') # Save integrated Seurat object
```

## Feature barcoding and hashing data to de-identify individual mice

The feature-barcoding (FB) data is normalized and the hashtag data is separated from the rest of the FB data. Demultiplexing of the cell hashtags is done to identify doublets and non-hashed cells that will be removed during the filtering steps. It also allows individual mice to be identified in the data. In this experiment, hashtags 1,2,3 used for PD1 and 2,3,4 were used for Spt and IgG samples. The hashing data is in the format of **Hash1-TotalSeqC**.
```{r hashtag demux of Seurat}
# Hashtags 1,2,3 used for PD1 and 2,3,4 used for Spt and IgG
Idents(seurat_integrated) <- seurat_integrated$treatment
hash123 <- subset(seurat_integrated, idents = "PD1")
hash234 <- subset(seurat_integrated, idents = c("IgG","Spt"))

# Separate the hashtag (HTO) data from other FB antibody counts. 4 hashtags were used in the following format: `Hash1-TotalSeqC`
hash123@assays$HTO <- subset(hash123@assays$FB, features = paste0("Hashtag", 1:3, "-TotalSeqC"))
hash123@assays$HTO@key <- "HTO_" # Change the key to preface the additional metadata columns with 'HTO_' instead of the default 'fb_' from the FB data
hash234@assays$HTO <- subset(hash234@assays$FB, features = paste0("Hashtag", 2:4, "-TotalSeqC"))
hash234@assays$HTO@key <- "HTO_"

hash123 <- NormalizeData(hash123, assay = "HTO", normalization.method = "CLR")
hash234 <- NormalizeData(hash234, assay = "HTO", normalization.method = "CLR")

# Demultiplex the hashing data
hash123 <- MULTIseqDemux(hash123, assay = "HTO", autoThresh = T)
hash234 <- MULTIseqDemux(hash234, assay = "HTO", autoThresh = T)

seurat_integrated <- merge(hash123,hash234) #merge the data again after demux

#subset and normalize the FB data as the only cite-seq antibody used which was for tetramer NRPv7-TotalSeqC
seurat_integrated@assays$FB <- subset(seurat_integrated@assays$FB, features = "NRPv7-TotalSeqC")  
seurat_integrated <- NormalizeData(seurat_integrated, assay = "FB", normalization.method = "CLR")
table(seurat_integrated$MULTI_classification)
table(seurat_integrated$MULTI_ID)

## Add mouse_id based on hashtag data but independent of hashtag # (as PD1 was hashtags 1-3 and Spt,IgG were hashtags 2-4)
# Preview of naming scheme that names mice by treatment_replicate (eg: IgG_1, IgG_2, IgG_3, PD1_1, PD1_2, etc.)
seurat_integrated@meta.data %>% 
    group_by(treatment) %>% 
    arrange(treatment,MULTI_ID) %>%
    mutate(mouse_id = paste0(treatment,'_', 
                              match(paste(treatment, MULTI_ID), unique(paste(treatment, MULTI_ID))))) %>% 
    dplyr::select(treatment, MULTI_ID, mouse_id) %>% 
    unique()

# Generate the new metadata for the mouse_id while keeping the barcode
new_metadata <- seurat_integrated@meta.data %>% 
    rownames_to_column('barcode') %>%
    group_by(treatment) %>% 
    arrange(treatment, MULTI_ID) %>%
    mutate(mouse_id = paste0(treatment,'_', 
                              match(paste(treatment, MULTI_ID), unique(paste(treatment, MULTI_ID))))) %>% 
    dplyr::select(treatment, MULTI_ID, mouse_id, barcode) %>%
    column_to_rownames('barcode') %>%
    dplyr::select(mouse_id)

seurat_integrated <- AddMetaData(seurat_integrated, metadata = new_metadata)
seurat_integrated$mouse_id <- factor(seurat_integrated$mouse_id, levels = c(paste0("IgG_",1:3), paste0("Spt_",1:3), paste0("PD1_",1:3)))
```

## T Cell Receptor (TCR) data

The TCR data can now be added to the seurat metadata by looping through the file corresponding to each individual patient. The data generated is clonotype-centric and is derived from **consensus_annotations.csv** because the data from **filtered_contig_annotations** contains multiple consensus sequences. The barcodes must have the suffix added to match those in the seurat object. The TRA and TRB information must also be transformed from long to wide format. 

For cells with multiple TCR assignments, select TCR by the highest number of UMIs first; then break ties by taking the TCR with the highest number of reads; then pick at random if both equal using dplyr::slice. Note that the raw_clonotype_id is added to the seurat object, but this must be re-calculated to consider all samples individually - this cannot be done by cellranger aggr as it does not consider feature-barcoding (hashtag) data that was used to distinguish individual mice: Clonotypes are PRIVATE to each individual mouse.
```{r add TCR information}
metadata <- data.frame() #initialize an empty dataframe for all of the tcr metadata combined

for (i in 1:length(list.files(path = "data/nod/"))) {
  condition <- list.files(path = "data/nod/")[i]
  
  tcrdata <- read.csv(paste0("data/nod/", condition, "/filtered_contig_annotations.csv"), header = T)
  tcrdata$barcode <- sapply(tcrdata$barcode, function(x) sub("-1", "", x), USE.NAMES = F) #remove suffix of -1 at the end of the barcode
  tcrdata$barcode <- paste(condition, tcrdata$barcode, sep = "_") #add prefix
  tcrdata <- tcrdata %>% dplyr::select(barcode, raw_clonotype_id) %>% unique()
  head(tcrdata) #clonotype-centric dataframe
  
  clonotype_data <- read.csv(paste0("data/nod/", condition, "/consensus_annotations.csv"), header = T)
  
  # Find the TCR beta chain for each cell
  trb <- clonotype_data %>% filter(productive == "true", chain == "TRB") %>% 
    group_by(clonotype_id) %>%  
    arrange(desc(reads), desc(umis)) %>% 
    top_n(1, umis) %>%  #select TCR by number of UMIs
    top_n(1, reads) %>% #break ties by taking TCR with highest number of reads
    dplyr::slice(1) %>% #pick top TCR at random if the above are equal using slice
    dplyr::select(clonotype_id, v_gene, d_gene, j_gene, c_gene, cdr3, cdr3_nt) %>% # remove unnecessary columns
    setNames(paste0("TRB.", names(.))) # append to each column name for TCR beta chain
  colnames(trb)[1] <- "raw_clonotype_id"
  
  #do the same with TRA
  tra <- clonotype_data %>% filter(productive == "true", chain == "TRA") %>% 
    group_by(clonotype_id) %>%  
    arrange(desc(reads), desc(umis)) %>% 
    top_n(1, umis) %>%  #select TCR by number of UMIs
    top_n(1, reads) %>% #break ties by taking TCR with highest number of reads
    dplyr::slice(1) %>%
    dplyr::select(clonotype_id, v_gene, d_gene, j_gene, c_gene, cdr3, cdr3_nt) %>%
    setNames(paste0("TRA.", names(.))) #append to each column name for TCR alpha chain
  colnames(tra)[1] <- "raw_clonotype_id"
  
  clonotype_data <- merge(tra, trb, by = "raw_clonotype_id") #merge the TCR alpha and beta chain data
  tcrdata <- merge(tcrdata, clonotype_data, by = "raw_clonotype_id") #using the clono-centric dataframe, add the TCR consensus data
  metadata <- rbind(metadata, tcrdata) #gather the TCR data for all the samples
}

metadata <- metadata %>% column_to_rownames("barcode") #add the merged dataframe to the Seurat object's metadata.
seurat_integrated <- AddMetaData(object = seurat_integrated, metadata = metadata)

#add metadata for whether a TCR was detected or not
metadata <- seurat_integrated@meta.data %>% 
              mutate(tcr_detected = ifelse(is.na(TRA.cdr3) & is.na(TRB.cdr3), FALSE, TRUE)) %>% 
              dplyr::select(tcr_detected)
seurat_integrated <- AddMetaData(object = seurat_integrated, metadata = metadata)

seurat_integrated@meta.data %>% filter(!is.na(TRA.cdr3) & !is.na(TRB.cdr3)) %>% count() / seurat_integrated@meta.data %>% count() * 100
```

Set the **clonotype_id** based on unique TCR VDJ genes and CDR3 nucleotide sequences for the TCR-alpha and TCR-beta chains for each cell with consideration of individual mice: Clonotypes are not shared across animals, but may be shared across tissues. Clonotypes are enumerated based on descending number of cells, so clonotype1 is the largest clonally expanded clonotype. 

Set the **tcrtype_id** based on the same TCR amino acid sequence only. There may be shared TCRs across animals and treatment groups. **tcrtype_convergence** is a measure of how many individual clonotypes converge into a single tcrtype.
```{r clonotype and tcrtype metadata}
## CLONOTYPE ID
new_metadata <- seurat_integrated@meta.data %>% 
  rownames_to_column('barcode') %>%
  filter(!is.na(raw_clonotype_id)) %>%
  group_by(mouse_id, TRB.v_gene, TRB.d_gene, TRB.j_gene, TRB.cdr3_nt ,TRA.v_gene, TRA.d_gene, TRA.j_gene, TRA.cdr3_nt) %>% 
  mutate(clonotype_count = n()) %>% #count the number of cells in each clonotype
  arrange(desc(clonotype_count)) %>% 
  ungroup() %>% 
  arrange(desc(clonotype_count)) %>% 
  mutate(clonotype_id = paste0('clonotype', ... = match(paste(mouse_id, TRB.v_gene, TRB.d_gene, TRB.j_gene, TRB.cdr3_nt, TRA.v_gene, TRA.d_gene, TRA.j_gene, TRA.cdr3_nt), unique(paste(mouse_id, TRB.v_gene, TRB.d_gene, TRB.j_gene, TRB.cdr3_nt, TRA.v_gene, TRA.d_gene, TRA.j_gene, TRA.cdr3_nt))))) %>%  #uniquely match VDJ for TCRB and TCRA
  mutate(clonotype_count_bin = cut(clonotype_count, 
                                   breaks=c(0,1,5,20,50,100,300), 
                                   labels=c("1","2-4","5-19","20-49","50-99","100-300"))) %>% #bin the clonotype counts
  column_to_rownames('barcode') %>%
  select(clonotype_count, clonotype_id, clonotype_count_bin)
seurat_integrated <- AddMetaData(seurat_integrated, metadata = new_metadata)

## TCRTYPE ID
new_metadata <- seurat_integrated@meta.data[!(is.na(seurat_integrated@meta.data$TRA.cdr3) & 
                              is.na(seurat_integrated@meta.data$TRB.cdr3)),] %>% # excludes TRA.TCR==NA & TRB.TCR==NA
    rownames_to_column('barcode') %>% 
    add_count(TRA.cdr3, TRB.cdr3) %>% arrange(desc(n)) %>% # orders by identical groups with greatest number of cells
    mutate(tcrtype_id = paste0('tcrtype', group_indices(., desc(n), TRA.cdr3, TRB.cdr3))) %>% # adds the new column according to number of cells in that tcrtype group (n)
    add_count(tcrtype_id, name='tcrtype_count') %>%
    mutate(tcrtype_count_bin = cut(tcrtype_count, 
                                   breaks=c(0,1,5,20,50,100,300), 
                                   labels=c("1","2-4","5-19","20-49","50-99","100-300"))) %>%
    column_to_rownames('barcode') %>%
    dplyr::select(tcrtype_id, tcrtype_count, tcrtype_count_bin)
seurat_integrated <- AddMetaData(seurat_integrated, metadata = new_metadata)

## TCRTYPE CONVERGENCE
obj_convergence <- seurat_integrated@meta.data %>% 
                      dplyr::select(tcrtype_id, clonotype_id) %>%
                      unique() %>% arrange(tcrtype_id) %>%
                      group_by(tcrtype_id) %>%
                      count() %>%
                      data.frame()
obj_convergence <- setNames(as.numeric(obj_convergence$n), obj_convergence$tcrtype_id)
new_metadata <- seurat_integrated@meta.data %>% 
                    rownames_to_column('barcode') %>%
                    mutate(tcrtype_convergence = obj_convergence[tcrtype_id]) %>%
                    column_to_rownames('barcode') %>%
                    dplyr::select(tcrtype_convergence) 
seurat_integrated <- AddMetaData(seurat_integrated, metadata = new_metadata)
```

## Clustering for filtering

Run the standard workflow for clustering to define CD4 and CD8 T cells. Cells that cluster together that are not clearly CD4 or CD8 T cells, based on expression of Cd4 and Cd8a, respectively, that do not have TCR data will be removed.
```{r clustering to filter non-T cells}
seurat_integrated[["percent.mt"]] <- PercentageFeatureSet(seurat_integrated, pattern = "^mt-", assay = "RNA") #percent mitochondrial RNA
VlnPlot(seurat_integrated, features = c("nFeature_RNA", "nCount_RNA", "percent.mt"), ncol = 3, pt.size=0.1)
FeatureScatter(seurat_integrated, feature1 = "nCount_RNA", feature2 = "percent.mt", pt.size=0.1) +
  FeatureScatter(seurat_integrated, feature1 = "nCount_RNA", feature2 = "nFeature_RNA", pt.size=0.1)

#the number of cells from each sample before filtering - will add to get a summary of how many cells are filtered
cell_counts <- seurat_integrated@meta.data %>% group_by(orig.ident) %>% count()

seurat_integrated <- subset(seurat_integrated, subset = (nFeature_RNA > 200 & #remove cells with few features
                                                   percent.mt < 10 & #remove cells with too much mitochondrial RNA
                                                   MULTI_ID != 'Doublet' & #remove doublets as determined by hashing
                                                   MULTI_ID != 'Negative')) #remove any cells that never got a hashtag

#add the number of cells after filtering to the data frame
cell_counts <- merge(cell_counts, 
                     seurat_integrated@meta.data %>% 
                       group_by(orig.ident) %>% 
                       count(), by = "orig.ident")
colnames(cell_counts) <- c("sample","before_filtering","after_filtering")
cell_counts <- cell_counts %>% 
                mutate(percent_filtered = round(100*(before_filtering-after_filtering)/before_filtering)) #add percent filtered

cell_counts

#standard workflow for scaling and clustering the integrated Seurat
seurat_integrated <- ScaleData(seurat_integrated, verbose = FALSE) #scale the RNA data
seurat_integrated <- RunPCA(seurat_integrated, npcs = 30, verbose = FALSE, features = rownames(x = seurat_integrated)) #run PCA
seurat_integrated <- RunUMAP(seurat_integrated, reduction = "pca", dims = 1:30)
seurat_integrated <- FindNeighbors(seurat_integrated, reduction = "pca", dims = 1:30)
seurat_integrated <- FindClusters(seurat_integrated, resolution = 7) #very high resolution results in overclustering to start
p_cluster <- DimPlot(seurat_integrated, reduction = "umap", label = TRUE, label.size = 5) + theme(legend.position = "none")
p_markers <- FeaturePlot(seurat_integrated, c("Cd4","Cd8a","Foxp3","Mki67"))
p_cluster + p_markers

#visually identify contaminating clusters based on gene expression and remove them - overclustering helps facilitate this
seurat_integrated@meta.data %>%
  group_by(seurat_clusters) %>% 
  summarize(cd4_percent = 100*sum(Cd4 > 0)/n(), 
            cd8_percent = 100*sum(Cd8a > 0)/n(), 
            cd3_percent = 100*sum(Cd3e > 0)/n()) %>% 
  mutate(to_be_filtered = cd4_percent < 10 & cd8_percent < 10 | cd3_percent < 80) %>% 
  filter(to_be_filtered == TRUE)
contaminants <- c(63,56,42,71,84,81,77,68,44,62,17,20,79,37,51,66,76,82)
seurat_integrated <- subset(seurat_integrated, idents = contaminants, invert = T)

#choose CD8 T cell clusters based on the ratio of Cd8a/Cd4 average expression. The rest are CD4
avgs <- AverageExpression(seurat_integrated, features = c("Cd4", "Cd8a"), assay = "RNA")
cd8_clusters <- as.data.frame(t(avgs$RNA)) %>% 
                  mutate(ratio = Cd8a/Cd4) %>% 
                  rownames_to_column("cluster") %>% 
                  filter(ratio > 1) %>% 
                  filter(cluster != 67) %>% #cluster 67 is CD4
                  pull(cluster) %>% 
                  as.numeric()

#visualize the cells
p_cd8s <- DimPlot(seurat_integrated,cells.highlight = WhichCells(seurat_integrated, idents = cd8_clusters)) + 
            theme(legend.position = "none")
p_cd4expression <- FeaturePlot(seurat_integrated, "Cd4") + theme(legend.position = "none")
p_cd8expression <- FeaturePlot(seurat_integrated, "Cd8a") + theme(legend.position = "none")
grid.arrange(p_cluster, p_cd4expression, p_cd8expression, p_cd8s, nrow=1)

#add the metadata containing the celltype (CD4 or CD8)
new_metadata <- seurat_integrated@meta.data %>% 
                  mutate(celltype = ifelse(seurat_clusters %in% cd8_clusters, "CD8","CD4")) %>%
                  dplyr::select(celltype)
seurat_integrated <- AddMetaData(seurat_integrated, new_metadata)
```

Clonotype analysis is important in this experiment, so clonotype_id is used to help determine whether a cell should be a CD4 or CD8 T cell. A clone is assumed to be either a CD8 or CD4 - it is assumed that it cannot be both. If a clonotype has more of one or the other, then the incorrectly classified cell is re-classified such that all cells in a clonotype are CD4 or CD8. This isn't perfect: ome clonotypes will have the same number of CD4 and CD8s and it cannot be determined whether they should be all CD4 or CD8s. These will later be split into two distinct clonotypes within CD4s and CD8s.
```{r Allow clonotype to determine T cell type}
#first, count the number of CD4 and CD8 T cells in each clonotype and filter to remove clonotypes where the number of CD4=number of CD8
clono_celltypes <- seurat_integrated@meta.data %>% 
                    filter(!is.na(clonotype_id)) %>% #filter NA clonotype group
                    group_by(clonotype_id, celltype) %>% 
                    tally() %>% 
                    cast(clonotype_id ~ celltype) %>% #numbers of CD4 and CD8 are now columns with rows as clonotype_id
                    mutate(CD4 = ifelse(is.na(CD4), 0, CD4), 
                           CD8 = ifelse(is.na(CD8), 0, CD8)) %>% #change NA to zero values
                    filter(CD4 != 0 & CD8 != 0 & CD4 != CD8) #only consider cells with more CD4 or CD8

clono_celltypes$clono_celltype <- colnames(clono_celltypes)[2:3][apply(clono_celltypes[,2:3], 1, which.max)] #reports the celltype with most cells in the clono_celltype column

dim(clono_celltypes)
clono_celltypes <- setNames(clono_celltypes$clono_celltype,clono_celltypes$clonotype_id)

#loop through each celltype in the Seurat to change the celltype based on the largest clonotype
for (i in 1:length(clono_celltypes)) {
  seurat_integrated@meta.data[which(seurat_integrated$clonotype_id == names(clono_celltypes[i])), "celltype"] <- paste(clono_celltypes[i])
}

#visualize results
p_celltypes <- DimPlot(seurat_integrated,group.by = "celltype")
grid.arrange(p_cd4expression, p_cd8expression, p_celltypes, nrow=1)
```

Cluster 12 and 6 are Lyz2+ macrophages, cluster 14 are Cd74+H2-Aa+ DCs, and cluster 9 are likely NKT cells (Xcl1, Cd3e, but low in both Cd4 and Cd8a). Both cluster 10 and cluster 13 (proliferative cluster high in Mki67) contain both CD8 and CD4 T cells. To distinguish the CD4 and CD8 T cells in these shared clusters, the cells will be overclustered and separated based on manual grouping of small clusters.
```{r re-cluster}
seurat_integrated <- ScaleData(seurat_integrated, verbose = FALSE) #scale the RNA data
seurat_integrated <- RunPCA(seurat_integrated, npcs = 30, verbose = FALSE, features = rownames(x = seurat_integrated)) #run PCA
seurat_integrated <- RunUMAP(seurat_integrated, reduction = "pca", dims = 1:30)
seurat_integrated <- FindNeighbors(seurat_integrated, reduction = "pca", dims = 1:30)
seurat_integrated <- FindClusters(seurat_integrated, resolution = 1.2)
p_cluster <- DimPlot(seurat_integrated, reduction = "umap", label = TRUE, label.size = 5, split.by = "celltype") + theme(legend.position = "none")
p_cluster

seurat_markers <- FindAllMarkers(seurat_integrated, min.pct = 0.25, only.pos = T, logfc.threshold = 0.25) %>%
                     group_by(cluster) %>%
                     filter(p_val_adj < 0.05) %>%
                     group_by(cluster) %>%
                     arrange(desc(avg_log2FC), .by_group = T) %>%
                     top_n(n = 50, wt = avg_log2FC)

#identify CD8 and CD4 T cells based on clusters
#cluster 9, 15, 16 are split between CD4 and CD8 (proliferative, acinar-contaminated, and IFN-sensing, respectively)
new_metadata <- seurat_integrated@meta.data %>% 
                  mutate(new_celltype = ifelse(seurat_clusters %in% c(19), "unknown", 
                                               ifelse(seurat_clusters %in% c(0,1,3,4,6,8,10,11,14,17), "CD4", 
                                                      ifelse(seurat_clusters %in% c(2,5,7,12,13), "CD8", paste0(celltype))))) %>%
                  dplyr::select(new_celltype)
colnames(new_metadata) <- "celltype"
seurat_integrated <- AddMetaData(seurat_integrated, new_metadata)

#some cells classified as CD8 T cells are CD4 T cells based on expression of CD4 and lack of CD8, so fix this
#VlnPlot(seurat_integrated,"Cd4") + geom_hline(yintercept = 0.75) + #0.75 is a good cutoff for Cd4
#VlnPlot(seurat_integrated,"Cd8a") + geom_hline(yintercept = 0.75) #0.75 is a good cutoff for Cd8
new_metadata[WhichCells(seurat_integrated, expression = celltype == "CD4" & Cd8a > 1 & Cd4 <= 1),"celltype"] <- "CD8"
new_metadata[WhichCells(seurat_integrated, expression = celltype == "CD8" & Cd4 > 1 & Cd8a <= 1),"celltype"] <- "CD4"
new_metadata[WhichCells(seurat_integrated, expression = celltype == "CD8" & Foxp3 > 0.5),"celltype"] <- "CD4"

#add the metadata where CD4 and CD8s are based on expression of Cd8a and Cd4
seurat_integrated <- AddMetaData(seurat_integrated, new_metadata)
seurat_integrated <- subset(seurat_integrated, subset = celltype %in% c("CD4","CD8"))

#re-plot with new celltype designations
p_cluster <- DimPlot(seurat_integrated, reduction = "umap", label = TRUE, label.size = 5, split.by = "celltype") + theme(legend.position = "none")
p_cluster
```


Note that the clustering looks messy between CD4 and CD8 T cells with some cells in every cluster. This is okay, as the cells will be re-clustered as CD4 and CD8 T cells separately and these new clusters will be re-projected onto the UMAP containing both T cell types.

Now that CD4s and CD8s have been distinguished, they will now be clustered separately into different UMAPs. This is both for visualization and also for better resolution of clusters of interest. These clusters, defined on the separate CD4 and CD8 UMAPS, will be "re-projected" back onto the UMAP containing all the cells. Some clusters will be shared (like an interferon-sensing cluster) while others will be unique to CD4 or CD8.
```{r re-cluster CD4 and CD8 separately}
#start by re-clustering the CD4 using the standard workflow
cd4 <- subset(seurat_integrated, subset = celltype == "CD4")
cd4 <- ScaleData(cd4, verbose = FALSE) #scale the RNA data
cd4 <- RunPCA(cd4, npcs = 30, verbose = FALSE, features = rownames(x = cd4)) #run PCA
cd4 <- RunUMAP(cd4, reduction = "pca", dims = 1:30)
cd4 <- FindNeighbors(cd4, reduction = "pca", dims = 1:30)
cd4 <- FindClusters(cd4, resolution = 0.8)
p4_cluster <- DimPlot(cd4, reduction = "umap", label = TRUE, label.size = 5, split.by = "celltype") + theme(legend.position = "none")
p4_cluster

new_metadata[WhichCells(cd4, idents = 12),"celltype"] <- "CD8" #these cells express Gzma but also CD4. Doublets?

#next re-clustering the CD8 using the standard workflow
cd8 <- subset(seurat_integrated, subset = celltype == "CD8")
cd8 <- ScaleData(cd8, verbose = FALSE) #scale the RNA data
cd8 <- RunPCA(cd8, npcs = 30, verbose = FALSE, features = rownames(x = cd8)) #run PCA
cd8 <- RunUMAP(cd8, reduction = "pca", dims = 1:30)
cd8 <- FindNeighbors(cd8, reduction = "pca", dims = 1:30)
cd8 <- FindClusters(cd8, resolution = 1.1)
p8_cluster <- DimPlot(cd8, reduction = "umap", label = TRUE, label.size = 5, split.by = "celltype") + theme(legend.position = "none")
p8_cluster

new_metadata[WhichCells(cd8, idents = 9),"celltype"] <- "CD4" #these cells express CD200 and look very similar to the large cluster in CD4
seurat_integrated <- AddMetaData(seurat_integrated, new_metadata) #add the corrected celltypes and try again...

#Re-cluster the CD4 using the standard workflow
cd4 <- subset(seurat_integrated, subset = celltype == "CD4")
cd4 <- ScaleData(cd4, verbose = FALSE) #scale the RNA data
cd4 <- RunPCA(cd4, npcs = 30, verbose = FALSE, features = rownames(x = cd4)) #run PCA
cd4 <- RunUMAP(cd4, reduction = "pca", dims = 1:30)
cd4 <- FindNeighbors(cd4, reduction = "pca", dims = 1:30)
cd4 <- FindClusters(cd4, resolution = 0.8)
p4_cluster <- DimPlot(cd4, reduction = "umap", label = TRUE, label.size = 5, split.by = "celltype") + theme(legend.position = "none")
p4_cluster

#determine the markers for all the CD4 clusters
cd4_markers <- FindAllMarkers(cd4, min.pct = 0.25, only.pos = T, logfc.threshold = 0.25) %>%
                     group_by(cluster) %>%
                     filter(p_val_adj < 0.05) %>%
                     group_by(cluster) %>%
                     arrange(desc(avg_log2FC), .by_group = T) %>%
                     top_n(n = 50, wt = avg_log2FC)
saveRDS(cd4_markers, "objects/cd4_seurat_markers.rds")

#merge clusters
cd4 <- RenameIdents(cd4,
                    "0" = "Tcon_mem",
                    "1" = "Treg_resting",
                    "2" = "Treg_eff",
                    "3" = "Tcon_cm",
                    "4" = "Tcon_prog",
                    "5" = "Tcon_effmem",
                    "6" = "Recently_activated",
                    "7" = "Tcon_cm",
                    "8" = "Tcon_Th21",
                    "9" = "Tcon_eff",
                    "10" = "Acinar_contaminated",
                    "11" = "Interferon_sensing",
                    "12" = "contaminant",
                    "13" = "Tcon_Tfh",
                    "14" = "Proliferating",
                    "15" = "contaminant",
                    "16" = "contaminant")

#next re-clustering the CD8 using the standard workflow
cd8 <- subset(seurat_integrated, subset = celltype == "CD8")
cd8 <- ScaleData(cd8, verbose = FALSE) #scale the RNA data
cd8 <- RunPCA(cd8, npcs = 30, verbose = FALSE, features = rownames(x = cd8)) #run PCA
cd8 <- RunUMAP(cd8, reduction = "pca", dims = 1:30)
cd8 <- FindNeighbors(cd8, reduction = "pca", dims = 1:30)
cd8 <- FindClusters(cd8, resolution = 1.1)
p8_cluster <- DimPlot(cd8, reduction = "umap", label = TRUE, label.size = 5, split.by = "celltype") + theme(legend.position = "none")
p8_cluster

#determine the markers for all the CD8 clusters
cd8_markers <- FindAllMarkers(cd8, min.pct = 0.25, only.pos = T, logfc.threshold = 0.25) %>%
                     group_by(cluster) %>%
                     filter(p_val_adj < 0.05) %>%
                     group_by(cluster) %>%
                     arrange(desc(avg_log2FC), .by_group = T) %>%
                     top_n(n = 50, wt = avg_log2FC)

#merge clusters
cd8 <- RenameIdents(cd8,
                    "0" = "CD8_slec",
                    "1" = "CD8_cm",
                    "2" = "CD8_eff",
                    "3" = "CD8_slec",
                    "4" = "CD8_texh",
                    "5" = "CD8_mem",
                    "6" = "CD8_pexh",
                    "7" = "CD8_cm",
                    "8" = "Acinar_contaminated",
                    "9" = "contaminant",
                    "10" = "CD8_slec",
                    "11" = "CD8_cm",
                    "12" = "contaminant",
                    "13" = "Interferon_sensing",
                    "14" = "CD8_effmem",                    
                    "15" = "Recently_activated",                    
                    "16" = "Proliferating",                    
                    "17" = "CD8_slec")

#re-project the newly defined clusters onto the UMAP containing all the cells
new_metadata <- rbind(data.frame(seurat_clusters = Idents(cd4)), data.frame(seurat_clusters = Idents(cd8)))
seurat_integrated <- AddMetaData(seurat_integrated, new_metadata)
Idents(seurat_integrated) <- seurat_integrated$seurat_clusters
seurat_integrated <- subset(seurat_integrated, idents = "contaminant", invert = T)
cluster_levels <- c("CD8_cm", #factor the Seurat clusters to order them by celltype (shared clusters are at the end)
                    "CD8_mem",
                    "CD8_effmem",
                    "CD8_slec",
                    "CD8_eff",
                    "CD8_pexh",
                    "CD8_texh",
                    "Tcon_cm",
                    "Tcon_mem",
                    "Tcon_effmem",
                    "Tcon_eff",
                    "Tcon_prog",
                    "Tcon_Th21",
                    "Tcon_Tfh",
                    "Treg_resting",
                    "Treg_eff",
                    "Acinar_contaminated",
                    "Recently_activated",
                    "Interferon_sensing",
                    "Proliferating")
seurat_integrated$seurat_clusters <- factor(seurat_integrated$seurat_clusters, levels = cluster_levels) #factor the Seurat clusters to order them by celltype (shared clusters are at the end)
Idents(seurat_integrated) <- seurat_integrated$seurat_clusters
```

Plot the genes associated with each cluster group. For example, which genes are conserved within Tcon and CD8 central memory clusters, etc.
```{r identify genes associated with clusters for heatmap}
#find the markers that distinguish all of the clusters
seurat_markers <- FindAllMarkers(seurat_integrated, min.pct = 0.25, only.pos = T, logfc.threshold = 0.25) %>%
                     group_by(cluster) %>%
                     filter(p_val_adj < 0.05) %>%
                     group_by(cluster) %>%
                     arrange(desc(avg_log2FC), .by_group = T)
write.table(seurat_markers,"output/all_clusters_markers.tsv", sep = "\t", quote = F, row.names = F) #save to file
write.table(seurat_markers %>% group_by(cluster) %>% top_n(50, wt = avg_log2FC), "output/all_clusters_top50markers.tsv", sep = "\t", quote = F, row.names = F) #save to file

#identify conserved genes within cluster groups
#central memory for example
seurat_markers %>% filter(cluster %in% c("CD8_cm","Tcon_cm")) %>% group_by(gene) %>% filter(n() > 1) %>% data.frame() #example
celltype_genes <- c("Cd4", "Cd8a", "Cd8b1", "Foxp3", "Ikzf2")
mem_genes <- seurat_markers %>% filter(cluster %in% c("CD8_mem","Tcon_mem")) %>% group_by(gene) %>% filter(n() > 1) %>% pull(gene) %>% unique()
cm_genes <- seurat_markers %>% filter(cluster %in% c("CD8_cm","Tcon_cm")) %>% group_by(gene) %>% filter(n() > 1) %>% pull(gene) %>% unique()
exh_genes <- c("Tox", seurat_markers %>% filter(cluster %in% c("CD8_pexh","CD8_texh")) %>% group_by(gene) %>% filter(n() > 1) %>% pull(gene) %>% unique())
eff_genes <- seurat_markers %>% filter(cluster %in% c("CD8_eff","Tcon_eff","CD8_effmem","Tcon_effmem")) %>% group_by(gene) %>% filter(n() > 1) %>% pull(gene) %>% unique()
slec_genes  <- unique(c("Klrg1","Cx3cr1","Zeb2", seurat_markers %>% filter(cluster == "CD8_slec") %>% top_n(20, wt = avg_log2FC) %>% pull(gene)))
treg_genes <- c(unique(seurat_markers %>% filter(cluster == "Treg_resting") %>% top_n(50, wt = avg_log2FC) %>% pull(gene) %>% unique(), 
                    seurat_markers %>% filter(cluster == "Treg_eff") %>% top_n(50, wt = avg_log2FC) %>% pull(gene) %>% unique()))
tfh_genes <- seurat_markers %>% filter(cluster == "Tcon_Tfh") %>% top_n(50, wt = avg_log2FC) %>% pull(gene) %>% unique()
th21_genes <- seurat_markers %>% filter(cluster == "Tcon_Th21") %>% top_n(50, wt = avg_log2FC) %>% pull(gene) %>% unique()
acinar_genes <- seurat_markers %>% filter(cluster == "Acinar_contaminated") %>% top_n(50, wt = avg_log2FC) %>% pull(gene) %>% unique()
ifn_genes <- seurat_markers %>% filter(cluster == "Interferon_sensing") %>% filter(grepl("Is|If|Stat1", gene)) %>% pull(gene)
activation_genes <- seurat_markers %>% filter(cluster == "Recently_activated") %>% top_n(50, wt = avg_log2FC) %>% pull(gene) %>% unique()
proliferating_genes <- seurat_markers %>% filter(cluster == "Proliferating") %>% top_n(20, wt = avg_log2FC) %>% pull(gene) %>% unique()

all_genes <- rbind(data.frame(group = "celltype", gene = celltype_genes),
                         data.frame(group = "memory", gene = mem_genes),
                        data.frame(group = "central memory", gene = cm_genes),
                        data.frame(group = "exhaustion", gene = exh_genes),
                        data.frame(group = "effector", gene = eff_genes),
                        data.frame(group = "slec-like", gene = slec_genes),
                        data.frame(group = "treg", gene = treg_genes),
                        data.frame(group = "tfh", gene = tfh_genes),
                        data.frame(group = "th21", gene = th21_genes),
                        data.frame(group = "acinar-contaminated", gene = acinar_genes),
                        data.frame(group = "interferon-sensing", gene = ifn_genes),
                        data.frame(group = "activation", gene = activation_genes),
                        data.frame(group = "proliferation", gene = proliferating_genes))
all_genes <- all_genes %>% filter(!duplicated(gene))
all_genes <- merge(all_genes, as.data.frame(AverageExpression(seurat_integrated, features = all_genes$gene, assays = "RNA", group.by = "seurat_clusters")$RNA) %>% rownames_to_column("gene"), by = "gene")

write.table(all_genes, 
            "output/cluster_gene_expression_for_morpheus.tsv", 
            quote = F, 
            row.names = F, 
            sep = "\t")

top50 <- merge(seurat_markers[,c("cluster","gene")], 
               as.data.frame(AverageExpression(seurat_integrated, features = all_genes$gene, assays = "RNA", group.by = "seurat_clusters")$RNA) %>% rownames_to_column("gene"), by = "gene") %>% 
         filter(!duplicated(gene))
write.table(top50, 
            "output/cluster_top50gene_expression_for_morpheus.tsv", 
            quote = F, 
            row.names = F, 
            sep = "\t")
```

Now that the CD4 and CD8 T cells have been identified, next fix the clonotype designations so that only CD8 or CD4 T cell clones are distinguished. Note that **celltype** is now considered:
```{r fix clonotype metadata}
## CLONOTYPE ID
new_metadata <- seurat_integrated@meta.data %>% 
  rownames_to_column('barcode') %>%
  filter(!is.na(raw_clonotype_id)) %>% #remove cells without a clonotype_id
  group_by(celltype, mouse_id, TRB.v_gene, TRB.d_gene, TRB.j_gene, TRB.cdr3_nt ,TRA.v_gene, TRA.d_gene, TRA.j_gene, TRA.cdr3_nt) %>% 
  mutate(clonotype_count = n()) %>% #number of cells in clonotype
  arrange(desc(clonotype_count)) %>% 
  ungroup() %>% 
  arrange(desc(clonotype_count)) %>% 
  mutate(clonotype_id = paste0('clonotype', ... = match(paste(celltype, mouse_id, TRB.v_gene, TRB.d_gene, TRB.j_gene, TRB.cdr3_nt, TRA.v_gene, TRA.d_gene, TRA.j_gene, TRA.cdr3_nt), unique(paste(celltype, mouse_id, TRB.v_gene, TRB.d_gene, TRB.j_gene, TRB.cdr3_nt, TRA.v_gene, TRA.d_gene, TRA.j_gene, TRA.cdr3_nt))))) %>% #uniquely match VDJ for TCRB and TCRA and celltype (CD4 or CD8)
  mutate(clonotype_count_bin = cut(clonotype_count, 
                                   breaks=c(0,1,5,20,50,100,300), 
                                   labels=c("1","2-4","5-19","20-49","50-99","100-300"))) %>% #bin the counts
  column_to_rownames('barcode') %>%
  dplyr::select(clonotype_count, clonotype_id, clonotype_count_bin)
seurat_integrated <- AddMetaData(seurat_integrated, metadata = new_metadata)
```

Factor some additional variables to order them properly:
```{r factor additional variables}
seurat_integrated$treatment <- factor(seurat_integrated$treatment, levels = c("IgG", "Spt", "PD1"))
```

As a measure of exhaustion, the number of co-inhibitory receptors expressed simultaneously in the same cell can be counted. This will be added to the metadata:
```{r coinhibitory receptor count}
new_metadata <- data.frame(coinh_count = rowSums(as.data.frame(t(as.data.frame(seurat_integrated@assays$RNA@data[c("Pdcd1","Havcr2","Ctla4","Lag3","Tigit","Cd160","Cd244a"),]))) > 0))
seurat_integrated <- AddMetaData(seurat_integrated, new_metadata)
```

Modify the celltype of the CD4+ T cells to distinguish Tcon and Treg based on the Seurat clusters for later analyses:
```{r distinguish Tcon and Treg}
new_metadata <- seurat_integrated@meta.data %>% 
                  mutate(cd4_or_cd8 = paste0(celltype), 
                         celltype = ifelse(seurat_clusters %in% c("Treg_resting","Treg_eff"), "Treg", paste0(celltype))) %>% 
                  select(cd4_or_cd8, celltype)
seurat_integrated <- AddMetaData(seurat_integrated, new_metadata)
```

```{r identify tissue shared clonotypes}
shared_clonotypes <- seurat_integrated@meta.data %>% 
  filter(!is.na(clonotype_id)) %>%
  group_by(cd4_or_cd8, clonotype_id, tissue) %>% 
  tally() %>% 
  summarize(clonotype_tissue_groups = paste(sort(tissue), collapse = "&")) %>% #be sure to sort the tissue terms before collapsing
  arrange(desc(clonotype_tissue_groups)) %>% 
  data.frame()

shared_clonotypes <- setNames(as.character(shared_clonotypes$clonotype_tissue_groups), shared_clonotypes$clonotype_id)

new_metadata <- seurat_integrated@meta.data %>% 
  rownames_to_column('barcode') %>%
  mutate(clonotype_tissue_groups = shared_clonotypes[clonotype_id]) %>%
  column_to_rownames('barcode') %>%
  dplyr::select(tissue, clonotype_tissue_groups) 

new_metadata <- new_metadata %>%
                  mutate(clonotype_shared_tissues = case_when(is.na(clonotype_tissue_groups) ~ NA_character_,
                                                   tissue == "blood" & clonotype_tissue_groups %in% c("pLN&blood&pancreas", "blood&pancreas") ~ "blood PM",
                                                   tissue == "pLN"  & clonotype_tissue_groups %in% c("pLN&blood&pancreas", "pLN&pancreas") ~ "pLN PM",
                                                   tissue == "pancreas"  & clonotype_tissue_groups %in% c("pLN&blood&pancreas", "pLN&pancreas", "blood&pancreas") ~ "pancreas PM",
                                                   tissue == "pancreas"  & clonotype_tissue_groups == "pancreas" ~ "pancreas non-PM",
                                                   tissue == "blood"  & clonotype_tissue_groups %in% c("pLN&blood", "blood") ~ "blood non-PM",
                                                   tissue == "pLN"  & clonotype_tissue_groups %in% c("pLN&blood", "blood") ~ "pLN non-PM"),
                         pancreas_matching = case_when(is.na(clonotype_tissue_groups) ~ NA_character_,
                                                   clonotype_tissue_groups %in% c("pLN&blood&pancreas", "blood&pancreas", "pLN&pancreas") ~ "PM",
                                                   TRUE ~ "non-PM")) %>%
  dplyr::select(clonotype_shared_tissues, pancreas_matching)
seurat_integrated <- AddMetaData(seurat_integrated, new_metadata)
seurat_integrated$clonotype_shared_tissues <- factor(seurat_integrated$clonotype_shared_tissues, levels = c("pancreas PM", "pancreas non-PM", "pLN PM", "blood PM", "blood non-PM", "pLN non-PM"))
seurat_integrated$pancreas_matching <- factor(seurat_integrated$pancreas_matching, levels = c("non-PM", "PM"))
```

```{r identify celltype shared clonotypes (Tcon/Treg)}
shared_clonotypes <- seurat_integrated@meta.data %>% 
                      filter(!is.na(clonotype_id) & cd4_or_cd8 == "CD4") %>%
                      group_by(celltype, clonotype_id) %>% 
                      tally() %>% 
                      group_by(clonotype_id) %>%
                      summarize(clonotype_celltype_groups = paste(sort(celltype), collapse = "&"), .groups = "keep") %>% #be sure to sort the tissue terms before collapsing
                      arrange(desc(clonotype_celltype_groups)) %>% 
                      data.frame()

shared_clonotypes <- setNames(as.character(shared_clonotypes$clonotype_celltype_groups), shared_clonotypes$clonotype_id)

new_metadata <- seurat_integrated@meta.data %>% 
                  rownames_to_column('barcode') %>%
                  mutate(clonotype_celltype_groups = shared_clonotypes[clonotype_id]) %>%
                  column_to_rownames('barcode') %>%
                  dplyr::select(clonotype_celltype_groups) 
seurat_integrated <- AddMetaData(seurat_integrated, new_metadata)
seurat_integrated$clonotype_celltype_groups <- factor(seurat_integrated$clonotype_celltype_groups, levels = c("Tcon", "Treg", "Tcon&Treg"))
```

Next, gene signatures will be added to the Seurat to plot gene module scores. These gene signatures were gathered from various papers and also msigdb:
```{r gene module scores}
#read in data from excel sheets
prog_exh <- read_excel("ref/signatures/Miller2019_NatureImmuno_SuppTable1.xlsx",sheet = "Progenitor Exh.")$gene
term_exh <- read_excel("ref/signatures/Miller2019_NatureImmuno_SuppTable1.xlsx",sheet = "Terminally Exh.")$gene
eff_like <- read_excel("ref/signatures/Miller2019_NatureImmuno_SuppTable1.xlsx",sheet = "Effector-like")$gene
trm <- read_excel("ref/signatures/20200711_Beura_Trm_2017_Immunity.xlsx", sheet = "Beura_FRT_TRM_UP",col_names = F)$`...1`
nonbystander <- read_excel("ref/signatures/20200718_Mognol_PNAS2017_bystanders.xlsx", sheet = "OT-1 UP (TS)",col_names = F)$`...1`
bystander <- read_excel("ref/signatures/20200718_Mognol_PNAS2017_bystanders.xlsx", sheet = "P14 UP (bystander)",col_names = F)$`...1`
cd8_activation <- read.table("ref/signatures/cd8_activation_sarkar_up.txt")$V1
cell_cycle <- read.table("ref/signatures/cell_cycle_all_strong.txt")$V1
lag3 <- read_excel("ref/signatures/stephanie_de_genes.xlsx") #these DE genes are from FindAllMarkers, so they will be parsed out into separate signatures
  lag3exh1 <- lag3 %>% filter(cluster == 1) %>% filter(p_val_adj<0.05 & avg_logFC > 0) %>% pull(gene)
  lag3progexh2 <- lag3 %>% filter(cluster == 2) %>% filter(p_val_adj<0.05 & avg_logFC > 0) %>% pull(gene)
  lag3exh3 <- lag3 %>% filter(cluster == 3) %>% filter(p_val_adj<0.05 & avg_logFC > 0) %>% pull(gene)
  lag3act4 <- lag3 %>% filter(cluster == 4) %>% filter(p_val_adj<0.05 & avg_logFC > 0) %>% pull(gene)
  lag3act6 <- lag3 %>% filter(cluster == 6) %>% filter(p_val_adj<0.05 & avg_logFC > 0) %>% pull(gene)
  lag3eff7 <- lag3 %>% filter(cluster == 7) %>% filter(p_val_adj<0.05 & avg_logFC > 0) %>% pull(gene)
  lag3mem8 <- lag3 %>% filter(cluster == 8) %>% filter(p_val_adj<0.05 & avg_logFC > 0) %>% pull(gene)

#some of these are human gene symbols that must be converted to mouse
trm <- convertHumanGeneList(trm)
cell_cycle <- convertHumanGeneList(cell_cycle)
cd8_activation <- convertHumanGeneList(cd8_activation)

#some gene signatures will be from msigdb
m_df <- msigdbr(species = "Mus musculus", category = "C7")
seurat_integrated <- AddModuleScore(seurat_integrated, features = list(m_df %>% 
                                             filter(gs_name == "GSE8678_IL7R_LOW_VS_HIGH_EFF_CD8_TCELL_UP") %>% 
                                             dplyr::select(gene_symbol) %>% pull()), 
                      name = "SLEC", assay = "RNA")
seurat_integrated <- AddModuleScore(seurat_integrated, features = list(m_df %>% 
                                             filter(gs_name == "GSE8678_IL7R_LOW_VS_HIGH_EFF_CD8_TCELL_DN") %>% 
                                             dplyr::select(gene_symbol) %>% pull()), 
                      name = "MPEC", assay = "RNA") 
seurat_integrated <- AddModuleScore(seurat_integrated, features = list(m_df %>% 
                                             filter(gs_name == "KAECH_DAY15_EFF_VS_MEMORY_CD8_TCELL_UP") %>% 
                                             dplyr::select(gene_symbol) %>% pull()), 
                      name = "KAECH_DAY15_EFF_VS_MEMORY_CD8_TCELL_UP", assay = "RNA")
seurat_integrated <- AddModuleScore(seurat_integrated, features = list(m_df %>% 
                                             filter(gs_name == "KAECH_DAY15_EFF_VS_MEMORY_CD8_TCELL_DN") %>% 
                                             dplyr::select(gene_symbol) %>% pull()), 
                      name = "KAECH_DAY15_EFF_VS_MEMORY_CD8_TCELL_DN", assay = "RNA")
seurat_integrated <- AddModuleScore(seurat_integrated, features = list(m_df %>% 
                                             filter(gs_name == "GOLDRATH_EFF_VS_MEMORY_CD8_TCELL_DN") %>% 
                                             dplyr::select(gene_symbol) %>% pull()), 
                      name = "GOLDRATH_EFF_VS_MEMORY_CD8_TCELL_DN", assay = "RNA")

seurat_integrated <- AddModuleScore(seurat_integrated, features = list(prog_exh), name = "Miller2019_progenitor_exh", assay = "RNA")
seurat_integrated <- AddModuleScore(seurat_integrated, features = list(term_exh), name = "Miller2019_terminally_exh", assay = "RNA")
seurat_integrated <- AddModuleScore(seurat_integrated, features = list(eff_like), name = "Miller2019_effector_like", assay = "RNA")
seurat_integrated <- AddModuleScore(seurat_integrated, features = list(trm), name = "beura_trm", assay = "RNA")
seurat_integrated <- AddModuleScore(seurat_integrated, features = list(nonbystander), name = "nonbystander", assay = "RNA")
seurat_integrated <- AddModuleScore(seurat_integrated, features = list(bystander), name = "bystander", assay = "RNA")
seurat_integrated <- AddModuleScore(seurat_integrated, features = list(cd8_activation), name = "activation_sarkar", assay = "RNA")
seurat_integrated <- AddModuleScore(seurat_integrated, features = list(cell_cycle), name = "cell_cycle", assay = "RNA")
seurat_integrated <- AddModuleScore(seurat_integrated, features = list(lag3exh1), name = "lag3exh1", assay = "RNA")
seurat_integrated <- AddModuleScore(seurat_integrated, features = list(lag3progexh2), name = "lag3progexh2", assay = "RNA")
seurat_integrated <- AddModuleScore(seurat_integrated, features = list(lag3exh3), name = "lag3exh3", assay = "RNA")
seurat_integrated <- AddModuleScore(seurat_integrated, features = list(lag3mem8), name = "lag3mem8", assay = "RNA")
seurat_integrated <- AddModuleScore(seurat_integrated, features = list(lag3eff7), name = "lag3eff7", assay = "RNA")
seurat_integrated <- AddModuleScore(seurat_integrated, features = list(lag3act4), name = "lag3act4", assay = "RNA")
seurat_integrated <- AddModuleScore(seurat_integrated, features = list(lag3act6), name = "lag3act6", assay = "RNA")

#this isn't needed yet, but could be useful in the future
gene_signatures <- c("SLEC1",
                     "MPEC1",
                     "KAECH_DAY15_EFF_VS_MEMORY_CD8_TCELL_UP1",
                      "Miller2019_progenitor_exh1",
                     "Miller2019_terminally_exh1",
                     "Miller2019_effector_like1",
                     "beura_trm1",
                     "nonbystander1",
                     "bystander1",
                     "cell_cycle1",
                     "activation_sarkar1",
                     "lag3exh11",
                     "lag3progexh21",
                     "lag3exh31",
                     "lag3mem81",
                     "lag3eff71",
                     "lag3act41",
                     "lag3act61")
```

Add the NRP-v7 tetramer staining to the Seurat meta data for ease of plotting. An NRP-v7+ T cell was identified if: 1) The cell was classified as a CD8 T cell; 2) The cell had an associated TCR; 3) the cell had a NRP-7 tetramer TotalSeqC staining value of over 1.5. NRP-v7+ T cells were sorted by the Scheitinger lab for scRNA-seq: these TCRs will be examined to determine the overlap with the tetramer+ T cells identified in this experiment.
```{r NRPv7 tetramer staining}
new_metadata <- as.data.frame(t(data.frame(seurat_integrated@assays$FB@data)))
colnames(new_metadata) <- "NRPv7_TotalSeqC"
seurat_integrated <- AddMetaData(seurat_integrated, new_metadata)
new_metadata <- seurat_integrated@meta.data %>% 
                  mutate(tetramer_binding = ifelse(NRPv7_TotalSeqC > 1.5 & celltype == "CD8", #cutoff for specificity
                                                 "NRPv7", 
                                                 "unknown")) %>%
                  dplyr::select(tetramer_binding)
seurat_integrated <- AddMetaData(seurat_integrated, new_metadata)

sample_names <- list.files("data/GSE151652") #all samples are within individual folders
sample_names <- sample_names[grepl("consensus", sample_names)]

tetramer_tcrs <- read.table(paste0("data/GSE151652/", sample_names[1]), sep = ",", header = T)[,c("clonotype_id","chain","cdr3","reads","umis")] %>%
                  group_by(clonotype_id, chain) %>%  
                  arrange(desc(reads), desc(umis)) %>% 
                  top_n(1, umis) %>%  #select TCR by number of UMIs
                  top_n(1, reads) %>% #break ties by taking TCR with highest number of reads
                  dplyr::slice(1) %>% #pick top TCR at random if the above are equal using slice
                  cast(clonotype_id ~ chain, value = "cdr3") %>% 
                  filter(!is.na(TRA) | is.na(TRB)) %>%
                  dplyr::select(-clonotype_id) %>%
                  unique()

for (sample_name in sample_names[-1]) {
  tetramer_tcrs <- rbind(tetramer_tcrs,
                         read.table(paste0("data/GSE151652/", sample_name), sep = ",", header = T)[,c("clonotype_id","chain","cdr3","reads","umis")] %>%
                  group_by(clonotype_id, chain) %>%  
                  arrange(desc(reads), desc(umis)) %>% 
                  top_n(1, umis) %>%  #select TCR by number of UMIs
                  top_n(1, reads) %>% #break ties by taking TCR with highest number of reads
                  dplyr::slice(1) %>% #pick top TCR at random if the above are equal using slice
                  cast(clonotype_id ~ chain, value = "cdr3") %>% 
                  filter(!is.na(TRA) | is.na(TRB)) %>%
                  dplyr::select(-clonotype_id) %>%
                  unique())
}

tetramer_tcrs <- unique(tetramer_tcrs)
tetramer_tcrs$cdr3_combined <- paste0(tetramer_tcrs$TRA, tetramer_tcrs$TRB)

new_metadata <- seurat_integrated@meta.data %>% 
                  mutate(andreas_tetramer_specificity = ifelse(paste0(TRA.cdr3,TRB.cdr3) %in% tetramer_tcrs$cdr3_combined,"NRPv7", NA_character_)) %>% 
                  dplyr::select(andreas_tetramer_specificity)
seurat_integrated <- AddMetaData(seurat_integrated, new_metadata)

#use TCRtypes to determine specificity
tbl <- data.frame(seurat_integrated@meta.data %>%
        filter(!is.na(tcrtype_id) & tissue == "pancreas") %>% #only confident in pancreas staining
        group_by(tetramer_binding, tcrtype_id) %>%
        tally() %>%
        cast(tcrtype_id ~ tetramer_binding, value = "n")) %>%
        arrange(desc(NRPv7))
tbl[is.na(tbl)] <- 0
nrpv7_tcrtypes <- tbl %>% filter(NRPv7 > unknown) %>% pull(tcrtype_id) #consider a cell NRPv7 positive if there is more NRPv7 positive than negative cells
new_metadata <- seurat_integrated@meta.data %>%
                  mutate(tetramer_specificity = ifelse(tcrtype_id %in% nrpv7_tcrtypes, "NRPv7", "unknown")) %>%
                  dplyr::select(tetramer_specificity)
seurat_integrated <- AddMetaData(seurat_integrated, new_metadata)

cumulative_sums <- seurat_integrated@meta.data %>% 
                            filter(!is.na(clonotype_id) & tissue == "pancreas") %>% 
                            group_by(treatment, clonotype_id) %>% 
                            tally() %>% 
                            group_by(treatment) %>% 
                            mutate(prcnt = n*100/sum(n)) %>% 
                            group_by(treatment) %>% 
                            arrange(treatment, desc(prcnt)) %>% 
                            mutate(cs = cumsum(prcnt))

cumulative_sums$clonotype_id <- factor(cumulative_sums$clonotype_id, 
                                       levels = cumulative_sums$clonotype_id)
nrpv7_clonos <- seurat_integrated@meta.data %>% filter(tetramer_specificity == "NRPv7") %>% pull(clonotype_id) %>% unique()
cumulative_sums$tetramer_specificity <- ifelse(cumulative_sums$clonotype_id %in% nrpv7_clonos, "NRPv7", "unknown")
cumulative_sums$clonotype_id <- factor(cumulative_sums$clonotype_id, 
                                       levels = cumulative_sums$clonotype_id)

p_top50_clonos <- cumulative_sums %>% 
                            group_by(treatment) %>% 
                            slice_head(n = 50) %>% 
                            ggplot(aes(x = clonotype_id, 
                                       y = cs, 
                                       color = treatment)) + 
                            geom_point() + 
                            geom_segment(aes(xend = clonotype_id, 
                                             yend = cs, 
                                             x = clonotype_id, 
                                             y = 0, 
                                             color = tetramer_specificity)) + 
                            scale_color_manual(values = custom_colors("tetramer")) +
                            scale_y_continuous(expand = c(0, 0)) +
                            theme_minimal() +
                            theme(panel.grid = element_blank(),
                                  panel.spacing.x = unit(0, "lines"),
                                  text = element_text(size = 15),
                                  axis.text.x = element_blank(),
                                  panel.border = element_rect(colour = "black", 
                                                              fill = NA, 
                                                              size = 1),
                                  legend.position = "none") +
                            labs(x = "Top 50 clonotypes", 
                                 y = paste0("Cumulative % CD8 TCRs")) +
                            facet_wrap(~ treatment, scales = "free_x")
```

```{r Save list of TCRs}
write.csv(seurat_integrated@meta.data %>% 
            dplyr::select(treatment, tissue, mouse_id, celltype, clonotype_id, TRA.v_gene, TRA.j_gene, TRA.c_gene, TRA.cdr3_nt, TRA.cdr3, TRB.v_gene, TRB.j_gene, TRB.c_gene, TRB.cdr3_nt, TRB.cdr3) %>% 
            filter(!is.na(clonotype_id)),
          file = "output/tableS7.csv")
```

Finally, add the new metadata from the integrated Seurat back to the separate Seurats of the CD4 and CD8 T cells and then save.
```{r save seurat object}
new_metadata <- seurat_integrated@meta.data #take all the metadata from the integrated Seurat and add to the subsetted objects
cd8 <- AddMetaData(cd8, new_metadata)
cd4 <- AddMetaData(cd4, new_metadata)

#save the Seurats
saveRDS(cd4, "objects/seurat_cd4_clustered.rds")
saveRDS(cd8, "objects/seurat_cd8_clustered.rds")
saveRDS(seurat_integrated, "objects/seurat_integrated_clustered.rds")
```
