==> scripts/scRNAseq/2_pancreas_only_seurat_generation.Rmd <==
---
title: "1_integrated_seurat_generation"
output: rmarkdown::github_document
date: '2022-07-08'
---

```{r setup, include=FALSE, echo=FALSE}
knitr::opts_chunk$set(echo = TRUE, root.dir = "C:/Users/Strix/GitRepos/NOD_spt_pd1")
knitr::opts_knit$set(echo = TRUE, root.dir = "C:/Users/Strix/GitRepos/NOD_spt_pd1")
```

```{r libraries, message = FALSE, warning = FALSE}
require(dplyr, quietly = T) #simple manipulation of data.frames
require(Seurat, quietly = T) #scRNA-seq analysis
require(reshape, quietly = T) #melting and casting data.frames
require(ggplot2, quietly = T) #plotting
require(gridExtra, quietly = T) #display multiple plots
require(ggrepel, quietly = T) #labels for plots
require(tibble, quietly = T) #rownames_to_column function
require(useful, quietly = T) #corner function to peek at data.frames
set.seed(1) #set a seed for reproducibility
```

```{r read in integrated Seurat}
seurat_integrated <- readRDS("objects/seurat_integrated_clustered.rds")
```

For visualization, the T cells in the pancreas only will be clustered separately into distinct UMAPs:
```{r cluster pancreas-only cells}
Idents(seurat_integrated) <- seurat_integrated$tissue
pan8 <- subset(seurat_integrated, idents = "pancreas")

Idents(pan8) <- pan8$celltype
pan4 <- subset(pan8, idents = c("Tcon","Treg"))
pan8 <- subset(pan8, idents = "CD8")

pan8 <- ScaleData(pan8, verbose = FALSE) #scale the RNA data
pan8 <- RunPCA(pan8, npcs = 30, verbose = FALSE, features = rownames(x = pan8)) #run PCA
pan8 <- RunUMAP(pan8, reduction = "pca", dims = 1:30)
pan8 <- FindNeighbors(pan8, reduction = "pca", dims = 1:30)
pan8 <- FindClusters(pan8, resolution = 1.1)

pan4 <- ScaleData(pan4, verbose = FALSE) #scale the RNA data
pan4 <- RunPCA(pan4, npcs = 30, verbose = FALSE, features = rownames(x = pan4)) #run PCA
pan4 <- RunUMAP(pan4, reduction = "pca", dims = 1:30)
pan4 <- FindNeighbors(pan4, reduction = "pca", dims = 1:30)
pan4 <- FindClusters(pan4, resolution = 1.1)

#add the Seurat clusters from clustering of all the T cells to the smaller UMAPs
new_metadata <- seurat_integrated@meta.data %>% select(seurat_clusters)
pan4 <- AddMetaData(pan4, new_metadata)
pan8 <- AddMetaData(pan8, new_metadata)
Idents(pan4) <- pan4$seurat_clusters
Idents(pan8) <- pan8$seurat_clusters

#visualize the results
pan4_cluster <- DimPlot(pan4, reduction = "umap", label = TRUE, label.size = 5, split.by = "treatment") + theme(legend.position = "none")
pan4_cluster
pan8_cluster <- DimPlot(pan8, reduction = "umap", label = TRUE, label.size = 5, split.by = "treatment") + theme(legend.position = "none")
pan8_cluster
```

Add the UMAP cell embeddings (UMAP positions) to the meta data for the Seurats to facilitate plotting in ggplot:
```{r add cell embeddings to Seurat meta data}
new_metadata <- as.data.frame(seurat_integrated@reductions$umap@cell.embeddings) #get the cell embeddings
colnames(new_metadata) <- paste0("allcells_", colnames(new_metadata)) #prepend the UMAP coordinates with `allcells`
seurat_integrated <- AddMetaData(seurat_integrated, new_metadata)

new_metadata <- as.data.frame(cd8@reductions$umap@cell.embeddings) #get the cell embeddings
cd8 <- AddMetaData(cd8, new_metadata) #add to the meta data

new_metadata <- as.data.frame(cd4@reductions$umap@cell.embeddings) #get the cell embeddings
cd4 <- AddMetaData(cd4, new_metadata) #add to the meta data

new_metadata <- as.data.frame(pan4@reductions$umap@cell.embeddings) #get the cell embeddings
pan4 <- AddMetaData(pan4, new_metadata) #add to the meta data

new_metadata <- as.data.frame(pan8@reductions$umap@cell.embeddings) #get the cell embeddings
pan8 <- AddMetaData(pan8, new_metadata) #add to the meta data
```

Next, re-calculate the number of each clonotype within the pancreas (not including other tissues). This data will be re-added back to the integrated Seurat such that each clonotype that is in the pancreas or shared with the pancreas has the count associated with that clonotype regardless of the tissue of origin of that cell:
```{r calculate pancreas clonotype counts}
#generate a clonotype-centric data frame with the clonotype counts of cells in the pancreas
pan4_metadata <- pan4@meta.data %>% 
                  rownames_to_column('barcode') %>%
                  filter(!is.na(clonotype_id)) %>%
                  group_by(clonotype_id) %>% 
                  mutate(pan_clonotype_count = n(), 
                  pan_clonotype_count_bin = cut(pan_clonotype_count, 
                                              breaks = c(0,1,5,20,100,250), 
                                              labels = c("1","2-4","5-19","20-99","100-250"))) %>%
                  dplyr::select(clonotype_id, pan_clonotype_count, pan_clonotype_count_bin) %>%
                  unique()
metadata_for_integrated <- pan4_metadata #collect clonotype-centric metadata for integrated Seurat object
pan4_metadata <- merge(pan4@meta.data %>% #generate a data frame with cell barcodes and corresponding clonotype
                        dplyr::select(clonotype_id) %>% 
                        filter(!is.na(clonotype_id)) %>% 
                        rownames_to_column("barcode"),
                      pan4_metadata, 
                      by = "clonotype_id") %>% #merge with the clonotype-centric data frame
                      dplyr::select(-clonotype_id) %>%
                      column_to_rownames("barcode")
pan4 <- AddMetaData(pan4, pan4_metadata) #add the new clonotype count within the pancreas only cells

#do the same for CD8 T cells
pan8_metadata <- pan8@meta.data %>% 
                  rownames_to_column('barcode') %>%
                  filter(!is.na(clonotype_id)) %>%
                  group_by(clonotype_id) %>% 
                  mutate(pan_clonotype_count = n(), 
                  pan_clonotype_count_bin = cut(pan_clonotype_count, 
                                              breaks = c(0,1,5,20,100,250), 
                                              labels = c("1","2-4","5-19","20-99","100-250"))) %>%
                  dplyr::select(clonotype_id, pan_clonotype_count, pan_clonotype_count_bin) %>%
                  unique()
metadata_for_integrated <- rbind(metadata_for_integrated, pan8_metadata)
pan8_metadata <- merge(pan8@meta.data %>% #generate a data frame with cell barcodes and corresponding clonotype
                        dplyr::select(clonotype_id) %>% 
                        filter(!is.na(clonotype_id)) %>% 
                        rownames_to_column("barcode"),
                      pan8_metadata, 
                      by = "clonotype_id") %>% #merge with the clonotype-centric data frame
                      dplyr::select(-clonotype_id) %>%
                      column_to_rownames("barcode")
pan8 <- AddMetaData(pan8, pan8_metadata) #add the new clonotype count within the pancreas only cells

#add this data back to the integrated Seurat
metadata_for_integrated <- merge(seurat_integrated@meta.data %>% #generate a data frame with cell barcodes and corresponding clonotype
                                dplyr::select(clonotype_id) %>% 
                                filter(!is.na(clonotype_id)) %>% 
                                rownames_to_column("barcode"),
                              metadata_for_integrated, 
                                by = "clonotype_id") %>% #merge with the clonotype-centric data frame
                              dplyr::select(-clonotype_id) %>%
                              column_to_rownames("barcode")
seurat_integrated <- AddMetaData(seurat_integrated, metadata_for_integrated)
```

Finally, save the Seurat objects with their new metadata:
```{r save Seurat objects}
saveRDS(pan4, "objects/seurat_cd4_panonly_clustered.rds")
saveRDS(pan8, "objects/seurat_cd8_panonly_clustered.rds")
saveRDS(seurat_integrated, "objects/seurat_integrated_clustered.rds")
```
