==> scripts/scRNAseq/6_figure5_plots.Rmd <==
---
title: "6_figure5_plots"
output: rmarkdown::github_document
date: '2022-07-10'
---

The purpose of this figure is to show CD4+ and CD8+ T cells can be identified in the pancreatic LN and blood using the TCR as a barcode that have shared clonality with cells in the pancreas.

```{r setup, include=FALSE, echo=FALSE}
knitr::opts_chunk$set(echo = TRUE, root.dir = "C:/Users/Strix/GitRepos/NOD_spt_pd1")
knitr::opts_knit$set(echo = TRUE, root.dir = "C:/Users/Strix/GitRepos/NOD_spt_pd1")
```

```{r libraries, message = FALSE, warning = FALSE}
require(tidyr)
require(dplyr, quietly = T) #simple manipulation of data.frames
require(Seurat, quietly = T) #scRNA-seq analysis
require(reshape, quietly = T) #melting and casting data.frames
require(ggplot2, quietly = T) #plotting
require(rstatix, quietly = T) #adding p values to plots
require(ggprism, quietly = T) #adding p values to plots
require(ggsci, quietly = T) #IGV color palette
require(ComplexUpset)
require(tidyr)
require(patchwork, quietly = T) #adding p values to plots
require(magrittr, quietly = T) #adding p values to plots
require(gridExtra, quietly = T) #display multiple plots
require(ggrepel, quietly = T) #labels for plots
require(tibble, quietly = T) #rownames_to_column function
require(useful, quietly = T) #corner function to peek at data.frames
require(viridis, quietly = T) #viridis color palettes
require(readxl, quietly = T) #reading Excel files
require(msigdbr, quietly = T) #accessing gene signatures from mSigDB
source("scripts/utils.R") #custom color palettes and other functions
require(tidyr, quietly = T) #pivot_wider for upset data frame preparation
library(ComplexUpset) #upset plots
set.seed(1) #set a seed for reproducibility
```

```{r read in Seurats}
seurat_integrated <- saveRDS("objects/seurat_integrated_clustered.rds")
pan4 <- readRDS("objects/seurat_cd4_panonly_clustered.rds")
pan8 <- readRDS("objects/seurat_cd8_panonly_clustered.rds")
```


```{r UMAPs of shared/non-shared}
p_allcells_shared <- seurat_integrated@meta.data %>% 
                      filter(!is.na(clonotype_shared_tissues)) %>%
                      ggplot(aes(x = allcells_UMAP_1, 
                                 y = allcells_UMAP_2, 
                                 color = seurat_clusters)) + 
                      geom_point(data = seurat_integrated@meta.data %>% 
                                     dplyr::select(-clonotype_shared_tissues), 
                                 shape = 16, 
                                 color = "grey90") +
                      geom_point(shape = 16, 
                                 alpha = 0.3) +
                      theme_minimal() +
                      theme(axis.title = element_blank(),
                            axis.text = element_blank(),
                            panel.grid = element_blank(),
                            text = element_text(size = 15),
                            legend.position = "none") +
                      labs(x = 'UMAP 1', 
                           y = 'UMAP 2', 
                           color=NULL) +
                      scale_alpha_continuous(range = c(0.4 ,1)) +
                      scale_color_manual(values = custom_colors("clusters")) +
                      scale_size_continuous(range=c(1, 3)) +
                      guides(color = guide_legend(override.aes = list(alpha = 1, 
                                                                      size = 5, 
                                                                      shape = 19)), 
                             size = guide_legend()) +
                      facet_wrap(~clonotype_shared_tissues)

#PIES OF CLUSTERS IN EACH GROUP
p_allcells_shared_pie <- seurat_integrated@meta.data %>% 
    filter(!is.na(clonotype_shared_tissues)) %>%
    group_by(clonotype_shared_tissues,
             seurat_clusters) %>% 
    tally() %>% 
    group_by(clonotype_shared_tissues) %>% 
    mutate(prcnt = n*100/sum(n),
           csum = rev(cumsum(rev(prcnt))),  #get the position for the labels
           pos = prcnt/2 + lead(csum, 1),
           pos = if_else(is.na(pos), prcnt/2, pos)) %>%
    ggplot(aes(x = "", 
               y = prcnt, 
               fill = seurat_clusters)) + # pct used here so slices add to 100
        geom_bar(stat="identity", width=1) +
        coord_polar("y", start=0) +
        #geom_text(aes(label = paste0(round(prcnt, 0), "%")), position = position_stack(vjust = 0.5)) +
        facet_wrap(~clonotype_shared_tissues, ncol = 3) +
        theme_void() +
        theme(legend.position = "none",
              text = element_text(size = 15)) +
      geom_label_repel(aes(y = pos, label = paste0(round(prcnt, 0), "%")),
                         size = 4.5, nudge_x = 0.5, show.legend = FALSE) +
  scale_fill_manual(values = custom_colors("clusters"))

grid.arrange(p_allcells_shared, p_allcells_shared_pie, nrow = 2)

#STACKED BAR OF PM/NON-PM ACROSS TISSUES
p_shared_per_cluster_by_tissue <- seurat_integrated@meta.data %>% 
    filter(!is.na(pancreas_matching)) %>% #remove T cells where no TCR detected
    ggplot(aes(x = factor(seurat_clusters, 
                          levels = rev(levels(seurat_integrated$seurat_clusters))), 
               fill = pancreas_matching)) + 
    geom_bar(stat="count", 
             position = "fill") + 
    scale_x_discrete(expand = c(0,0)) +
    scale_y_continuous(labels = scales::percent_format(accuracy = 1), 
                       expand=c(0,0)) +
    theme_minimal() +
    theme(panel.grid = element_blank(),
          text = element_text(size = 15),
          legend.position = "right") +
    labs(title = "T cells in pancreas", 
         y = "Percent of Seurat Cluster", 
         x = NULL,
         fill = NULL) +
  facet_wrap(~tissue) +
    scale_fill_manual(values = custom_colors("pm")) +
    coord_flip()
p_shared_per_cluster_by_tissue

```

```{r UMAP of pancreas only cells}
#UMAP of pancreas only shared/non-shared for CD4 and CD8 T cell clusters
p_pan8_shared <- pan8@meta.data %>% 
                  filter(!is.na(pancreas_matching)) %>%
                  ggplot(aes(x = UMAP_1, 
                             y = UMAP_2, 
                             color = pancreas_matching)) + 
                  geom_point(data = pan8@meta.data %>% 
                               dplyr::select(-treatment), 
                             shape = 16, 
                             color = "grey90") +
                  geom_point(shape = 16, 
                             alpha = 0.3) +
                  theme_minimal() +
                  theme(axis.title = element_blank(),
                        axis.text = element_blank(),
                        panel.grid = element_blank(),
                        text = element_blank(),
                        legend.position = "none") +
                  labs(color = "Pancreas matching (PM)",
                       title = "CD8+ T cells in pancreas") +
                  scale_color_manual(values = custom_colors("pm")) +
                  guides(color = guide_legend(override.aes = list(alpha = 1, 
                                                                size = 5, 
                                                                shape = 19))) +
                  facet_wrap(~ treatment, ncol =1)

p_pan4_shared <- pan4@meta.data %>% 
                  filter(!is.na(pancreas_matching)) %>%
                  ggplot(aes(x = UMAP_1, 
                             y = UMAP_2, 
                             color = pancreas_matching)) + 
                  geom_point(data = pan4@meta.data %>% 
                               dplyr::select(-treatment), 
                             shape = 16, 
                             color = "grey90") +
                  geom_point(shape = 16, 
                             alpha = 0.3) +
                  theme_minimal() +
                  theme(axis.title = element_blank(),
                        axis.text = element_blank(),
                        panel.grid = element_blank(),
                        text =  element_blank(),
                        legend.position = "none") +
                  labs(color = NULL,
                       title = NULL) +
                  scale_color_manual(values = custom_colors("pm")) +
                  guides(color = guide_legend(override.aes = list(alpha = 1, 
                                                                size = 5, 
                                                                shape = 19))) +
                  facet_wrap(~ treatment, ncol = 1)

p_cd8_panonly_cluster <- pan8@meta.data %>% 
                            ggplot(aes(x = UMAP_1, 
                                     y = UMAP_2, 
                                     color = seurat_clusters)) + 
                            geom_point(shape = 16, 
                                       alpha = 0.6) +
                            theme_minimal() +
                            theme(axis.title = element_blank(),
                                  axis.text = element_blank(),
                                  text =  element_blank(),
                                  panel.grid = element_blank(),
                                  legend.position = "none") +
                            labs(x = 'UMAP 1', 
                                 y = 'UMAP 2', 
                                 color = NULL,
                                 size = NULL) +
                            scale_alpha_continuous(range = c(0.4,1)) +
                            scale_size_continuous(range = c(1,3)) +
                            # geom_label_repel(data = pan8_cls, #change depending on Seurat used
                            #      mapping = aes(x = U1,
                            #                    y = U2, 
                            #                    label = seurat_clusters), 
                            #      size = 5, 
                            #      min.segment.length = Inf, 
                            #      show.legend = FALSE) +
                            guides(color = guide_legend(override.aes = list(alpha=1, 
                                                                          size=5, 
                                                                          shape=19)), 
                                   size = guide_legend()) +
                            scale_color_manual(values = custom_colors("clusters"),
                                               limits = force) #drops unused clusters in legend

p_cd4_panonly_cluster <- pan4@meta.data %>% 
                            ggplot(aes(x = UMAP_1, 
                                     y = UMAP_2, 
                                     color = seurat_clusters)) + 
                            geom_point(shape = 16, 
                                       alpha = 0.6) +
                            theme_minimal() +
                            theme(axis.title = element_blank(),
                                  axis.text = element_blank(),
                                  text = element_text(size = 15),
                                  panel.grid = element_blank(),
                                  legend.position = "none") +
                            labs(x = 'UMAP 1', 
                                 y = 'UMAP 2', 
                                 color = "Seurat clusters",
                                 size = "Size of clonotype\n(pancreas only)") +
                            scale_alpha_continuous(range = c(0.4,1)) +
                            scale_size_continuous(range = c(1,3)) +
                            # geom_label_repel(data = pan8_cls, #change depending on Seurat used
                            #      mapping = aes(x = U1,
                            #                    y = U2, 
                            #                    label = seurat_clusters), 
                            #      size = 5, 
                            #      min.segment.length = Inf, 
                            #      show.legend = FALSE) +
                            guides(color = guide_legend(override.aes = list(alpha=1, 
                                                                          size=5, 
                                                                          shape=19)), 
                                   size = guide_legend()) +
                            scale_color_manual(values = custom_colors("clusters"),
                                               limits = force) #drops unused clusters in legend

grid.arrange(p_cd8_panonly_cluster, p_cd4_panonly_cluster, p_pan8_shared, p_pan4_shared, ncol = 2, heights = c(1,3))

#STACKED BAR PLOT OF SHARED/NON-SHARED
p_shared_per_cluster <- seurat_integrated@meta.data %>% 
                          filter(tissue == "pancreas" & !is.na(pancreas_matching)) %>% #remove T cells where no TCR detected
                            ggplot(aes(x = factor(seurat_clusters, 
                                                levels = rev(levels(seurat_integrated$seurat_clusters))), 
                                                fill = pancreas_matching)) + 
                            geom_bar(stat="count", 
                                     position = "fill") + 
                            scale_x_discrete(expand = c(0,0)) +
                            scale_y_continuous(labels = scales::percent_format(accuracy = 1), 
                                               expand=c(0,0)) +
                            theme_minimal() +
                            theme(panel.grid = element_blank(),
                                  text = element_text(size = 15),
                                  legend.position = "right") +
                            labs(title = "T cells in pancreas", 
                                 y = "Percent of Seurat Cluster", 
                                 x = NULL,
                                 fill = "Pancreas matching (PM)") +
                            scale_fill_manual(values = custom_colors("pm")) +
                            coord_flip()
p_shared_per_cluster

#ratio of shared/non-shared
rbind(pan4@meta.data %>% 
        filter(!is.na(pancreas_matching)) %>% 
        group_by(treatment, mouse_id, pancreas_matching) %>% 
        tally() %>% 
        mutate(celltype="CD4") %>% 
        cast(treatment+mouse_id+celltype ~ pancreas_matching, value = "n"),
      pan8@meta.data %>% 
        filter(!is.na(pancreas_matching)) %>% 
        group_by(treatment, mouse_id, pancreas_matching) %>% 
        tally() %>% 
        mutate(celltype="CD8") %>% 
        cast(treatment+mouse_id+celltype ~ pancreas_matching, value = "n")) %>% 
        group_by(treatment, mouse_id, celltype) %>% 
        summarize(prcnt = 100*PM/(`non-PM`+PM)) %>% 
  ggplot(aes(x = treatment, 
             color = treatment, 
             fill = treatment, 
             y = prcnt)) + 
    geom_violin(alpha = 0.4, 
                adjust = 0.8, 
                scale = "width", 
                draw_quantiles = 0.5) + 
    geom_point(size = 4) + 
    theme_minimal() +
    theme(panel.grid = element_blank(),
          panel.border = element_rect(color= "black", fill =NA, size=0.5),
          panel.spacing = unit(0, "cm"),
          axis.title.x = element_blank(),
          text = element_text(size = 15)) +
    scale_color_manual(values = custom_colors("treatments")) +
    scale_fill_manual(values = custom_colors("treatments")) +
    facet_wrap(~ celltype, 
               scales = "free_y",
               ncol = 1) +
    labs(y = "Pancreas-matching (%)",
         title = "T cells in pancreas")
```

Show the top 120 clonotypes shared across all three tissues to look at patterns in seurat clusters.
```{r top 100 clonotype seurat clusters shared across tissues}
seurat_integrated@meta.data %>% 
  filter(clonotype_id %in% paste0("clonotype",c(2:70)) & 
           clonotype_tissue_groups == "pLN&blood&pancreas" & 
           celltype == "CD8") %>% 
  group_by(clonotype_id, treatment, tissue, seurat_clusters) %>% 
  tally() %>% 
  ggplot(aes(x=tissue, fill = seurat_clusters, y = n)) + 
  geom_bar(stat="identity", position = "fill") + 
  facet_wrap(treatment~clonotype_id) + 
  scale_fill_manual(values = custom_colors("clusters")) + 
  theme_minimal() + 
  theme(legend.position = "none", 
        panel.grid = element_blank(), 
        axis.text = element_blank(), 
        axis.title = element_blank())

seurat_integrated@meta.data %>% 
  filter(clonotype_id %in% paste0("clonotype",c(2:70)) & 
           clonotype_tissue_groups == "pLN&blood&pancreas" & 
           celltype == "CD8") %>% 
  group_by(clonotype_id, treatment, tissue, seurat_clusters) %>% 
  tally() %>% 
  ggplot(aes(x=tissue, fill = seurat_clusters, y = n+1)) + 
  geom_bar(stat="identity", position = "stack") + 
  facet_wrap(treatment~clonotype_id) + 
  scale_fill_manual(values = custom_colors("clusters")) + 
  theme_minimal() + 
  scale_y_continuous(trans = "log10") +
  theme(legend.position = "none", 
        panel.grid = element_blank(), 
        axis.text = element_blank(), 
        axis.title = element_blank())

clonotypes_of_interest <- c(paste0("clonotype", c(1,7,18,26,28,31,32,35)), paste0("clonotype", c(3,5,10,12,13,14,17,20)), paste0("clonotype", c(4,6,9,30,54,90,96,105)))
clonotypes_of_interest <- paste0("clonotype", c(3,4,7))

p_topshared_clonotypes <- seurat_integrated@meta.data %>% 
                            filter(clonotype_id %in% clonotypes_of_interest) %>%
                            dplyr::select(treatment,
                                          clonotype_id, 
                                          seurat_clusters,
                                          allcells_UMAP_1, 
                                          allcells_UMAP_2) %>% 
                            group_by(seurat_clusters) %>% 
                            mutate(U1 = mean(allcells_UMAP_1),
                                   U2 = mean(allcells_UMAP_2)) %>%
                            group_by(treatment,
                                     clonotype_id, 
                                     seurat_clusters, U1, U2) %>% 
                            summarise(num_cells = n(),
                                      .groups = "keep") %>% 
                            ggplot() +
                            geom_point(data = seurat_integrated@meta.data %>%
                                         dplyr::select(-treatment) %>%
                                         dplyr::select(-clonotype_id),
                                       aes(x = allcells_UMAP_1, 
                                           y = allcells_UMAP_2),
                                       shape = 16,
                                       color = "grey90") +
                            geom_point(aes(x = U1,
                                           y = U2,
                                           size = num_cells,
                                           color = seurat_clusters),
                                       shape = 16) +
                            theme_minimal() +
                            theme(axis.title = element_blank(),
                                  axis.text = element_blank(),
                                  panel.grid = element_blank(),
                                  text = element_blank(),
                                  legend.position = "none") +
                            labs(color = NULL) +
                            scale_color_manual(values = custom_colors("clusters")) +
                            scale_size_continuous(range = c(5,20)) +
                            guides(color = guide_legend(override.aes = list(alpha = 1, 
                                                                            size = 5, 
                                                                            shape = 19))) +
                            facet_wrap(treatment ~ clonotype_id, ncol = 9) +
                            geom_text(aes(x = U1, y = U2, label = num_cells), color = "white", size = 6, fontface = "bold")
p_topshared_clonotypes

#stacked bar plots of shared clusters in blood and pancreas
seurat_integrated@meta.data %>% 
    filter(!is.na(pancreas_matching)) %>% 
    mutate(peripheral = ifelse(tissue == "pancreas", "pancreas", "periphery")) %>%
    group_by(treatment, pancreas_matching, cd4_or_cd8, seurat_clusters, peripheral) %>% 
    tally() %>% 
    ggplot(aes(x = treatment, 
               fill = seurat_clusters, 
               y = n)) + 
    geom_bar(stat="identity", 
             position = "fill") + 
    facet_wrap(peripheral ~ cd4_or_cd8 + pancreas_matching, ncol = 4) + 
    theme_minimal() + 
    theme(panel.grid = element_blank()) +
    scale_fill_manual(values = custom_colors("clusters"))
# clonos_of_interest <- seurat_integrated@meta.data %>% 
#                               filter(tissue == "pancreas" & 
#                                        clonotype_tissue_groups %in% c("pLN&blood&pancreas")) %>% 
#                               pull(unique(clonotype_id))
# seurat_integrated@meta.data %>% 
#     filter(clonotype_id %in% clonos_of_interest) %>% 
#   group_by(treatment, tissue, cd4_or_cd8, seurat_clusters) %>% 
#   tally() %>% 
#   ggplot(aes(x = tissue, 
#              fill = seurat_clusters, 
#              y = n)) + 
#   geom_bar(stat="identity", 
#            position = "fill") + 
#   facet_wrap(cd4_or_cd8 ~ treatment) + 
#   theme_minimal() + 
#   theme(panel.grid = element_blank()) +
#   scale_fill_manual(values = custom_colors("clusters"))

#UPSET PLOT
clonos_of_interest <- seurat_integrated@meta.data %>% 
                              filter(tissue == "pancreas" & clonotype_tissue_groups %in% c("pLN&blood&pancreas", "blood&pancreas") & seurat_clusters %in% c("CD8_texh", "CD8_eff", "CD8_mem")) %>% 
                              dplyr::select(treatment, seurat_clusters, clonotype_id)
colnames(clonos_of_interest)[which(colnames(clonos_of_interest) == "seurat_clusters")] <- "pan_seurat_cluster"

merge(clonos_of_interest, seurat_integrated@meta.data %>%
          filter(tissue != "pancreas" & clonotype_id %in% clonos_of_interest$clonotype_id) %>%
          group_by(clonotype_id, tissue, seurat_clusters) %>%
          tally() %>%
          cast(clonotype_id + seurat_clusters ~ tissue, value = "n", fill = 0), by = "clonotype_id") %>% group_by(treatment,pan_seurat_cluster, seurat_clusters) %>% 
    summarize(pLN = sum(pLN), blood = sum(blood)) %>%
    data.frame() %>%
    melt(id.vars = c("treatment", "pan_seurat_cluster", "seurat_clusters")) %>%
    ggplot(aes(x = variable, 
               y = value, 
               fill = seurat_clusters)) +
    geom_bar(stat = "identity", 
             position = "fill") + 
    facet_wrap(treatment ~ pan_seurat_cluster, nrow = 3) + 
    theme_minimal() + 
    theme(panel.grid = element_blank(),
          legend.position = "none") +
    scale_fill_manual(values = custom_colors("clusters"))

#stacked bar plot of clusters in pancreas by what cluster they are in blood
merge(clonos_of_interest, seurat_integrated@meta.data %>% 
          filter(tissue == "blood" & clonotype_id %in% clonos_of_interest$clonotype_id) %>%
          group_by(clonotype_id, seurat_clusters) %>% 
          tally(), by = "clonotype_id") %>% group_by(treatment, pan_seurat_cluster, seurat_clusters) %>% summarize(num_cells = sum(n)) %>% group_by(treatment, pan_seurat_cluster) %>%
    mutate(label_y = ((sum(num_cells) - cumsum(num_cells))/sum(num_cells)))   %>% ggplot(aes(x = pan_seurat_cluster, y = num_cells, fill = seurat_clusters)) +
    geom_bar(stat="identity", 
             position = "fill") + 
    facet_wrap(~treatment, nrow = 3) + 
    theme_minimal() + 
    theme(panel.grid = element_blank()) +
    scale_fill_manual(values = custom_colors("clusters")) + 
    geom_text(aes(y = label_y, 
                label = num_cells), 
            color = "white")

```

```{r boxplot of gene expression}
new_metadata <- data.frame(t(as.matrix(seurat_integrated@assays$RNA[c("Entpd1","Slamf6","Tcf7"),])))
seurat_integrated <- AddMetaData(seurat_integrated, new_metadata)

expression_means_df <- seurat_integrated@meta.data %>% 
                      filter(cd4_or_cd8 == "CD4") %>%
                      group_by(mouse_id, seurat_clusters) %>%
                      summarize(Entpd1 = mean(Entpd1), Slamf6 = mean(Slamf6), Tcf7 = mean(Tcf7)) %>%
                      ungroup() %>%
                      summarize(Entpd1 = mean(Entpd1), Slamf6 = mean(Slamf6), Tcf7 = mean(Tcf7)) %>%
                      data.frame()
expression_means <- expression_means_df[1,]
names(expression_means_df) <- colnames(expression_means_df)

#No dashed lines
seurat_integrated@meta.data %>% 
    filter(cd4_or_cd8 == "CD4" & tissue == "pancreas") %>%
    group_by(mouse_id, seurat_clusters) %>%
    summarize(Entpd1 = mean(Entpd1), Slamf6 = mean(Slamf6), Tcf7 = mean(Tcf7)) %>%
    reshape2::melt(id.vars = c("mouse_id","seurat_clusters")) %>%
    mutate(seurat_clusters = factor(seurat_clusters, levels = rev(levels(pan4$seurat_clusters)))) %>%
    ggplot(aes(y = seurat_clusters, x = value, fill = seurat_clusters, color = seurat_clusters)) +
    geom_boxplot(alpha = 0.5,
                 outlier.shape = NA) + 
    theme_minimal() + 
    theme(panel.spacing = unit(0,"lines"),
          axis.title.x = element_blank(),
          panel.grid = element_blank(),
          legend.position = "none",
          panel.border = element_rect(color = "black", 
                                      size = 0.5, 
                                      fill = NA)) +
    scale_fill_manual(values = custom_colors("clusters")) + 
    scale_color_manual(values = custom_colors("clusters")) +
    labs(x = "Entpd1 expression", y = NULL) +
    facet_wrap(~variable, scales = "free_x")

#DASHED MEAN LINES
p_entpd1 <- seurat_integrated@meta.data %>% 
    filter(cd4_or_cd8 == "CD4" & tissue == "pancreas") %>%
    group_by(mouse_id, seurat_clusters) %>%
    summarize(Entpd1 = mean(Entpd1), Slamf6 = mean(Slamf6), Tcf7 = mean(Tcf7)) %>%
    mutate(seurat_clusters = factor(seurat_clusters, levels = rev(levels(pan4$seurat_clusters)))) %>%
    ggplot(aes(y = seurat_clusters, x = Entpd1, fill = seurat_clusters, color = seurat_clusters)) +
    geom_boxplot(alpha = 0.5,
                 outlier.shape = NA) + 
    theme_minimal() + 
    theme(panel.spacing = unit(0,"lines"),
          axis.title.x = element_blank(),
          panel.grid = element_blank(),
          legend.position = "none",
          panel.border = element_rect(color = "black", 
                                      size = 0.5, 
                                      fill = NA)) +
    scale_fill_manual(values = custom_colors("clusters")) + 
    scale_color_manual(values = custom_colors("clusters")) +
    labs(x = "Entpd1 expression", y = NULL) +
    geom_vline(xintercept =  mean(seurat_integrated@meta.data %>% 
                        filter(cd4_or_cd8 == "CD4" & tissue == "pancreas") %>%
                        group_by(mouse_id, seurat_clusters) %>%
                        summarize(Entpd1 = mean(Entpd1), Slamf6 = mean(Slamf6), Tcf7 = mean(Tcf7)) %>% 
                          dplyr::select(Entpd1) %>% 
                          pull()))
p_slamf6 <- seurat_integrated@meta.data %>% 
    filter(cd4_or_cd8 == "CD4" & tissue == "pancreas") %>%
    group_by(mouse_id, seurat_clusters) %>%
    summarize(Entpd1 = mean(Entpd1), Slamf6 = mean(Slamf6), Tcf7 = mean(Tcf7)) %>%
    mutate(seurat_clusters = factor(seurat_clusters, levels = rev(levels(pan4$seurat_clusters)))) %>%
    ggplot(aes(y = seurat_clusters, x = Slamf6, fill = seurat_clusters, color = seurat_clusters)) +
    geom_boxplot(alpha = 0.5,
                 outlier.shape = NA) + 
    theme_minimal() + 
    theme(panel.spacing = unit(0,"lines"),
          axis.title.x = element_blank(),
          panel.grid = element_blank(),
          legend.position = "none",
          panel.border = element_rect(color = "black", 
                                      size = 0.5, 
                                      fill = NA)) +
    scale_fill_manual(values = custom_colors("clusters")) + 
    scale_color_manual(values = custom_colors("clusters")) +
    labs(x = "Entpd1 expression", y = NULL) +
    geom_vline(xintercept =  mean(seurat_integrated@meta.data %>% 
                        filter(cd4_or_cd8 == "CD4" & tissue == "pancreas") %>%
                        group_by(mouse_id, seurat_clusters) %>%
                        summarize(Entpd1 = mean(Entpd1), Slamf6 = mean(Slamf6), Tcf7 = mean(Tcf7)) %>% 
                          dplyr::select(Slamf6) %>% 
                          pull()))
p_tcf7 <- seurat_integrated@meta.data %>% 
    filter(cd4_or_cd8 == "CD4" & tissue == "pancreas") %>%
    group_by(mouse_id, seurat_clusters) %>%
    summarize(Entpd1 = mean(Entpd1), Slamf6 = mean(Slamf6), Tcf7 = mean(Tcf7)) %>%
    mutate(seurat_clusters = factor(seurat_clusters, levels = rev(levels(pan4$seurat_clusters)))) %>%
    ggplot(aes(y = seurat_clusters, x = Tcf7, fill = seurat_clusters, color = seurat_clusters)) +
    geom_boxplot(alpha = 0.5,
                 outlier.shape = NA) + 
    theme_minimal() + 
    theme(panel.spacing = unit(0,"lines"),
          axis.title.x = element_blank(),
          panel.grid = element_blank(),
          legend.position = "none",
          panel.border = element_rect(color = "black", 
                                      size = 0.5, 
                                      fill = NA)) +
    scale_fill_manual(values = custom_colors("clusters")) + 
    scale_color_manual(values = custom_colors("clusters")) +
    labs(x = "Entpd1 expression", y = NULL) +
    geom_vline(xintercept =  mean(seurat_integrated@meta.data %>% 
                        filter(cd4_or_cd8 == "CD4" & tissue == "pancreas") %>%
                        group_by(mouse_id, seurat_clusters) %>%
                        summarize(Entpd1 = mean(Entpd1), Slamf6 = mean(Slamf6), Tcf7 = mean(Tcf7)) %>% 
                          dplyr::select(Tcf7) %>% 
                          pull()))

grid.arrange(p_entpd1, p_slamf6, p_tcf7, ncol = 3) #plot the two genes together

seurat_integrated@meta.data %>% 
    filter(cd4_or_cd8 == "CD4") %>%
    group_by(mouse_id, seurat_clusters) %>%
    summarize(Entpd1 = mean(Entpd1), Slamf6 = mean(Slamf6)) %>%
  reshape2::melt(id.vars = c("mouse_id","seurat_clusters")) %>%
  group_by(variable) %>%
    t_test(value ~ seurat_clusters)
```

``` {r violin of scRNA-seq pancreas-matching}
cd8s <- seurat_integrated@meta.data %>% 
    filter(tissue != "pancreas" & !is.na(pancreas_matching)) %>% 
    dplyr::select(treatment, tissue, cd4_or_cd8, mouse_id, pancreas_matching, clonotype_id) %>% 
    unique() %>%
    group_by(treatment, tissue, cd4_or_cd8, mouse_id, pancreas_matching) %>% 
    tally() %>% 
    cast(treatment+mouse_id+tissue+cd4_or_cd8~pancreas_matching, value = "n") %>% 
    group_by(treatment, cd4_or_cd8, mouse_id, tissue) %>% 
    mutate(prcnt = 100*PM/sum(PM,`non-PM`)) %>% 
    filter(cd4_or_cd8 == "CD8") %>%
    ggplot(aes(x = treatment, 
               y = prcnt,
               fill = treatment,
               color = treatment)) + 
    geom_violin(alpha = 0.5, draw_quantiles = 0.5,
                scale = "width",
                adjust = 1) + 
    geom_point(size = 2) + 
    facet_wrap(~tissue) + 
    theme_minimal() +
    theme(panel.spacing = unit(0,"lines"),
          axis.title.x = element_blank(),
          panel.grid = element_blank(),
          axis.ticks.y = element_line(),
          legend.position = "none",
          panel.border = element_rect(color = "black", 
                                      size = 0.5, 
                                      fill = NA)) +
    scale_fill_manual(values = custom_colors("treatments")) + 
    scale_color_manual(values = custom_colors("treatments"))

cd4s <- seurat_integrated@meta.data %>% 
    filter(tissue != "pancreas" & !is.na(pancreas_matching)) %>% 
    dplyr::select(treatment, tissue, cd4_or_cd8, mouse_id, pancreas_matching, clonotype_id) %>% 
    unique() %>%
    group_by(treatment, tissue, cd4_or_cd8, mouse_id, pancreas_matching) %>% 
    tally() %>% 
    cast(treatment+mouse_id+tissue+cd4_or_cd8~pancreas_matching, value = "n") %>% 
    group_by(treatment, cd4_or_cd8, mouse_id, tissue) %>% 
    mutate(prcnt = 100*PM/sum(PM,`non-PM`)) %>% 
    filter(cd4_or_cd8 == "CD4") %>%
    ggplot(aes(x = treatment, 
               y = prcnt,
               fill = treatment,
               color = treatment)) + 
    geom_violin(alpha = 0.5, draw_quantiles = 0.5,
                scale = "width",
                adjust = 1) + 
    geom_point(size = 2) + 
    facet_wrap(~tissue) + 
    theme_minimal() +
    theme(panel.spacing = unit(0,"lines"),
          axis.title.x = element_blank(),
          panel.grid = element_blank(),
          axis.ticks.y = element_line(),
          legend.position = "none",
          panel.border = element_rect(color = "black", 
                                      size = 0.5, 
                                      fill = NA)) +
    scale_fill_manual(values = custom_colors("treatments")) + 
    scale_color_manual(values = custom_colors("treatments"))

cd8s_pan <- seurat_integrated@meta.data %>% 
    filter(tissue == "pancreas" & !is.na(pancreas_matching)) %>% 
    dplyr::select(treatment, tissue, cd4_or_cd8, mouse_id, pancreas_matching, clonotype_id) %>% 
    unique() %>%
    group_by(treatment, tissue, cd4_or_cd8, mouse_id, pancreas_matching) %>% 
    tally() %>% 
    cast(treatment+mouse_id+tissue+cd4_or_cd8~pancreas_matching, value = "n") %>% 
    group_by(treatment, cd4_or_cd8, mouse_id, tissue) %>% 
    mutate(prcnt = 100*PM/sum(PM,`non-PM`)) %>% 
    filter(cd4_or_cd8 == "CD8") %>%
    ggplot(aes(x = treatment, 
               y = prcnt,
               fill = treatment,
               color = treatment)) + 
    geom_violin(alpha = 0.5, draw_quantiles = 0.5,
                scale = "width",
                adjust = 1) + 
    geom_point(size = 2) + 
    facet_wrap(~tissue) + 
    theme_minimal() +
    theme(panel.spacing = unit(0,"lines"),
          axis.title.x = element_blank(),
          panel.grid = element_blank(),
          axis.ticks.y = element_line(),
          legend.position = "none",
          panel.border = element_rect(color = "black", 
                                      size = 0.5, 
                                      fill = NA)) +
    scale_fill_manual(values = custom_colors("treatments")) + 
    scale_color_manual(values = custom_colors("treatments"))

cd4s_pan <- seurat_integrated@meta.data %>% 
    filter(tissue == "pancreas" & !is.na(pancreas_matching)) %>% 
    dplyr::select(treatment, tissue, cd4_or_cd8, mouse_id, pancreas_matching, clonotype_id) %>% 
    unique() %>%
    group_by(treatment, tissue, cd4_or_cd8, mouse_id, pancreas_matching) %>% 
    tally() %>% 
    cast(treatment+mouse_id+tissue+cd4_or_cd8~pancreas_matching, value = "n") %>% 
    group_by(treatment, cd4_or_cd8, mouse_id, tissue) %>% 
    mutate(prcnt = 100*PM/sum(PM,`non-PM`)) %>% 
    filter(cd4_or_cd8 == "CD4") %>%
    ggplot(aes(x = treatment, 
               y = prcnt,
               fill = treatment,
               color = treatment)) + 
    geom_violin(alpha = 0.5, draw_quantiles = 0.5,
                scale = "width",
                adjust = 1) + 
    geom_point(size = 2) + 
    facet_wrap(~tissue) + 
    theme_minimal() +
    theme(panel.spacing = unit(0,"lines"),
          axis.title.x = element_blank(),
          panel.grid = element_blank(),
          axis.ticks.y = element_line(),
          legend.position = "none",
          panel.border = element_rect(color = "black", 
                                      size = 0.5, 
                                      fill = NA)) +
    scale_fill_manual(values = custom_colors("treatments")) + 
    scale_color_manual(values = custom_colors("treatments"))

grid.arrange(cd4s_pan,cd4s, cd8s_pan, cd8s, nrow=2, widths = c(1,2))

```

```{r boxplot of clone size PM and non-PM in blood and pLN}
## BOXPLOT OF CLONE SIZE
p_boxplot_pLN <- seurat_integrated@meta.data %>% 
                  filter(!is.na(pancreas_matching) & tissue == "pLN" & celltype == "CD8") %>%
                  group_by(treatment, clonotype_id, pancreas_matching) %>%
                  tally() %>%
                  ggplot() + 
                  geom_boxplot(aes(x = treatment, 
                                   y = log10(n), 
                                   fill = treatment,
                                   color = treatment),
                               outlier.shape = NA, 
                               alpha = 0.5) +
                  geom_point(aes(x = treatment, 
                                 y = log10(n), 
                                 fill = treatment,
                                 color = treatment),
                             position = position_jitter(0.2),
                             alpha = 0.1) +
                  scale_fill_manual(values = custom_colors("treatments")) +
                  scale_color_manual(values = custom_colors("treatments")) +
                  theme_minimal() +
                  #scale_y_continuous(limits = c(0, 2.9)) +
                  theme(panel.grid = element_blank(),
                        axis.title.x = element_blank(),
                        legend.position = "none",
                        text = element_text(size = 15),
                        panel.spacing = unit(0, "cm"),
                        panel.border = element_rect(colour = "black", fill=NA, size=1)) +
                  labs(y = expression(log[10](clone~size)),
                       color = "Treatment",
                       fill = "Treatment") +
                  facet_wrap(~ pancreas_matching, scales = "free_y") +
                  add_pvalue(data.frame(seurat_integrated@meta.data %>% 
                                          filter(!is.na(pancreas_matching) & tissue == "pLN" & celltype == "CD8") %>%
                                          group_by(treatment, clonotype_id, pancreas_matching) %>%
                                          tally() %>% 
                                          group_by(pancreas_matching) %>%
                                          wilcox_test(n ~ treatment) %>% 
                                          add_xy_position() %>%
                                          mutate(y.position = log10(y.position + 1))),
                             label = "p.adj.signif",
                             step.increase = 0.05,
                             tip.length = 0,
                             label.size = 7) 

p_boxplot_bld <- seurat_integrated@meta.data %>% 
                  filter(!is.na(pancreas_matching) & tissue == "blood" & celltype == "CD8") %>%
                  group_by(treatment, clonotype_id, pancreas_matching) %>%
                  tally() %>%
                  ggplot() + 
                  geom_boxplot(aes(x = treatment, 
                                   y = log10(n), 
                                   fill = treatment,
                                   color = treatment),
                               outlier.shape = NA, 
                               alpha = 0.5) +
                  geom_point(aes(x = treatment, 
                                 y = log10(n), 
                                 fill = treatment,
                                 color = treatment),
                             position = position_jitter(0.2),
                             alpha = 0.1) +
                  scale_fill_manual(values = custom_colors("treatments")) +
                  scale_color_manual(values = custom_colors("treatments")) +
                  theme_minimal() +
                  #scale_y_continuous(limits = c(0, 2.9)) +
                  theme(panel.grid = element_blank(),
                        axis.title.x = element_blank(),
                        legend.position = "none",
                        text = element_text(size = 15),
                        panel.spacing = unit(0, "cm"),
                        panel.border = element_rect(colour = "black", fill=NA, size=1)) +
                  labs(y = expression(log[10](clone~size)),
                       color = "Treatment",
                       fill = "Treatment") +
                  facet_wrap(~ pancreas_matching, scales = "free_y") +
                  add_pvalue(data.frame(seurat_integrated@meta.data %>% 
                                          filter(!is.na(pancreas_matching) & tissue == "blood" & celltype == "CD8") %>%
                                          group_by(treatment, clonotype_id, pancreas_matching) %>%
                                          tally() %>% 
                                          group_by(pancreas_matching) %>%
                                          wilcox_test(n ~ treatment) %>% 
                                          add_xy_position() %>%
                                          mutate(y.position = log10(y.position + 1))),
                             label = "p.adj.signif",
                             step.increase = 0.05,
                             tip.length = 0,
                             label.size = 7) 

#by mouse
p_boxplot_bld_pm <- seurat_integrated@meta.data %>% 
    filter(!is.na(pancreas_matching) & tissue == "pancreas" & clonotype_tissue_groups %in% c("blood&pancreas", "pancreas") ) %>%
    group_by(treatment, mouse_id, pancreas_matching, cd4_or_cd8) %>%
    tally() %>% cast(treatment+mouse_id+cd4_or_cd8 ~ pancreas_matching, value = "n") %>% 
    mutate(prcnt = 100*PM/(PM+`non-PM`)) %>%
                  ggplot() + 
                  geom_violin(aes(x = treatment, 
                                   y = prcnt, 
                                   fill = treatment,
                                   color = treatment),
                               scale= "width", 
                               alpha = 0.5,
                              draw_quantiles = 0.5) +
                  geom_point(aes(x = treatment, 
                                 y = prcnt, 
                                 fill = treatment,
                                 color = treatment),
                             size = 3) +
                  scale_fill_manual(values = custom_colors("treatments")) +
                  scale_color_manual(values = custom_colors("treatments")) +
                  theme_minimal() +
                  #scale_y_continuous(limits = c(0, 2.9)) +
                  theme(panel.grid = element_blank(),
                        axis.title.x = element_blank(),
                        legend.position = "none",
                        text = element_text(size = 15),
                        panel.spacing = unit(0, "cm"),
                        panel.border = element_rect(colour = "black", fill=NA, size=1)) +
                  labs(y = "Blood PM T cells in pancreas (%)",
                       color = "Treatment",
                       fill = "Treatment") +
  facet_wrap(~ cd4_or_cd8)

## TOP 50 CLONOTYPES
cumulative_sums <- seurat_integrated@meta.data %>% filter(!is.na(pancreas_matching) & tissue == "pLN" & celltype == "CD8" & pancreas_matching == "PM") %>% 
    group_by(treatment, clonotype_id) %>% 
    tally() %>% 
    group_by(treatment) %>% 
    mutate(prcnt = n*100/sum(n)) %>% 
    group_by(treatment) %>% 
    arrange(treatment, desc(prcnt)) %>% 
    mutate(cs = cumsum(prcnt))

cumulative_sums$clonotype_id <- factor(cumulative_sums$clonotype_id, 
                                       levels = cumulative_sums$clonotype_id)

p_top50_clonos <- cumulative_sums %>% 
    group_by(treatment) %>% 
    slice_head(n = 20) %>% 
    ggplot(aes(x = clonotype_id, 
               y = cs, 
               color = treatment)) + 
    geom_point() + 
    geom_segment(aes(xend = clonotype_id, 
                     yend = cs, 
                     x = clonotype_id, 
                     y = 0, 
                     color = treatment)) + 
    scale_color_manual(values = custom_colors("treatments")) +
    scale_y_continuous(expand = c(0, 0)) +
    theme_minimal() +
    theme(panel.grid = element_blank(),
          panel.spacing.x = unit(0, "lines"),
          text = element_text(size = 15),
          axis.text.x = element_blank(),
          panel.border = element_rect(colour = "black", 
                                      fill = NA, 
                                      size = 1),
          legend.position = "none") +
    labs(x = "Top 50 clonotypes", 
         y = "Cumulative % CD8 TCRs") +
    facet_wrap(~ treatment, scales = "free_x")
p_top50_clonos

grid.arrange(p_boxplot_pLN, p_boxplot_bld, ncol = 1)
```

```{r stacked bar plots of terminally exhausted CD8+ T cells}
eff_clonos <- seurat_integrated@meta.data %>% 
                filter(tissue == "pancreas" & clonotype_tissue_groups == "pLN&blood&pancreas" & celltype == "CD8" & seurat_clusters == "CD8_eff") %>%
                pull(clonotype_id) %>% 
                unique()
mem_clonos <- seurat_integrated@meta.data %>% 
                filter(tissue == "pancreas" & clonotype_tissue_groups == "pLN&blood&pancreas" & celltype == "CD8" & seurat_clusters == "CD8_mem") %>%
                pull(clonotype_id) %>% 
                unique()
texh_clonos <- seurat_integrated@meta.data %>% 
                filter(tissue == "pancreas" & clonotype_tissue_groups == "pLN&blood&pancreas" & celltype == "CD8" & seurat_clusters == "CD8_texh") %>%
                pull(clonotype_id) %>% 
                unique()

p1 <- seurat_integrated@meta.data %>% 
  filter(clonotype_id %in% mem_clonos) %>% 
  group_by(tissue,seurat_clusters) %>% 
  tally() %>% 
  ggplot(aes(x=tissue, 
             y = n, 
             fill = seurat_clusters)) + 
  geom_bar(stat="identity", 
           position = "fill") + 
  theme_minimal() +
  theme(panel.grid = element_blank(),
        text = element_text(size = 10),
        axis.title.x = element_blank(),
        panel.spacing.x = unit(0, "cm"),
        legend.position = "none") +
  scale_fill_manual(values = custom_colors("clusters"))

p2 <- seurat_integrated@meta.data %>% 
  filter(clonotype_id %in% eff_clonos) %>% 
  group_by(tissue,seurat_clusters) %>% 
  tally() %>% 
  ggplot(aes(x=tissue, 
             y = n, 
             fill = seurat_clusters)) + 
  geom_bar(stat="identity", 
           position = "fill") + 
  theme_minimal() +
  theme(panel.grid = element_blank(),
        text = element_text(size = 10),
        axis.title.x = element_blank(),
        panel.spacing.x = unit(0, "cm"),
        legend.position = "none") +
  scale_fill_manual(values = custom_colors("clusters"))

p3 <- seurat_integrated@meta.data %>% 
  filter(clonotype_id %in% texh_clonos) %>% 
  group_by(tissue,seurat_clusters) %>% 
  tally() %>% 
  ggplot(aes(x=tissue, 
             y = n, 
             fill = seurat_clusters)) + 
  geom_bar(stat="identity", 
           position = "fill") + 
  theme_minimal() +
  theme(panel.grid = element_blank(),
        text = element_text(size = 10),
        axis.title.x = element_blank(),
        panel.spacing.x = unit(0, "cm"),
        legend.position = "none") +
  scale_fill_manual(values = custom_colors("clusters"))

grid.arrange(p1,p2,p3, nrow = 1)
```


Next, examine how clones in various Seurat clusters in the pancreas differ in transcriptional profile with shared clones in other tissues. Two gene signatures are examined and compared:
```{r line plots of gene signature changes}
#line plots of exhaustion signatures: choose each cluster in pancreas and track that clonotype through the other tissues
clonotypes_pancreas_means <- seurat_integrated@meta.data %>% 
                              filter(tissue == "pancreas" & clonotype_tissue_groups == "pLN&blood&pancreas" & celltype == "CD8") %>% 
                              group_by(seurat_clusters, clonotype_id, tissue) %>%
                              summarize(mean = mean(Miller2019_terminally_exh1, na.rm = T)) %>%
                              cast(seurat_clusters + clonotype_id ~ tissue, value = "mean")
clonotypes_periphery_means <- seurat_integrated@meta.data %>% 
                                filter(tissue != "pancreas" & clonotype_id %in% clonotypes_pancreas_means$clonotype_id) %>%
                                group_by(clonotype_id, tissue) %>%
                                summarize(mean = mean(Miller2019_terminally_exh1, na.rm = T)) %>%
                                cast(clonotype_id ~ tissue, value = "mean")
clonotypes_means <- merge(clonotypes_pancreas_means, clonotypes_periphery_means , by = "clonotype_id") %>% 
                      melt(id.vars = c("clonotype_id","seurat_clusters"))
clonotypes_means$variable <- factor(clonotypes_means$variable, levels = c("pLN","blood","pancreas"))
colnames(clonotypes_means)[which(colnames(clonotypes_means) == "variable")] <- "tissue"
colnames(clonotypes_means)[which(colnames(clonotypes_means) == "value")] <- "mean"

stat.test <- clonotypes_means %>% 
              filter(seurat_clusters
                     %in% c("CD8_mem", "CD8_eff", "CD8_texh")) %>% 
              group_by(seurat_clusters) %>% 
              t_test(mean ~ tissue, paired = T) %>%
              add_xy_position()


p_termexh_clonotypes <- clonotypes_means %>%
                          filter(seurat_clusters %in% c("CD8_eff", "CD8_mem", "CD8_texh")) %>% #remove clusters with few cells/not interesting
                          ggplot(aes(x = tissue, y = mean)) + 
                          geom_point(alpha = 0.05) +
                          geom_line(aes(group = clonotype_id), alpha = 0.2) + 
                          geom_violin(adjust = 1, 
                                      scale = "area",
                                      trim=F, 
                                      alpha = 0.2, 
                                      aes(color = tissue, 
                                          fill = tissue)) +
                          geom_boxplot(outlier.color = NA, 
                                       notch = T, 
                                       width = 0.2,
                                       alpha=0.5, 
                                       fatten=T, 
                                       aes(color = tissue,
                                           fill = tissue)) +
                          scale_color_manual(values = custom_colors("tissues")) +
                          scale_fill_manual(values = custom_colors("tissues")) +
                        #scale_y_continuous(limits = c(0,1)) +
                          theme_minimal() +
                          theme(panel.grid = element_blank(),
                                text = element_text(size = 10),
                                axis.title.x = element_blank(),
                                panel.spacing.x = unit(0, "cm"),
                                panel.border = element_rect(color = "black", fill=NA, size=1),
                                legend.position = "none") +
                          labs(y = expression("Terminally-exhausted score (Miller 2019)")) +
                          facet_wrap(~ seurat_clusters, ncol = 5) + 
                          add_pvalue(stat.test %>% filter(p.adj < 0.05))
#progenitor-like
clonotypes_pancreas_means <- seurat_integrated@meta.data %>% 
                              filter(tissue == "pancreas" & clonotype_tissue_groups == "pLN&blood&pancreas" & celltype == "CD8") %>% 
                              group_by(seurat_clusters, clonotype_id, tissue) %>%
                              summarize(mean = mean(Miller2019_progenitor_exh1, na.rm = T)) %>%
                              cast(seurat_clusters + clonotype_id ~ tissue, value = "mean")
clonotypes_periphery_means <- seurat_integrated@meta.data %>% 
                                filter(tissue != "pancreas" & clonotype_id %in% clonotypes_pancreas_means$clonotype_id) %>%
                                group_by(clonotype_id, tissue) %>%
                                summarize(mean = mean(Miller2019_progenitor_exh1, na.rm = T)) %>%
                                cast(clonotype_id ~ tissue, value = "mean")
clonotypes_means <- merge(clonotypes_pancreas_means, clonotypes_periphery_means , by = "clonotype_id") %>% 
                      melt(id.vars = c("clonotype_id","seurat_clusters"))
clonotypes_means$variable <- factor(clonotypes_means$variable, levels = c("pLN","blood","pancreas"))
colnames(clonotypes_means)[which(colnames(clonotypes_means) == "variable")] <- "tissue"
colnames(clonotypes_means)[which(colnames(clonotypes_means) == "value")] <- "mean"

stat.test <- clonotypes_means %>% 
              filter(seurat_clusters
                     %in% c("CD8_mem", "CD8_eff", "CD8_texh")) %>% 
              group_by(seurat_clusters) %>% 
              t_test(mean ~ tissue, paired = T) %>%
              add_xy_position()


p_progexh_clonotypes <- clonotypes_means %>%
                          filter(seurat_clusters %in% c("CD8_eff", "CD8_mem", "CD8_texh")) %>% #remove clusters with few cells/not interesting
                          ggplot(aes(x = tissue, y = mean)) + 
                          geom_point(alpha = 0.05) +
                          geom_line(aes(group = clonotype_id), alpha = 0.2) + 
                          geom_violin(adjust = 1, 
                                      scale = "area",
                                      trim=F, 
                                      alpha = 0.2, 
                                      aes(color = tissue, 
                                          fill = tissue)) +
                          geom_boxplot(outlier.color = NA, 
                                       notch = T, 
                                       width = 0.2,
                                       alpha=0.5, 
                                       fatten=T, 
                                       aes(color = tissue,
                                           fill = tissue)) +
                          scale_color_manual(values = custom_colors("tissues")) +
                          scale_fill_manual(values = custom_colors("tissues")) +
                        #scale_y_continuous(limits = c(0,1)) +
                          theme_minimal() +
                          theme(panel.grid = element_blank(),
                                text = element_text(size = 10),
                                axis.title.x = element_blank(),
                                panel.spacing.x = unit(0, "cm"),
                                panel.border = element_rect(color = "black", fill=NA, size=1),
                                legend.position = "none") +
                          labs(y = expression("Progenitor-exhausted score (Miller 2019)")) +
                          facet_wrap(~ seurat_clusters, ncol = 5) + 
                          add_pvalue(stat.test %>% filter(p.adj < 0.05))


#effector-like
clonotypes_pancreas_means <- seurat_integrated@meta.data %>% 
                              filter(tissue == "pancreas" & clonotype_tissue_groups == "pLN&blood&pancreas" & celltype == "CD8") %>% 
                              group_by(seurat_clusters, clonotype_id, tissue) %>%
                              summarize(mean = mean(Miller2019_effector_like1, na.rm = T)) %>%
                              cast(seurat_clusters + clonotype_id ~ tissue, value = "mean")
clonotypes_periphery_means <- seurat_integrated@meta.data %>% 
                                filter(tissue != "pancreas" & clonotype_id %in% clonotypes_pancreas_means$clonotype_id) %>%
                                group_by(clonotype_id,tissue) %>%
                                summarize(mean = mean(Miller2019_effector_like1, na.rm = T)) %>%
                                cast(clonotype_id ~ tissue, value = "mean")
clonotypes_means <- merge(clonotypes_pancreas_means, clonotypes_periphery_means , by = "clonotype_id") %>% 
                      melt(id.vars = c("clonotype_id","seurat_clusters"))
clonotypes_means$variable <- factor(clonotypes_means$variable, levels = c("pLN","blood","pancreas"))
colnames(clonotypes_means)[which(colnames(clonotypes_means) == "variable")] <- "tissue"
colnames(clonotypes_means)[which(colnames(clonotypes_means) == "value")] <- "mean"

stat.test <- clonotypes_means %>% 
              filter(seurat_clusters
                     %in% c("CD8_mem", "CD8_eff", "CD8_texh")) %>% 
              group_by(seurat_clusters) %>% 
              t_test(mean ~ tissue, paired = T) %>%
              add_xy_position()


p_eff_clonotypes <- clonotypes_means %>%
                      filter(seurat_clusters %in% c("CD8_eff", "CD8_mem", "CD8_texh")) %>% #remove clusters with few cells/not interesting
                      ggplot(aes(x = tissue, y = mean)) + 
                      geom_point(alpha = 0.05) +
                      geom_line(aes(group = clonotype_id), alpha = 0.2) + 
                      geom_violin(adjust = 1, 
                                  scale = "area",
                                  trim=F, 
                                  alpha = 0.2, 
                                  aes(color = tissue, 
                                      fill = tissue)) +
                      geom_boxplot(outlier.color = NA, 
                                   notch = T, 
                                   width = 0.2,
                                   alpha=0.5, 
                                   fatten=T, 
                                   aes(color = tissue,
                                       fill = tissue)) +
                      scale_color_manual(values = custom_colors("tissues")) +
                      scale_fill_manual(values = custom_colors("tissues")) +
                    #scale_y_continuous(limits = c(0,1)) +
                      theme_minimal() +
                      theme(panel.grid = element_blank(),
                            text = element_text(size = 10),
                            axis.title.x = element_blank(),
                            panel.spacing.x = unit(0, "cm"),
                            panel.border = element_rect(color = "black", fill=NA, size=1),
                            legend.position = "none") +
                      labs(y = expression("Effector-like score (Miller 2019)")) +
                      facet_wrap(~ seurat_clusters, ncol = 5) + 
                      add_pvalue(stat.test %>% filter(p.adj < 0.05))

#memory-vs-effector
clonotypes_pancreas_means <- seurat_integrated@meta.data %>% 
                              filter(tissue == "pancreas" & clonotype_tissue_groups == "pLN&blood&pancreas" & celltype == "CD8") %>% 
                              group_by(seurat_clusters, clonotype_id, tissue) %>%
                              summarize(mean = mean(GOLDRATH_EFF_VS_MEMORY_CD8_TCELL_DN1, na.rm = T)) %>%
                              cast(seurat_clusters + clonotype_id ~ tissue, value = "mean")
clonotypes_periphery_means <- seurat_integrated@meta.data %>% 
                                filter(tissue != "pancreas" & clonotype_id %in% clonotypes_pancreas_means$clonotype_id) %>%
                                group_by(clonotype_id,tissue) %>%
                                summarize(mean = mean(GOLDRATH_EFF_VS_MEMORY_CD8_TCELL_DN1, na.rm = T)) %>%
                                cast(clonotype_id ~ tissue, value = "mean")
clonotypes_means <- merge(clonotypes_pancreas_means, clonotypes_periphery_means , by = "clonotype_id") %>% 
                      melt(id.vars = c("clonotype_id","seurat_clusters"))
clonotypes_means$variable <- factor(clonotypes_means$variable, levels = c("pLN","blood","pancreas"))
colnames(clonotypes_means)[which(colnames(clonotypes_means) == "variable")] <- "tissue"
colnames(clonotypes_means)[which(colnames(clonotypes_means) == "value")] <- "mean"

stat.test <- clonotypes_means %>% 
              filter(seurat_clusters
                     %in% c("CD8_mem", "CD8_eff", "CD8_texh")) %>% 
              group_by(seurat_clusters) %>% 
              t_test(mean ~ tissue, paired = T) %>%
              add_xy_position()


p_mem_clonotypes <- clonotypes_means %>%
                      filter(seurat_clusters %in% c("CD8_eff", "CD8_mem", "CD8_texh")) %>% #remove clusters with few cells/not interesting
                      ggplot(aes(x = tissue, y = mean)) + 
                      geom_point(alpha = 0.05) +
                      geom_line(aes(group = clonotype_id), alpha = 0.2) + 
                      geom_violin(adjust = 1, 
                                  scale = "area",
                                  trim=F, 
                                  alpha = 0.2, 
                                  aes(color = tissue, 
                                      fill = tissue)) +
                      geom_boxplot(outlier.color = NA, 
                                   notch = T, 
                                   width = 0.2,
                                   alpha=0.5, 
                                   fatten=T, 
                                   aes(color = tissue,
                                       fill = tissue)) +
                      scale_color_manual(values = custom_colors("tissues")) +
                      scale_fill_manual(values = custom_colors("tissues")) +
                    #scale_y_continuous(limits = c(0,1)) +
                      theme_minimal() +
                      theme(panel.grid = element_blank(),
                            text = element_text(size = 10),
                            axis.title.x = element_blank(),
                            panel.spacing.x = unit(0, "cm"),
                            panel.border = element_rect(color = "black", fill=NA, size=1),
                            legend.position = "none") +
                      labs(y = expression("Memory vs. Effector (Goldrath)")) +
                      facet_wrap(~ seurat_clusters, ncol = 5)+ 
                      add_pvalue(stat.test %>% filter(p.adj < 0.05))

# COINHIBITORY RECEPTOR EXPRESSION
clonotypes_pancreas_means <- seurat_integrated@meta.data %>% 
                              filter(tissue == "pancreas" & clonotype_tissue_groups == "pLN&blood&pancreas" & celltype == "CD8") %>% 
                              group_by(seurat_clusters, clonotype_id, tissue) %>%
                              summarize(mean = mean(coinh_count, na.rm = T)) %>%
                              cast(seurat_clusters + clonotype_id ~ tissue, value = "mean")
clonotypes_periphery_means <- seurat_integrated@meta.data %>% 
                                filter(tissue != "pancreas" & clonotype_id %in% clonotypes_pancreas_means$clonotype_id) %>%
                                group_by(clonotype_id,tissue) %>%
                                summarize(mean = mean(coinh_count, na.rm = T)) %>%
                                cast(clonotype_id ~ tissue, value = "mean")
clonotypes_means <- merge(clonotypes_pancreas_means, clonotypes_periphery_means , by = "clonotype_id") %>% 
                      melt(id.vars = c("clonotype_id","seurat_clusters"))
clonotypes_means$variable <- factor(clonotypes_means$variable, levels = c("pLN","blood","pancreas"))
colnames(clonotypes_means)[which(colnames(clonotypes_means) == "variable")] <- "tissue"
colnames(clonotypes_means)[which(colnames(clonotypes_means) == "value")] <- "mean"

stat.test <- clonotypes_means %>% 
              filter(seurat_clusters
                     %in% c("CD8_mem", "CD8_eff", "CD8_texh")) %>% 
              group_by(seurat_clusters) %>% 
              t_test(mean ~ tissue, paired = T) %>%
              add_xy_position()

p_coinh_clonotypes <- clonotypes_means %>%
                        filter(seurat_clusters %in% c("CD8_eff", "CD8_mem", "CD8_texh")) %>% #remove clusters with few cells/not interesting
                        ggplot(aes(x = tissue, y = mean)) + 
                        geom_point(alpha = 0.05) +
                        geom_line(aes(group = clonotype_id), alpha = 0.2) + 
                        geom_violin(adjust = 1, 
                                    scale = "area",
                                    trim=F, 
                                    alpha = 0.2, 
                                    aes(color = tissue, 
                                        fill = tissue)) +
                        geom_boxplot(outlier.color = NA, 
                                     notch = T, 
                                     width = 0.2,
                                     alpha=0.5, 
                                     fatten=T, 
                                     aes(color = tissue,
                                         fill = tissue)) +
                        scale_color_manual(values = custom_colors("tissues")) +
                        scale_fill_manual(values = custom_colors("tissues")) +
                        #scale_y_continuous(limits = c(0,1)) +
                        theme_minimal() +
                        theme(panel.grid = element_blank(),
                              text = element_text(size = 10),
                              axis.title.x = element_blank(),
                              panel.spacing.x = unit(0, "cm"),
                              panel.border = element_rect(color = "black", fill=NA, size=1),
                              legend.position = "none") +
                        labs(y = expression("Average co-inhibitory receptors expressed")) +
                        facet_wrap(~ seurat_clusters, ncol = 5) + 
                        add_pvalue(stat.test %>% filter(p.adj < 0.05))

grid.arrange(p_coinh_clonotypes, p_termexh_clonotypes, p_progexh_clonotypes, p_eff_clonotypes, p_mem_clonotypes, ncol = 1)
```

Show the line plots across different treatment groups.
```{r line plots of gene signature changes}
#line plots of exhaustion signatures: choose each cluster in pancreas and track that clonotype through the other tissues
clonotypes_pancreas_means <- seurat_integrated@meta.data %>% 
                              filter(tissue == "pancreas" & clonotype_tissue_groups == "pLN&blood&pancreas") %>% 
                              group_by(treatment, clonotype_id, tissue) %>%
                              summarize(mean = mean(Miller2019_terminally_exh1, na.rm = T)) %>%
                              cast(treatment + clonotype_id ~ tissue, value = "mean")
clonotypes_periphery_means <- seurat_integrated@meta.data %>% 
                                filter(tissue != "pancreas" & clonotype_id %in% clonotypes_pancreas_means$clonotype_id) %>%
                                group_by(treatment, clonotype_id, tissue) %>%
                                summarize(mean = mean(Miller2019_terminally_exh1, na.rm = T)) %>%
                                cast(treatment + clonotype_id ~ tissue, value = "mean")
clonotypes_means <- merge(clonotypes_pancreas_means, clonotypes_periphery_means , by = c("clonotype_id","treatment")) %>% 
                      melt(id.vars = c("clonotype_id","treatment"))
clonotypes_means$variable <- factor(clonotypes_means$variable, levels = c("pLN","blood","pancreas"))
colnames(clonotypes_means)[which(colnames(clonotypes_means) == "variable")] <- "tissue"
colnames(clonotypes_means)[which(colnames(clonotypes_means) == "value")] <- "mean"

p_termexh_clonotypes <- clonotypes_means %>%
                          ggplot(aes(x = tissue, y = mean)) + 
                          geom_point(alpha = 0.05) +
                          geom_line(aes(group = clonotype_id), alpha = 0.2) + 
                          geom_violin(adjust = 1, 
                                      scale = "area",
                                      trim=F, 
                                      alpha = 0.2, 
                                      aes(color = tissue, 
                                          fill = tissue)) +
                          geom_boxplot(outlier.color = NA, 
                                       notch = T, 
                                       width = 0.2,
                                       alpha=0.5, 
                                       fatten=T, 
                                       aes(fill = tissue)) +
                          scale_color_manual(values = custom_colors("tissues")) +
                          scale_fill_manual(values = custom_colors("tissues")) +
                        #scale_y_continuous(limits = c(0,1)) +
                          theme_minimal() +
                          theme(panel.grid = element_blank(),
                                text = element_text(size = 10),
                                axis.title.x = element_blank(),
                                panel.spacing.x = unit(0, "cm"),
                                panel.border = element_rect(color = "black", fill=NA, size=1),
                                legend.position = "none") +
                          labs(y = expression("Terminally-exhausted score (Miller 2019)")) +
                        facet_wrap(~ treatment, ncol = 5)

clonotypes_pancreas_means <- seurat_integrated@meta.data %>% 
                              filter(tissue == "pancreas" & clonotype_tissue_groups == "pLN&blood&pancreas") %>% 
                              group_by(treatment, clonotype_id, tissue) %>%
                              summarize(mean = mean(Miller2019_effector_like1, na.rm = T)) %>%
                              cast(treatment + clonotype_id ~ tissue, value = "mean")
clonotypes_periphery_means <- seurat_integrated@meta.data %>% 
                                filter(tissue != "pancreas" & clonotype_id %in% clonotypes_pancreas_means$clonotype_id) %>%
                                group_by(treatment, clonotype_id, tissue) %>%
                                summarize(mean = mean(Miller2019_effector_like1, na.rm = T)) %>%
                                cast(treatment + clonotype_id ~ tissue, value = "mean")
clonotypes_means <- merge(clonotypes_pancreas_means, clonotypes_periphery_means , by = c("clonotype_id","treatment")) %>% 
                      melt(id.vars = c("clonotype_id","treatment"))
clonotypes_means$variable <- factor(clonotypes_means$variable, levels = c("pLN","blood","pancreas"))
colnames(clonotypes_means)[which(colnames(clonotypes_means) == "variable")] <- "tissue"
colnames(clonotypes_means)[which(colnames(clonotypes_means) == "value")] <- "mean"

p_eff_clonotypes <- clonotypes_means %>%
                      ggplot(aes(x = tissue, y = mean)) + 
                      geom_point(alpha = 0.05) +
                      geom_line(aes(group = clonotype_id), alpha = 0.2) + 
                      geom_violin(adjust = 1, 
                                  scale = "area",
                                  trim=F, 
                                  alpha = 0.2, 
                                  aes(color = tissue, 
                                      fill = tissue)) +
                      geom_boxplot(outlier.color = NA, 
                                   notch = T, 
                                   width = 0.2,
                                   alpha=0.5, 
                                   fatten=T, 
                                   aes(fill = tissue)) +
                      scale_color_manual(values = custom_colors("tissues")) +
                      scale_fill_manual(values = custom_colors("tissues")) +
                    #scale_y_continuous(limits = c(0,1)) +
                      theme_minimal() +
                      theme(panel.grid = element_blank(),
                            text = element_text(size = 10),
                            axis.title.x = element_blank(),
                            panel.spacing.x = unit(0, "cm"),
                            panel.border = element_rect(color = "black", fill=NA, size=1),
                            legend.position = "none") +
                      labs(y = expression("Effector-like score (Miller 2019)")) +
                    facet_wrap(~ treatment, ncol = 5)

#memory-vs-effector
clonotypes_pancreas_means <- seurat_integrated@meta.data %>% 
                              filter(tissue == "pancreas" & clonotype_tissue_groups == "pLN&blood&pancreas") %>% 
                              group_by(treatment, clonotype_id, tissue) %>%
                              summarize(mean = mean(GOLDRATH_EFF_VS_MEMORY_CD8_TCELL_DN1, na.rm = T)) %>%
                              cast(treatment + clonotype_id ~ tissue, value = "mean")
clonotypes_periphery_means <- seurat_integrated@meta.data %>% 
                                filter(tissue != "pancreas" & clonotype_id %in% clonotypes_pancreas_means$clonotype_id) %>%
                                group_by(treatment, clonotype_id, tissue) %>%
                                summarize(mean = mean(GOLDRATH_EFF_VS_MEMORY_CD8_TCELL_DN1, na.rm = T)) %>%
                                cast(treatment + clonotype_id ~ tissue, value = "mean")
clonotypes_means <- merge(clonotypes_pancreas_means, clonotypes_periphery_means , by = c("clonotype_id","treatment")) %>% 
                      melt(id.vars = c("clonotype_id","treatment"))
clonotypes_means$variable <- factor(clonotypes_means$variable, levels = c("pLN","blood","pancreas"))
colnames(clonotypes_means)[which(colnames(clonotypes_means) == "variable")] <- "tissue"
colnames(clonotypes_means)[which(colnames(clonotypes_means) == "value")] <- "mean"

p_mem_clonotypes <- clonotypes_means %>%
                      ggplot(aes(x = tissue, y = mean)) + 
                      geom_point(alpha = 0.05) +
                      geom_line(aes(group = clonotype_id), alpha = 0.2) + 
                      geom_violin(adjust = 1, 
                                  scale = "area",
                                  trim=F, 
                                  alpha = 0.2, 
                                  aes(color = tissue, 
                                      fill = tissue)) +
                      geom_boxplot(outlier.color = NA, 
                                   notch = T, 
                                   width = 0.2,
                                   alpha=0.5, 
                                   fatten=T, 
                                   aes(fill = tissue)) +
                      scale_color_manual(values = custom_colors("tissues")) +
                      scale_fill_manual(values = custom_colors("tissues")) +
                    #scale_y_continuous(limits = c(0,1)) +
                      theme_minimal() +
                      theme(panel.grid = element_blank(),
                            text = element_text(size = 10),
                            axis.title.x = element_blank(),
                            panel.spacing.x = unit(0, "cm"),
                            panel.border = element_rect(color = "black", fill=NA, size=1),
                            legend.position = "none") +
                      labs(y = expression("Memory vs. Effector (Goldrath)")) +
                    facet_wrap(~ treatment, ncol = 5)

grid.arrange(p_termexh_clonotypes, p_eff_clonotypes, p_mem_clonotypes, ncol = 1)
#show UMAPs of certain clonotypes as examples

# COINHIBITORY RECEPTOR EXPRESSION
clonotypes_pancreas_means <- seurat_integrated@meta.data %>% 
                              filter(tissue == "pancreas" & clonotype_tissue_groups == "pLN&blood&pancreas") %>% 
                              group_by(treatment, clonotype_id, tissue) %>%
                              summarize(mean = mean(coinh_count, na.rm = T)) %>%
                              cast(treatment + clonotype_id ~ tissue, value = "mean")
clonotypes_periphery_means <- seurat_integrated@meta.data %>% 
                                filter(tissue != "pancreas" & clonotype_id %in% clonotypes_pancreas_means$clonotype_id) %>%
                                group_by(treatment, clonotype_id, tissue) %>%
                                summarize(mean = mean(coinh_count, na.rm = T)) %>%
                                cast(treatment + clonotype_id ~ tissue, value = "mean")
clonotypes_means <- merge(clonotypes_pancreas_means, clonotypes_periphery_means , by = c("clonotype_id","treatment")) %>% 
                      melt(id.vars = c("clonotype_id","treatment"))
clonotypes_means$variable <- factor(clonotypes_means$variable, levels = c("pLN","blood","pancreas"))
colnames(clonotypes_means)[which(colnames(clonotypes_means) == "variable")] <- "tissue"
colnames(clonotypes_means)[which(colnames(clonotypes_means) == "value")] <- "mean"

p_coinh_clonotypes <- clonotypes_means %>%
                      ggplot(aes(x = tissue, y = mean)) + 
                      geom_point(alpha = 0.05) +
                      geom_line(aes(group = clonotype_id), alpha = 0.2) + 
                      geom_violin(adjust = 1, 
                                  scale = "area",
                                  trim=F, 
                                  alpha = 0.2, 
                                  aes(color = tissue, 
                                      fill = tissue)) +
                      geom_boxplot(outlier.color = NA, 
                                   notch = T, 
                                   width = 0.2,
                                   alpha=0.5, 
                                   fatten=T, 
                                   aes(fill = tissue)) +
                      scale_color_manual(values = custom_colors("tissues")) +
                      scale_fill_manual(values = custom_colors("tissues")) +
                    #scale_y_continuous(limits = c(0,1)) +
                      theme_minimal() +
                      theme(panel.grid = element_blank(),
                            text = element_text(size = 10),
                            axis.title.x = element_blank(),
                            panel.spacing.x = unit(0, "cm"),
                            panel.border = element_rect(color = "black", fill=NA, size=1),
                            legend.position = "none") +
                      labs(y = expression("Average co-inhibitory receptors expressed")) +
                    facet_wrap(~ treatment, ncol = 5)
p_coinh_clonotypes
```

Plot some example UMAPs of shared clusters.
```{r example UMAPs of clonotypes across tissues}
seurat_integrated@meta.data %>% 
                      filter(clonotype_id %in% paste0("clonotype", c(1,3:4))) %>%
                      ggplot(aes(x = allcells_UMAP_1, 
                                 y = allcells_UMAP_2, 
                                 color = seurat_clusters)) + 
                      geom_point(data = seurat_integrated@meta.data %>% 
                                     dplyr::select(-treatment), 
                                 shape = 16, 
                                 color = "grey90") +
                      geom_point(shape = 16, 
                                 alpha = 1) +
                      theme_minimal() +
                      theme(axis.title = element_blank(),
                            axis.text = element_blank(),
                            panel.grid = element_blank(),
                            text = element_text(size = 15),
                            legend.position = "none") +
                      labs(x = 'UMAP 1', 
                           y = 'UMAP 2', 
                           color=NULL,
                           title = "Top clonotype shared with pancreas") +
                      scale_alpha_continuous(range = c(0.4 ,1)) +
                      scale_color_manual(values = custom_colors("clusters")) +
                      scale_size_continuous(range=c(1, 3)) +
                      guides(color = guide_legend(override.aes = list(alpha = 1, 
                                                                      size = 5, 
                                                                      shape = 19)), 
                             size = guide_legend()) +
                      facet_wrap(~treatment)
```

Track the NRP-v7+ tcrtypes across tissues:
```{r track NRP-v7+ tcrtypes}
#terminally exhausted
tcrtypes_pancreas_means <- seurat_integrated@meta.data %>% 
                              filter(tissue == "pancreas" & clonotype_tissue_groups == "pLN&blood&pancreas" & tetramer_specificity == "NRPv7") %>% 
                              group_by(seurat_clusters, tcrtype_id, tissue) %>%
                              summarize(mean = mean(Miller2019_terminally_exh1, na.rm = T)) %>%
                              cast(seurat_clusters + tcrtype_id ~ tissue, value = "mean")
tcrtypes_periphery_means <- seurat_integrated@meta.data %>% 
                                filter(tissue != "pancreas" & tcrtype_id %in% tcrtypes_pancreas_means$tcrtype_id) %>%
                                group_by(tcrtype_id, tissue) %>%
                                summarize(mean = mean(Miller2019_terminally_exh1, na.rm = T)) %>%
                                cast(tcrtype_id ~ tissue, value = "mean")
tcrtypes_means <- merge(tcrtypes_pancreas_means, tcrtypes_periphery_means , by = "tcrtype_id") %>% 
                      melt(id.vars = c("tcrtype_id","seurat_clusters"))
tcrtypes_means$variable <- factor(tcrtypes_means$variable, levels = c("pLN","blood","pancreas"))
colnames(tcrtypes_means)[which(colnames(tcrtypes_means) == "variable")] <- "tissue"
colnames(tcrtypes_means)[which(colnames(tcrtypes_means) == "value")] <- "mean"

ptet_termexh_tcrtypes <- tcrtypes_means %>%
                          filter(seurat_clusters %in% c("CD8_eff_Cxcr6", "CD8_mem", "CD8_texh", "Acinar_contaminated")) %>% #remove clusters with few cells/not interesting
                          ggplot(aes(x = tissue, y = mean)) + 
                          geom_point(alpha = 0.05) +
                          geom_line(aes(group = tcrtype_id), alpha = 0.2) + 
                          geom_violin(adjust = 1, 
                                      scale = "area",
                                      trim=F, 
                                      alpha = 0.2, 
                                      aes(color = tissue, 
                                          fill = tissue)) +
                          geom_boxplot(outlier.color = NA, 
                                       notch = T, 
                                       width = 0.2,
                                       alpha=0.5, 
                                       fatten=T, 
                                       aes(fill = tissue)) +
                          scale_color_manual(values = custom_colors("tissues")) +
                          scale_fill_manual(values = custom_colors("tissues")) +
                        #scale_y_continuous(limits = c(0,1)) +
                          theme_minimal() +
                          theme(panel.grid = element_blank(),
                                text = element_text(size = 10),
                                axis.title.x = element_blank(),
                                panel.spacing.x = unit(0, "cm"),
                                panel.border = element_rect(color = "black", fill=NA, size=1),
                                legend.position = "none") +
                          labs(y = expression("Terminally-exhausted score (Miller 2019)")) +
                        facet_wrap(~ seurat_clusters, ncol = 4)

#Miller 2019 effector-like
tcrtypes_pancreas_means <- seurat_integrated@meta.data %>% 
                              filter(tissue == "pancreas" & clonotype_tissue_groups == "pLN&blood&pancreas" & tetramer_specificity == "NRPv7") %>% 
                              group_by(seurat_clusters, tcrtype_id, tissue) %>%
                              summarize(mean = mean(Miller2019_effector_like1, na.rm = T)) %>%
                              cast(seurat_clusters + tcrtype_id ~ tissue, value = "mean")
tcrtypes_periphery_means <- seurat_integrated@meta.data %>% 
                                filter(tissue != "pancreas" & tcrtype_id %in% tcrtypes_pancreas_means$tcrtype_id) %>%
                                group_by(tcrtype_id, tissue) %>%
                                summarize(mean = mean(Miller2019_effector_like1, na.rm = T)) %>%
                                cast(tcrtype_id ~ tissue, value = "mean")
tcrtypes_means <- merge(tcrtypes_pancreas_means, tcrtypes_periphery_means , by = "tcrtype_id") %>% 
                      melt(id.vars = c("tcrtype_id","seurat_clusters"))
tcrtypes_means$variable <- factor(tcrtypes_means$variable, levels = c("pLN","blood","pancreas"))
colnames(tcrtypes_means)[which(colnames(tcrtypes_means) == "variable")] <- "tissue"
colnames(tcrtypes_means)[which(colnames(tcrtypes_means) == "value")] <- "mean"

ptet_eff_tcrtypes <- tcrtypes_means %>%
                      filter(seurat_clusters %in% c("CD8_eff_Cxcr6", "CD8_mem", "CD8_texh", "Acinar_contaminated")) %>% #remove clusters with few cells/not interesting
                      ggplot(aes(x = tissue, y = mean)) + 
                      geom_point(alpha = 0.05) +
                      geom_line(aes(group = tcrtype_id), alpha = 0.2) + 
                      geom_violin(adjust = 1, 
                                  scale = "area",
                                  trim=F, 
                                  alpha = 0.2, 
                                  aes(color = tissue, 
                                      fill = tissue)) +
                      geom_boxplot(outlier.color = NA, 
                                   notch = T, 
                                   width = 0.2,
                                   alpha=0.5, 
                                   fatten=T, 
                                   aes(fill = tissue)) +
                      scale_color_manual(values = custom_colors("tissues")) +
                      scale_fill_manual(values = custom_colors("tissues")) +
                    #scale_y_continuous(limits = c(0,1)) +
                      theme_minimal() +
                      theme(panel.grid = element_blank(),
                            text = element_text(size = 10),
                            axis.title.x = element_blank(),
                            panel.spacing.x = unit(0, "cm"),
                            panel.border = element_rect(color = "black", fill=NA, size=1),
                            legend.position = "none") +
                      labs(y = expression("Effector-like (Miller 2019)")) +
                    facet_wrap(~ seurat_clusters, ncol = 4)

#GOLDRATH memory vs effector
tcrtypes_pancreas_means <- seurat_integrated@meta.data %>% 
                              filter(tissue == "pancreas" & clonotype_tissue_groups == "pLN&blood&pancreas" & tetramer_specificity == "NRPv7") %>% 
                              group_by(seurat_clusters, tcrtype_id, tissue) %>%
                              summarize(mean = mean(GOLDRATH_EFF_VS_MEMORY_CD8_TCELL_DN1, na.rm = T)) %>%
                              cast(seurat_clusters + tcrtype_id ~ tissue, value = "mean")
tcrtypes_periphery_means <- seurat_integrated@meta.data %>% 
                                filter(tissue != "pancreas" & tcrtype_id %in% tcrtypes_pancreas_means$tcrtype_id) %>%
                                group_by(tcrtype_id, tissue) %>%
                                summarize(mean = mean(GOLDRATH_EFF_VS_MEMORY_CD8_TCELL_DN1, na.rm = T)) %>%
                                cast(tcrtype_id ~ tissue, value = "mean")
tcrtypes_means <- merge(tcrtypes_pancreas_means, tcrtypes_periphery_means , by = "tcrtype_id") %>% 
                      melt(id.vars = c("tcrtype_id","seurat_clusters"))
tcrtypes_means$variable <- factor(tcrtypes_means$variable, levels = c("pLN","blood","pancreas"))
colnames(tcrtypes_means)[which(colnames(tcrtypes_means) == "variable")] <- "tissue"
colnames(tcrtypes_means)[which(colnames(tcrtypes_means) == "value")] <- "mean"

ptet_mem_tcrtypes <- tcrtypes_means %>%
                      filter(seurat_clusters %in% c("CD8_eff_Cxcr6", "CD8_mem", "CD8_texh", "Acinar_contaminated")) %>% #remove clusters with few cells/not interesting
                      ggplot(aes(x = tissue, y = mean)) + 
                      geom_point(alpha = 0.05) +
                      geom_line(aes(group = tcrtype_id), alpha = 0.2) + 
                      geom_violin(adjust = 1, 
                                  scale = "area",
                                  trim=F, 
                                  alpha = 0.2, 
                                  aes(color = tissue, 
                                      fill = tissue)) +
                      geom_boxplot(outlier.color = NA, 
                                   notch = T, 
                                   width = 0.2,
                                   alpha=0.5, 
                                   fatten=T, 
                                   aes(fill = tissue)) +
                      scale_color_manual(values = custom_colors("tissues")) +
                      scale_fill_manual(values = custom_colors("tissues")) +
                    #scale_y_continuous(limits = c(0,1)) +
                      theme_minimal() +
                      theme(panel.grid = element_blank(),
                            text = element_text(size = 10),
                            axis.title.x = element_blank(),
                            panel.spacing.x = unit(0, "cm"),
                            panel.border = element_rect(color = "black", fill=NA, size=1),
                            legend.position = "none") +
                      labs(y = expression("Memory vs. Effector (Goldrath)")) +
                    facet_wrap(~ seurat_clusters, ncol = 4)

grid.arrange(ptet_termexh_tcrtypes, ptet_eff_tcrtypes, ptet_mem_tcrtypes, ncol = 1)
```

Upset plot to show the sharing of clones between Treg and Tcon:
```{r clonotypes shared between Treg and Tcon}
p_tcon_treg_shared_clonotypes <- seurat_integrated@meta.data %>% 
                                  filter(!is.na(clonotype_celltype_groups)) %>%
                                  ggplot(aes(x = allcells_UMAP_1, 
                                           y = allcells_UMAP_2, 
                                           color = seurat_clusters)) + 
                                  geom_point(data = seurat_integrated@meta.data %>%
                                               dplyr::select(-clonotype_celltype_groups), 
                                               shape = 16,
                                               color = "grey90") +
                                  geom_point(shape = 16) +
                                  theme_minimal() +
                                  theme(axis.title = element_blank(),
                                        axis.text = element_blank(),
                                        strip.text = element_blank(),
                                        text = element_text(size = 15),
                                        panel.grid = element_blank(),
                                        legend.position = "none") +
                                  labs(x = 'UMAP 1', 
                                       y = 'UMAP 2', 
                                       color = "Seurat clusters",
                                       size = "Size of clonotype\n(pancreas only)") +
                                  facet_wrap(~ clonotype_celltype_groups) +
                                  scale_color_manual(values = custom_colors("clusters"),
                                                     limits = force) #drops unused clusters in legend
p_tcon_treg_shared_clonotypes

df <- seurat_integrated@meta.data %>% 
        filter(clonotype_celltype_groups == "Tcon&Treg") %>% 
          group_by(treatment,mouse_id,clonotype_id,clonotype_count,seurat_clusters) %>% 
          unique() %>% 
          select(treatment,mouse_id,clonotype_id,clonotype_count,seurat_clusters) %>% 
          summarize() %>%
            mutate(index = T) %>% 
            pivot_wider(c(treatment,mouse_id,clonotype_id,clonotype_count), names_from = seurat_clusters, values_from = index) %>% 
            mutate_all(replace_na, F)
cluster_labels <- rev(levels(pan4$seurat_clusters)) #determines the order of the clusters in the table
df$num_unique_clusters <- rowSums(df[,5:15])
df$grouping <- apply(df, 1, function(x) paste(colnames(df)[5:15][x[5:15] == T], collapse = "&")) #add a column that concatenates all the seurat clusters that a clonotype includes
upset(df, cluster_labels, set_sizes = F, sort_sets = F, width_ratio = 0.1, min_size = 3,  keep_empty_groups=TRUE) +
    ggtitle("Treg & Tcon shared clusters") + 
  theme_minimal() +
  theme(text = element_text(size = 15))

#plot example UMAPs of individual clonotypes showing distribution of Tcon and Treg
seurat_integrated@meta.data %>% 
    filter(clonotype_celltype_groups == "Tcon&Treg") %>% 
    group_by(clonotype_id,celltype) %>% tally() %>%
    cast(clonotype_id ~ celltype, value = "n") %>% 
    mutate(total = Tcon + Treg) %>%
    arrange(desc(Treg), desc(total)) %>%
    head(20)

p_tcon_treg_shared_example_clonotypes <- seurat_integrated@meta.data %>% 
                            filter(clonotype_id %in% paste0("clonotype",c(59,68,65,86,87))) %>%
                            dplyr::select(treatment,
                                          clonotype_id, 
                                          seurat_clusters,
                                          allcells_UMAP_1, 
                                          allcells_UMAP_2) %>% 
                            group_by(treatment,
                                     clonotype_id, 
                                     seurat_clusters) %>% 
                            summarise(U1 = mean(allcells_UMAP_1),
                                      U2 = mean(allcells_UMAP_2),
                                      num_cells = n(),
                                      .groups = "keep") %>% 
                            ggplot() +
                            geom_point(data = seurat_integrated@meta.data %>%
                                         dplyr::select(-treatment) %>%
                                         dplyr::select(-clonotype_id),
                                       aes(x = allcells_UMAP_1, 
                                           y = allcells_UMAP_2),
                                       shape = 16,
                                       color = "grey90") +
                            geom_point(aes(x = U1,
                                           y = U2,
                                           size = num_cells,
                                           color = seurat_clusters),
                                       shape = 16) +
                            theme_minimal() +
                            theme(axis.title = element_blank(),
                                  axis.text = element_blank(),
                                  panel.grid = element_blank(),
                                  text = element_text(size = 15),
                                  legend.position = "none") +
                            labs(color = NULL) +
                            scale_color_manual(values = custom_colors("clusters")) +
                            scale_size_continuous(range = c(3,10)) +
                            guides(color = guide_legend(override.aes = list(alpha = 1, 
                                                                            size = 5, 
                                                                            shape = 19))) +
                            facet_wrap(treatment ~ clonotype_id, ncol = 9) +
                            geom_text(aes(x = U1, y = U2, label = num_cells), color = "white", size = 3, fontface = "bold")

p_tcon_treg_shared_example_clonotypes
```

Show how the tetramer+ T cells differ between the treatment groups
```{r andreas tetramer positive T cells}
seurat_integrated@meta.data %>% 
    filter(andreas_tetramer_specificity == "NRPv7" & celltype == "CD8") %>% 
    group_by(treatment, seurat_clusters) %>% 
    tally() %>% group_by(treatment) %>%     mutate(prcnt = n*100/sum(n),
                                                             csum = rev(cumsum(rev(prcnt))),  #get the position for the labels
                                                             pos = prcnt/2 + lead(csum, 1),
                                                             pos = if_else(is.na(pos), prcnt/2, pos)) %>%
    ggplot(aes(x = "", 
               y = prcnt, 
               fill = seurat_clusters)) + # pct used here so slices add to 100
    geom_bar(stat="identity", width=1) +
    coord_polar("y", start=0) +
    #geom_text(aes(label = paste0(round(prcnt, 0), "%")), position = position_stack(vjust = 0.5)) +
    facet_wrap(~treatment, ncol = 3) +
    theme_void() +
    theme(legend.position = "none",
          text = element_text(size = 15)) +
    geom_label_repel(aes(y = pos, label = paste0(round(prcnt, 0), "%")),
                     size = 4.5, nudge_x = 0.5, show.legend = FALSE) +
    scale_fill_manual(values = custom_colors("clusters"))

seurat_integrated@meta.data %>% 
  filter(andreas_tetramer_specificity == "NRPv7" & celltype == "CD8" & pancreas_matching == "PM") %>% 
  group_by(clonotype_id, treatment, tissue, seurat_clusters) %>% 
  tally() %>% 
  ggplot(aes(x=tissue, fill = seurat_clusters, y = n+1)) + 
  geom_bar(stat="identity", position = "stack") + 
  facet_wrap(treatment~clonotype_id) + 
  scale_fill_manual(values = custom_colors("clusters")) + 
  theme_minimal() + 
  scale_y_continuous(trans = "log10") +
  theme(legend.position = "none", 
        panel.grid = element_blank(), 
        axis.text = element_blank(), 
        axis.title = element_blank())
```

```{r pancreas matching across clusters}
rbind(seurat_integrated@meta.data %>% 
          filter(tissue == "blood" & !is.na(pancreas_matching) & celltype == "CD8") %>% 
          group_by(seurat_clusters, 
                   pancreas_matching, 
                   mouse_id, 
                   treatment) %>% 
          tally() %>% 
          group_by(mouse_id, 
                   treatment) %>% 
          mutate(prcnt = 100*n/sum(n)) %>% 
          cast(seurat_clusters + mouse_id + treatment ~ pancreas_matching, 
               value = "prcnt", 
               fill = 0) %>% 
          dplyr::select(seurat_clusters, treatment, PM),
      data.frame(seurat_clusters = rep("CD8_effmem", 5), 
                 treatment = c("IgG","Spt","PD1","PD1","PD1"), 
                 PM = rep(0,5))) %>%
    filter(seurat_clusters != "Acinar_contaminated") %>%
    ggplot(aes(x = factor(seurat_clusters, levels = rev(levels(seurat_integrated$seurat_clusters))), 
               y = PM, 
               color = treatment, 
               fill = treatment)) + 
    theme_minimal() +
    theme(legend.position = "none",
          axis.ticks.y = element_line(),
          panel.grid = element_blank(),
          panel.spacing = unit(0, "cm"),
          strip.text = element_blank(),
          panel.border = element_rect(color = "black", fill = NA)) +
    geom_violin(scale = "width", 
                adjust = 0.8,
                width = 0.6,
                alpha = 0.2, 
                draw_quantiles = c(0.5),
                position = position_dodge(0.9)) + 
    geom_point(position = position_dodge(0.9), size = 3) + 
    scale_fill_manual(values = custom_colors("treatments")) + 
    scale_color_manual(values = custom_colors("treatments")) + 
    scale_y_continuous(expand = c(0,0), limits = c(0,30)) +
    facet_wrap(~seurat_clusters, nrow = 1, scales = "free_x")
```


```{r show association between clusters}
shared_clonotypes <- seurat_integrated@meta.data %>% 
  filter(!is.na(clonotype_id)) %>%
  group_by(cd4_or_cd8, clonotype_id, seurat_clusters) %>% 
  tally() %>% 
  summarize(clonotype_cluster_groups = paste(sort(seurat_clusters), collapse = "&")) %>% #be sure to sort the seurat_clusters terms before collapsing
  arrange(desc(clonotype_cluster_groups)) %>% 
  data.frame()

shared_clonotypes <- setNames(as.character(shared_clonotypes$clonotype_cluster_groups), shared_clonotypes$clonotype_id)

new_metadata <- seurat_integrated@meta.data %>% 
  rownames_to_column('barcode') %>%
  mutate(clonotype_cluster_groups = shared_clonotypes[clonotype_id]) %>%
  column_to_rownames('barcode') %>%
  dplyr::select(clonotype_cluster_groups) 

seurat_integrated <- AddMetaData(seurat_integrated, new_metadata)

merge(seurat_integrated@meta.data %>% 
        filter(cd4_or_cd8 == "CD8" & !is.na(clonotype_id) & grepl("Proliferating", clonotype_cluster_groups)) %>% 
        dplyr::select(clonotype_id, seurat_clusters) %>% unique() %>% 
        group_by(seurat_clusters) %>% 
        tally() %>% 
        mutate(prcnt = round(100*n/sum(n))), 
      seurat_integrated@meta.data %>% 
        group_by(seurat_clusters) %>% 
        summarize(U1 = mean(allcells_UMAP_1),
                  U2 = mean(allcells_UMAP_2)), 
      by = "seurat_clusters") %>%
  ggplot() +
  geom_point(data = seurat_integrated@meta.data,
             aes(x = allcells_UMAP_1, 
                 y = allcells_UMAP_2),
             shape = 16,
             color = "grey90") +
  geom_point(aes(x = U1,
                 y = U2,
                 size = prcnt,
                 color = seurat_clusters),
             shape = 16) +
  theme_minimal() +
  theme(axis.title = element_blank(),
        axis.text = element_blank(),
        panel.grid = element_blank(),
        text = element_text(size = 15),
        legend.position = "none") +
  labs(color = NULL) +
  scale_color_manual(values = custom_colors("clusters")) +
  scale_size_continuous(range = c(3,10)) +
  guides(color = guide_legend(override.aes = list(alpha = 1, 
                                                  size = 5, 
                                                  shape = 19))) +
  geom_text(aes(x = U1, 
                y = U2, 
                label = prcnt), 
            color = "white", 
            size = 3, 
            fontface = "bold")

df <- list()
for (current_treatment in c("IgG", "Spt", "PD1")) {
  df[[current_treatment]] <- data.frame(matrix(NA, ncol = length(levels(factor(pan8$seurat_clusters))), nrow = length(levels(factor(pan8$seurat_clusters)))))
  colnames(df[[current_treatment]]) <- levels(factor(pan8$seurat_clusters))
  rownames(df[[current_treatment]]) <- levels(factor(pan8$seurat_clusters))
  for (i in 1:(length(levels(factor(pan8$seurat_clusters)))-1)) {
    for (j in (i+1):length(levels(factor(pan8$seurat_clusters))))
      df[[current_treatment]][i,j] <- seurat_integrated@meta.data %>% 
      filter(cd4_or_cd8 == "CD8" & treatment == current_treatment & !is.na(clonotype_id) & grepl(levels(factor(pan8$seurat_clusters))[i], clonotype_cluster_groups) & grepl(levels(factor(pan8$seurat_clusters))[j], clonotype_cluster_groups)) %>% 
        dplyr::select(clonotype_id) %>% 
        unique() %>% 
        count() %>% 
        pull(n)
  }
  df[[current_treatment]] <- df[[current_treatment]] %>% 
    rownames_to_column("seurat_clusters") %>%
    mutate(treatment = current_treatment) %>% 
    reshape2::melt(id.vars = c("seurat_clusters", "treatment"), na.rm = T)
}

merged_df <- rbind(df[["IgG"]], df[["Spt"]], df[["PD1"]]) %>% group_by(treatment) %>% mutate(prcnt = 100*value/sum(value))
merged_df$seurat_clusters <- factor(merged_df$seurat_clusters, levels =levels(factor(pan8$seurat_clusters)))
merged_df$variable <- factor(merged_df$variable, levels = rev(levels(factor(pan8$seurat_clusters))))
merged_df$treatment <- factor(merged_df$treatment, levels = c("IgG", "Spt", "PD1"))

merged_df %>% 
    ggplot(aes(x = seurat_clusters, y = variable, fill = prcnt)) + 
    geom_tile() +
  geom_text(aes(label = value)) +
    facet_wrap(~ treatment) +
    theme_minimal() + 
    theme(axis.text.x = element_text(angle = 90, hjust = 1),
          axis.title = element_blank()) +
    scale_fill_gradient2(low = "white", mid = "blue", high = "red", midpoint = 4, limits = c(0,8)) +
  scale_x_discrete(expand = c(0,0)) +
  scale_y_discrete(expand = c(0,0))

merged_df$group <- paste0("_", merged_df$seurat_clusters,"_&_", merged_df$variable, "_") #ensures CD8_eff doesn't match to both CD8_eff and CD8_effmem

#loop through treatments and clusters to build list of contingency tables - NEED TO FIX
res <- data.frame()
for (current_treatment in c("IgG", "Spt", "PD1")) {
  for (i in 1:(length(levels(factor(pan8$seurat_clusters)))-1)) {
    for (j in (i+1):length(levels(factor(pan8$seurat_clusters)))) {
      contingency_tbl <- data.frame(cluster = c(merged_df %>% 
                                                   filter(treatment == current_treatment & 
                                                            grepl(paste0("_", levels(factor(pan8$seurat_clusters))[i], "_"), group) & 
                                                            grepl(paste0("_", levels(factor(pan8$seurat_clusters))[j], "_"), group)) %>% 
                                                   summarize(total = sum(value)) %>% 
                                                   pull(total), 
                                                 merged_df %>% 
                                                   filter(treatment == current_treatment & 
                                                            grepl(paste0("_", levels(factor(pan8$seurat_clusters))[i], "_"), group) & 
                                                            !grepl(paste0("_", levels(factor(pan8$seurat_clusters))[j], "_"), group)) %>% 
                                                   summarize(total = sum(value)) %>% 
                                                   pull(total)),
                                     
                                     not_cluster = c(merged_df %>% 
                                                       filter(treatment == current_treatment & 
                                                                !grepl(paste0("_", levels(factor(pan8$seurat_clusters))[i], "_"), group) & 
                                                                grepl(paste0("_", levels(factor(pan8$seurat_clusters))[j], "_"), group)) %>% 
                                                       summarize(total = sum(value)) %>% 
                                                       pull(total), 
                                                     merged_df %>% 
                                                       filter(treatment == current_treatment & 
                                                                !grepl(paste0("_", levels(factor(pan8$seurat_clusters))[i], "_"), group) & 
                                                                !grepl(paste0("_", levels(factor(pan8$seurat_clusters))[j], "_"), group)) %>% 
                                                       summarize(total = sum(value)) %>% 
                                                       pull(total)),
                                     
                                     row.names = c("shared","not_shared"))
      res <- rbind(res, 
                   data.frame(treatment = current_treatment, 
                              seurat_clusters = levels(factor(pan8$seurat_clusters))[i], 
                              variable = levels(factor(pan8$seurat_clusters))[j],
                              fisher.p.value = fisher.test(contingency_tbl, alternative = "greater")$p.value))
    }
  }
}

res$fisher.p.value <- p.adjust(res$fisher.p.value, method = "BH")

res <- merge(merged_df, res, by = c("treatment", "seurat_clusters", "variable"))

res %>% 
    ggplot(aes(x = seurat_clusters, y = variable, fill = prcnt)) + 
    geom_tile() +
    geom_tile(data = res %>% filter(fisher.p.value < 0.05), fill = NA, color = "black", size = 1) +
    geom_text(aes(label = value), color = "black", size = 5) +
    facet_wrap(~ treatment) +
    theme_minimal() + 
    theme(axis.text.x = element_text(angle = 90, hjust = 1),
          axis.title = element_blank()) +
    scale_fill_gradient2(low = "white", mid = "blue", high = "red", midpoint = 4, limits = c(0,8))

```

Fisher's exact test with plot to look at sharing between clusters. Number of unique clonotypes in box and p values as color.
