==> scripts/scRNAseq/8_seurat_initialization_colitis.Rmd <==
---
title: "7_figure6_plots"
output: rmarkdown::github_document
date: '2022-07-10'
---


This code documents the initialization of the Seurat objects for patient colitis.

```{r setup, include=FALSE, echo=FALSE}
knitr::opts_chunk$set(echo = TRUE, root.dir = "C:/Users/Strix/GitRepos/NOD_spt_pd1")
knitr::opts_knit$set(echo = TRUE, root.dir = "C:/Users/Strix/GitRepos/NOD_spt_pd1")
```

```{r libraries, message = FALSE, warning = FALSE}
require(dplyr, quietly = T) #simple manipulation of data.frames
require(Seurat, quietly = T) #scRNA-seq analysis
require(reshape, quietly = T) #melting and casting data.frames
require(ggplot2, quietly = T) #plotting
require(rstatix, quietly = T) #adding p values to plots
require(ggprism, quietly = T) #adding p values to plots
require(ggsci, quietly = T) #IGV color palette
require(patchwork, quietly = T) #adding p values to plots
require(magrittr, quietly = T) #adding p values to plots
require(gridExtra, quietly = T) #display multiple plots
require(ggrepel, quietly = T) #labels for plots
require(tibble, quietly = T) #rownames_to_column function
require(useful, quietly = T) #corner function to peek at data.frames
require(viridis, quietly = T) #viridis color palettes
require(readxl, quietly = T) #reading Excel files
require(msigdbr, quietly = T) #accessing gene signatures from mSigDB
source("scripts/utils.R") #custom color palettes and other functions
set.seed(1) #set a seed for reproducibility
```

```{r get gene signature of exhaustion}
texh <- FindMarkers(seurat_integrated, 
                       ident.1 = "CD8_texh",
                      group.by = "seurat_clusters") %>% 
            filter(p_val_adj < 0.05) %>%
            arrange(desc(avg_log2FC))
term_exh <- read_excel("ref/signatures/Miller2019_NatureImmuno_SuppTable1.xlsx", sheet = "Terminally Exh.")$gene
common_to_all_treatments
```

Recode the pathology levels. Then plot the expression of co-inhibitory receptors:
```{r coinhibitory receptor expression}
cd8_merge <- merge(luoma_cd8, list(boland_cd8, corri))
new_metadata <- data.frame(coinh_count = rowSums(as.data.frame(t(as.data.frame(cd8_merge@assays$RNA@data[c("PDCD1","HAVCR2","CTLA4","LAG3","TIGIT","CD160"),]))) > 0))
cd8_merge <- AddMetaData(cd8_merge, new_metadata)
cd8_merge$pathology <- recode_factor(cd8_merge$pathology, 
                                        "Healthy Control" = "healthy",
                                        "Ulcerative Colitis" = "colitis",
                                        "ICI No Colitis" = "ICI",
                                        "ICI Colitis" = "ICI+colitis")

tcon_merge <- merge(luoma_tcon, boland_tcon)
new_metadata <- data.frame(coinh_count = rowSums(as.data.frame(t(as.data.frame(tcon_merge@assays$RNA@data[c("PDCD1","HAVCR2","CTLA4","LAG3","TIGIT","CD160"),]))) > 0))
tcon_merge <- AddMetaData(tcon_merge, new_metadata)
tcon_merge$pathology <- recode_factor(tcon_merge$pathology,
                                        "Healthy Control" = "healthy",
                                        "Ulcerative Colitis" = "colitis",
                                        "ICI No Colitis" = "ICI",
                                        "ICI Colitis" = "ICI+colitis")

p_cd8_coinh <- cd8_merge@meta.data %>% 
                  group_by(pathology, paper, sample) %>% 
                  summarize(avg = mean(coinh_count)) %>%
                  ggplot() + 
                  geom_violin(aes(x = pathology, 
                             fill = pathology, 
                             y = avg, 
                             color = pathology),
                             alpha = 0.5,
                              adjust = 1,
                              draw_quantiles = c(0.25,0.5,0.75),
                              scale = "width") + 
                  geom_point(aes(x = pathology, 
                             fill = pathology, 
                             y = avg, 
                             color = pathology),
                             size = 4,
                             position = position_jitter()) + 
                  theme_minimal() + 
                  theme(panel.spacing = unit(0,"lines"),
                        axis.title.x = element_blank(),
                        panel.grid = element_blank(),
                        panel.border = element_rect(color = "black", size = 0.5, fill = NA)) +
                  facet_wrap(~paper, scales = "free_x") + 
                  scale_fill_manual(values = custom_colors("colitis")) + 
                  scale_color_manual(values = custom_colors("colitis")) +
                  labs(y = "Average number of co-inhibitory receptors") +
                  add_pvalue(data.frame(cd8_merge@meta.data %>% 
                                            group_by(pathology, paper, sample) %>% 
                                            summarize(avg = mean(coinh_count)) %>% 
                                            ungroup() %>%
                                          group_by(paper) %>%
                                            t_test(avg ~ pathology) %>% 
                                            add_xy_position()),
                             tip.length = 0)

p_pan8_coinh <- pan8@meta.data %>% 
                  group_by(treatment, mouse_id) %>% 
                  summarize(avg = mean(coinh_count)) %>%
                  ggplot() + 
                  geom_violin(aes(x = treatment, 
                                  fill = treatment, 
                                  y = avg, 
                                  color = treatment),
                              alpha = 0.5,
                              adjust = 1,
                              draw_quantiles = c(0.25,0.5,0.75),
                              scale = "width") + 
                  geom_point(aes(x = treatment, 
                                 y = avg, 
                                 color = treatment),
                             size = 4,
                             position = position_jitter()) + 
                  theme_minimal() + 
                  theme(panel.spacing = unit(0,"lines"),
                        legend.position = "none",
                        axis.title.x = element_blank(),
                        panel.grid = element_blank(),
                        panel.border = element_rect(color = "black", size = 0.5, fill = NA)) +
                  scale_fill_manual(values = custom_colors("treatments")) + 
                  scale_color_manual(values = custom_colors("treatments")) +
                  labs(y = "Average number of co-inhibitory receptors", title = "CD8+ T cells") +
                  add_pvalue(pan8@meta.data %>% 
                               group_by(treatment, mouse_id) %>% 
                               summarize(avg = mean(coinh_count)) %>% 
                               ungroup() %>% 
                               t_test(avg ~ treatment) %>% 
                               add_xy_position(),
                             tip.length = 0)

p_tcon_coinh <- tcon_merge@meta.data %>% 
                  group_by(pathology, paper, sample) %>% 
                  summarize(avg = mean(coinh_count)) %>%
                  ggplot() + 
                  geom_violin(aes(x = pathology, 
                             fill = pathology, 
                             y = avg, 
                             color = pathology),
                             alpha = 0.5,
                              adjust = 1,
                              draw_quantiles = c(0.25,0.5,0.75),
                              scale = "width") + 
                  geom_point(aes(x = pathology, 
                             fill = pathology, 
                             y = avg, 
                             color = pathology),
                             size = 4,
                             position = position_jitter()) + 
                  theme_minimal() + 
                  theme(panel.spacing = unit(0,"lines"),
                        axis.title.x = element_blank(),
                        panel.grid = element_blank(),
                        panel.border = element_rect(color = "black", size = 0.5, fill = NA)) +
                  facet_wrap(~paper, scales = "free_x") + 
                  scale_fill_manual(values = custom_colors("colitis")) + 
                  scale_color_manual(values = custom_colors("colitis")) +
                  labs(y = "Average number of co-inhibitory receptors") +
                  add_pvalue(data.frame(tcon_merge@meta.data %>% 
                                            group_by(pathology, paper, sample) %>% 
                                            summarize(avg = mean(coinh_count)) %>% 
                                            ungroup() %>%
                                          group_by(paper) %>%
                                            t_test(avg ~ pathology) %>% 
                                            add_xy_position()),
                             tip.length = 0)

p_pan4_coinh <- pan4@meta.data %>% 
                  filter(celltype == "Tcon") %>%
                  group_by(treatment, mouse_id) %>% 
                  summarize(avg = mean(coinh_count)) %>%
                  ggplot() + 
                  geom_violin(aes(x = treatment, 
                                  fill = treatment, 
                                  y = avg, 
                                  color = treatment),
                              alpha = 0.5,
                              adjust = 1,
                              draw_quantiles = c(0.25,0.5,0.75),
                              scale = "width") + 
                  geom_point(aes(x = treatment, 
                                 y = avg, 
                                 color = treatment),
                             size = 4,
                             position = position_jitter()) + 
                  theme_minimal() + 
                  theme(panel.spacing = unit(0,"lines"),
                        legend.position = "none",
                        axis.title.x = element_blank(),
                        panel.grid = element_blank(),
                        panel.border = element_rect(color = "black", size = 0.5, fill = NA)) +
                  scale_fill_manual(values = custom_colors("treatments")) + 
                  scale_color_manual(values = custom_colors("treatments")) +
                  labs(y = "Average number of co-inhibitory receptors", title = "Tcon") +
                  add_pvalue(pan4@meta.data %>% 
                               filter(celltype == "Tcon") %>%
                               group_by(treatment, mouse_id) %>% 
                               summarize(avg = mean(coinh_count)) %>% 
                               ungroup() %>% 
                               t_test(avg ~ treatment) %>% 
                               add_xy_position(),
                             tip.length = 0)

grid.arrange(p_cd8_coinh, p_pan8_coinh, p_tcon_coinh, p_pan4_coinh, nrow = 2, widths = c(2,1))

```

```{r add Tcon data}
tcon_merge <- merge(luoma_tcon, boland_tcon)
prog_exh <- FindMarkers(pan4, ident.1 = "Tcon_prog", group.by="seurat_clusters", only.pos = T)%>%filter(p_val_adj<0.05)%>%arrange(desc(avg_log2FC))
prog_exh <- convertMouseGeneList(rownames(prog_exh))
tcon_merge <- AddModuleScore(tcon_merge, features = list(prog_exh), name = "tcon_prog", assay = "RNA")

p_miller_texh <- tcon_merge@meta.data %>% 
                  group_by(pathology, paper, sample) %>% 
                  summarize(avg = mean(tcon_prog1)) %>%
                  ggplot() + 
                  geom_violin(aes(x = pathology, 
                                  fill = pathology, 
                                  y = avg, 
                                  color = pathology),
                              alpha = 0.5,
                              adjust = 1,
                              draw_quantiles = c(0.25,0.5,0.75),
                              scale = "width") + 
                  geom_point(aes(x = pathology, 
                                 fill = pathology, 
                                 y = avg, 
                                 color = pathology),
                             size = 4,
                             position = position_jitter()) + 
                  theme_minimal() + 
                  theme(panel.spacing = unit(0,"lines"),
                        axis.title.x = element_blank(),
                        panel.grid = element_blank(),
                        panel.border = element_rect(color = "black", size = 0.5, fill = NA)) +
                  facet_wrap(~paper, scales = "free_x") + 
                  scale_fill_manual(values = custom_colors("colitis")) + 
                  scale_color_manual(values = custom_colors("colitis")) +
                  scale_y_continuous(limits = c(0, 0.3), expand = c(0,0)) +
                  labs(y = "Progenitor Tcon gene signature from NOD") +
                  add_pvalue(data.frame(tcon_merge@meta.data %>% 
                                            group_by(pathology, paper, sample) %>% 
                                            summarize(avg = mean(tcon_prog1)) %>% 
                                            ungroup() %>%
                                            group_by(paper) %>%
                                            t_test(avg ~ pathology) %>% 
                                            add_xy_position()),
                             tip.length = 0)

```

```{r add the exhaustion gene signature}
term_exh <- read_excel("ref/signatures/Miller2019_NatureImmuno_SuppTable1.xlsx", sheet = "Terminally Exh.")$gene
term_exh <- convertMouseGeneList(term_exh)
cd8_merge <- AddModuleScore(cd8_merge, features = list(term_exh), name = "Miller2019_terminally_exh", assay = "RNA")

p_miller_texh <- cd8_merge@meta.data %>% 
                  group_by(pathology, paper, sample) %>% 
                  summarize(avg = mean( Miller2019_terminally_exh1)) %>%
                  ggplot() + 
                  geom_violin(aes(x = pathology, 
                                  fill = pathology, 
                                  y = avg, 
                                  color = pathology),
                              alpha = 0.5,
                              adjust = 1,
                              draw_quantiles = c(0.25,0.5,0.75),
                              scale = "width") + 
                  geom_point(aes(x = pathology, 
                                 fill = pathology, 
                                 y = avg, 
                                 color = pathology),
                             size = 4,
                             position = position_jitter()) + 
                  theme_minimal() + 
                  theme(panel.spacing = unit(0,"lines"),
                        axis.title.x = element_blank(),
                        panel.grid = element_blank(),
                        panel.border = element_rect(color = "black", size = 0.5, fill = NA)) +
                  facet_wrap(~paper, scales = "free_x") + 
                  scale_fill_manual(values = custom_colors("colitis")) + 
                  scale_color_manual(values = custom_colors("colitis")) +
                  labs(y = "Terminally exhausted module score (Miller 2019)") +
                  add_pvalue(data.frame(cd8_merge@meta.data %>% 
                                            group_by(pathology, paper, sample) %>% 
                                            summarize(avg = mean( Miller2019_terminally_exh1)) %>% 
                                            ungroup() %>%
                                            group_by(paper) %>%
                                            t_test(avg ~ pathology) %>% 
                                            add_xy_position()),
                             tip.length = 0)

p_nod_texh <- cd8_merge@meta.data %>% 
                group_by(pathology, paper, sample) %>% 
                summarize(avg = mean(texh_signature1)) %>%
                ggplot() + 
                geom_violin(aes(x = pathology, 
                                fill = pathology, 
                                y = avg, 
                                color = pathology),
                            alpha = 0.5,
                            adjust = 1,
                            draw_quantiles = c(0.25,0.5,0.75),
                            scale = "width") + 
                geom_point(aes(x = pathology, 
                               fill = pathology, 
                               y = avg, 
                               color = pathology),
                           size = 4,
                           position = position_jitter()) + 
                theme_minimal() + 
                theme(panel.spacing = unit(0,"lines"),
                      axis.title.x = element_blank(),
                      panel.grid = element_blank(),
                      panel.border = element_rect(color = "black", size = 0.5, fill = NA)) +
                facet_wrap(~paper, scales = "free_x") + 
                scale_fill_manual(values = custom_colors("colitis")) + 
                scale_color_manual(values = custom_colors("colitis")) +
  scale_y_continuous(limits = c(0.15,0.4), breaks = c(0.2,0.3,0.4)) +
                labs(y = "Terminally exhausted/effector-like module score") +
                add_pvalue(data.frame(cd8_merge@meta.data %>% 
                                          group_by(pathology, paper, sample) %>% 
                                          summarize(avg = mean(texh_signature1)) %>% 
                                          ungroup() %>%
                                          group_by(paper) %>%
                                          t_test(avg ~ pathology) %>% 
                                          add_xy_position()),
                           tip.length = 0)

p_nod_eff <- cd8_merge@meta.data %>% 
                group_by(pathology, paper, sample) %>% 
                summarize(avg = mean(eff_signature1)) %>%
                ggplot() + 
                geom_violin(aes(x = pathology, 
                                fill = pathology, 
                                y = avg, 
                                color = pathology),
                            alpha = 0.5,
                            adjust = 1,
                            draw_quantiles = c(0.25,0.5,0.75),
                            scale = "width") + 
                geom_point(aes(x = pathology, 
                               fill = pathology, 
                               y = avg, 
                               color = pathology),
                           size = 4,
                           position = position_jitter()) + 
                theme_minimal() + 
                theme(panel.spacing = unit(0,"lines"),
                      axis.title.x = element_blank(),
                      panel.grid = element_blank(),
                      panel.border = element_rect(color = "black", size = 0.5, fill = NA)) +
                facet_wrap(~paper, scales = "free_x") + 
                scale_fill_manual(values = custom_colors("colitis")) + 
  scale_y_continuous(limits = c(0.15,0.4), breaks = c(0.2,0.3,0.4)) +
                scale_color_manual(values = custom_colors("colitis")) +
                labs(y = "Terminally exhausted/effector-like module score") +
                add_pvalue(data.frame(cd8_merge@meta.data %>% 
                                          group_by(pathology, paper, sample) %>% 
                                          summarize(avg = mean(eff_signature1)) %>% 
                                          ungroup() %>%
                                          group_by(paper) %>%
                                          t_test(avg ~ pathology) %>% 
                                          add_xy_position()),
                           tip.length = 0)

grid.arrange(p_cd8_coinh, p_nod_texh, p_nod_eff, nrow = 3)
```

```{r identify overlapping genes in exhaustion signatures}
length(setdiff(term_exh, PD1_signature))
length(setdiff(PD1_signature, term_exh))
length(intersect(PD1_signature, term_exh))

setdiff(term_exh, PD1_signature)
setdiff(PD1_signature, term_exh)
intersect(PD1_signature, term_exh)
```

```{r violin plot of Cd38 expression}
new_metadata <- data.frame(t(as.matrix(seurat_integrated@assays$RNA["Cd38",])))
seurat_integrated <- AddMetaData(seurat_integrated, new_metadata)

seurat_integrated@meta.data %>%
    group_by(treatment, mouse_id, tissue) %>%
    summarize(Cd38 = mean(Cd38)) %>%
    ggplot(aes(x = treatment, y = Cd38, fill = treatment, color = treatment)) +
    geom_violin(alpha = 0.5,
                adjust = 1,
                draw_quantiles = c(0.25,0.5,0.75),
                scale = "width") + 
    geom_point(size = 4,
               position = position_dodge(0.9)) + 
    theme_minimal() + 
    theme(panel.spacing = unit(0,"lines"),
          axis.title.x = element_blank(),
          panel.grid = element_blank(),
          panel.border = element_rect(color = "black", 
                                      size = 0.5, 
                                      fill = NA)) +
    scale_fill_manual(values = custom_colors("treatments")) + 
    scale_color_manual(values = custom_colors("treatments")) +
    labs(y = "Cd38 expression") +
    facet_wrap(~ tissue, scales = "free_x")

seurat_integrated@meta.data %>% 
  filter(tissue == "blood" & seurat_clusters != "Acinar_contaminated") %>%
    group_by(mouse_id, seurat_clusters) %>%
    summarize(Cd38 = mean(Cd38)) %>%
    ggplot(aes(y = seurat_clusters, x = Cd38, fill = seurat_clusters, color = seurat_clusters)) +
    geom_boxplot(alpha = 0.5,
                outlier.shape = NA) + 
    theme_minimal() + 
    theme(panel.spacing = unit(0,"lines"),
          axis.title.x = element_blank(),
          panel.grid = element_blank(),
          panel.border = element_rect(color = "black", 
                                      size = 0.5, 
                                      fill = NA)) +
    scale_fill_manual(values = custom_colors("clusters")) + 
    scale_color_manual(values = custom_colors("clusters")) +
    labs(x = "Cd38 expression", y = NULL)

seurat_integrated@meta.data %>%
    group_by(treatment, mouse_id, tissue) %>%
    summarize(Cd38 = mean(Cd38)) %>% group_by(tissue) %>%
 t_test(Cd38 ~ treatment)
  
```

```{r UMAPs of Cd38 expression in patient colitis}
new_metadata <- as.data.frame(cd8_merge@reductions$umap@cell.embeddings)
cd8_merge <- AddMetaData(cd8_merge, new_metadata)

new_metadata <- data.frame(t(as.matrix(cd8_merge@assays$RNA["CD38",])))
cd8_merge <- AddMetaData(cd8_merge, new_metadata)

p_umaps <- cd8_merge@meta.data %>% 
    ggplot(aes(x = UMAP_1, 
               y = UMAP_2, 
               color = CD38)) + 
    geom_point(shape = 16) +
    theme_minimal() +
    theme(axis.title = element_blank(),
          axis.text = element_blank(),
          panel.grid = element_blank(),
          text = element_text(size = 15),
          legend.position = "none",
          legend.key.height = unit(0.25,"cm")) +
    labs(color= NULL,
         title = NULL) +
    scale_color_gradient(low = "azure2", high = "blue4") +
  facet_wrap(paper ~ pathology) +
    guides(color = guide_colorbar(ticks = F)) #removes ticks from colorbar

cd8_merge@meta.data %>%
    group_by(treatment, orig.ident, paper) %>%
    summarize(Cd38 = mean(CD38)) %>%
    ggplot() +
    geom_violin(aes(x = treatment, y = Cd38, fill = treatment, color = treatment),
                alpha = 0.5,
                adjust = 1,
                draw_quantiles = c(0.25,0.5,0.75),
                scale = "width") + 
    geom_point(aes(x = treatment, y = Cd38, fill = treatment, color = treatment),
               size = 4,
               position = position_dodge(0.9)) + 
    theme_minimal() + 
    theme(panel.spacing = unit(0,"lines"),
          axis.title.x = element_blank(),
          panel.grid = element_blank(),
          panel.border = element_rect(color = "black", 
                                      size = 0.5, 
                                      fill = NA)) +
    scale_fill_manual(values = custom_colors("colitis")) + 
    scale_color_manual(values = custom_colors("colitis")) +
    labs(y = "Cd38 expression") +
    facet_wrap(~ paper, scales = "free_x") +
    add_pvalue(cd8_merge@meta.data %>%
                   group_by(treatment, orig.ident, paper) %>%
                   summarize(Cd38 = mean(CD38)) %>% 
                   group_by(paper) %>% 
                   t_test(Cd38 ~ treatment) %>%
                   add_xy_position(),
               tip.length = 0)

```

Remove acinar cells and get gene signatures. UC shows a shift from memory to effectorness wheras PD1 shows shift towards Texhaustion. Look at 2D (texh vs eff and teff vs mem):
```{r identify Cd8 gene signatures}
new_metadata <- pan8@meta.data %>% 
                  mutate(patient_signature = case_when(seurat_clusters %in% c("CD8_pexh","CD8_eff","CD8_mem") ~ "Spt_clusters",
                                                       seurat_clusters %in% c("CD8_texh","Proliferating") ~ "PD1_clusters",
                                                       TRUE ~ NA_character_)) %>% 
                  dplyr::select(patient_signature)

pan8 <- AddMetaData(pan8, new_metadata)

PD1_markers <- FindMarkers(pan8, 
                       ident.1 = "CD8_texh", 
                       group.by = "seurat_clusters", only.pos = T) %>% 
            filter(p_val_adj<0.05) %>%
            arrange(desc(avg_log2FC))
slec_markers <- FindMarkers(pan8, 
                       ident.1 = "CD8_slec", 
                       group.by = "seurat_clusters") %>% 
            filter(p_val_adj<0.05) %>%
            arrange(desc(avg_log2FC))
Spt_markers <- FindMarkers(pan8, 
                       ident.1 = "CD8_eff", 
                       group.by = "seurat_clusters") %>% 
            filter(p_val_adj<0.05) %>%
            arrange(desc(avg_log2FC))
mem_markers <- FindMarkers(pan8, 
                       ident.1 = "CD8_mem", 
                       group.by = "seurat_clusters") %>% 
            filter(p_val_adj<0.05) %>%
            arrange(desc(avg_log2FC))
cm_markers <- FindMarkers(pan8, 
                       ident.1 = "CD8_cm", 
                       group.by = "seurat_clusters") %>% 
            filter(p_val_adj<0.05) %>%
            arrange(desc(avg_log2FC))
eff_markers <- FindMarkers(pan8, 
                       ident.1 = "CD8_eff", 
                       group.by = "seurat_clusters") %>% 
            filter(p_val_adj<0.05) %>%
            arrange(desc(avg_log2FC))

#aim for around 150 genes
PD1_markers %>% filter(avg_log2FC>0) %>% count()

#genes are mouse, so convert them to human
PD1_signature <- convertMouseGeneList(PD1_markers %>% filter(avg_log2FC > 0) %>% rownames()) #convert the positive logFC genes to human
Spt_signature <- convertMouseGeneList(Spt_markers %>% filter(avg_log2FC > 0) %>% rownames()) #convert the positive logFC genes to human
mem_signature <- convertMouseGeneList(mem_markers %>% filter(avg_log2FC > 0) %>% rownames()) #convert the positive logFC genes to human
slec_signature <- convertMouseGeneList(slec_markers %>% filter(avg_log2FC > 0) %>% rownames()) #convert the positive logFC genes to human
cm_signature <- convertMouseGeneList(cm_markers %>% filter(avg_log2FC > 0) %>% rownames()) #convert the positive logFC genes to human
eff_signature <- convertMouseGeneList(eff_markers %>% filter(avg_log2FC > 0) %>% rownames()) #convert the positive logFC genes to human

luoma_cd8 <- AddModuleScore(luoma_cd8, features = list(PD1_signature), name = "CD8_PD1_signature")
luoma_cd8 <- AddModuleScore(luoma_cd8, features = list(mem_signature), name = "CD8_mem_signature")
luoma_cd8 <- AddModuleScore(luoma_cd8, features = list(cm_signature), name = "CD8_cm_signature")
boland_cd8 <- AddModuleScore(boland_cd8, features = list(PD1_signature), name = "CD8_PD1_signature")
boland_cd8 <- AddModuleScore(boland_cd8, features = list(mem_signature), name = "CD8_mem_signature")
boland_cd8 <- AddModuleScore(boland_cd8, features = list(cm_signature), name = "CD8_cm_signature")
luoma_cd8 <- AddModuleScore(luoma_cd8, features = list(Spt_signature), name = "CD8_Spt_signature")
boland_cd8 <- AddModuleScore(boland_cd8, features = list(Spt_signature), name = "CD8_Spt_signature")
corri <- AddModuleScore(corri, features = list(cm_signature), name = "CD8_cm_signature")
corri <- AddModuleScore(corri, features = list(PD1_signature), name = "CD8_PD1_signature")
corri <- AddModuleScore(corri, features = list(Spt_signature), name = "CD8_Spt_signature")
corri <- AddModuleScore(corri, features = list(mem_signature), name = "CD8_mem_signature")
```

```{r identify CD4 gene signatures}
PD1_markers <- FindMarkers(pan4, 
                       ident.1 = "Tcon_prog", 
                       group.by = "seurat_clusters") %>% 
            filter(p_val_adj<0.05) %>%
            arrange(desc(avg_log2FC))
Spt_markers <- FindMarkers(pan4, 
                       ident.1 = "Tcon_effmem", 
                       group.by = "seurat_clusters") %>% 
            filter(p_val_adj<0.05) %>%
            arrange(desc(avg_log2FC))

#aim for around 150 genes
PD1_markers %>% filter(avg_log2FC>0) %>% count()

#genes are mouse, so convert them to human
PD1_signature <- convertMouseGeneList(PD1_markers %>% filter(avg_log2FC > 0) %>% rownames()) #convert the positive logFC genes to human
Spt_signature <- convertMouseGeneList(Spt_markers %>% filter(avg_log2FC > 0) %>% rownames()) #convert the positive logFC genes to human

luoma_tcon <- AddModuleScore(luoma_tcon, features = list(PD1_signature), name = "CD4_PD1_signature")
boland_tcon <- AddModuleScore(boland_tcon, features = list(PD1_signature), name = "CD4_PD1_signature")
luoma_tcon <- AddModuleScore(luoma_tcon, features = list(Spt_signature), name = "CD4_Spt_signature")
boland_tcon <- AddModuleScore(boland_tcon, features = list(Spt_signature), name = "CD4_Spt_signature")
```

With the human genes associated with PD1 compared to Spt, project these gene signatures onto the patient colitis data.
```{r cd8 signature violins}
#temporary fix to adjust naming scheme
luoma_cd8$treatment <- factor(recode_factor(luoma_cd8$treatment, "CT" = "healthy", "NC" = "ICI", "C" = "ICI+colitis"), 
                              levels = c("healthy","ICI","ICI+colitis"))
corri$treatment <- factor(recode_factor(corri$treatment, "HC" = "healthy", "UC" = "colitis"), levels = c("healthy","colitis"))
boland_cd8$treatment <- factor(recode_factor(boland_cd8$treatment, "HC" = "healthy", "UC" = "colitis"), levels = c("healthy","colitis"))

stat.test <- data.frame(luoma_cd8@meta.data %>%
                            group_by(treatment,sample) %>%
                            summarize(CD8_PD1_signature1 = mean(CD8_PD1_signature1)) %>%
                            ungroup() %>%
                            t_test(CD8_PD1_signature1 ~ treatment) %>% 
                            add_xy_position())

p_luoma <- luoma_cd8@meta.data %>%
    group_by(treatment, sample) %>%
    summarize(CD8_PD1_signature1 = mean(CD8_PD1_signature1)) %>%
    ggplot(aes(x = treatment,
               y = CD8_PD1_signature1)) +
    geom_violin(aes(
               color = treatment,
               fill = treatment),
               scale = "width",
                adjust = 1,
                trim = F,
                alpha = 0.5,
                draw_quantiles = c(0.25,0.5,0.75),
                lwd = 1) +
    geom_point(aes(color = treatment,
               fill = treatment),
               size = 5,
               position = position_jitter(width = 0.4)) +
    theme_minimal() +
    theme(panel.grid = element_blank(),
          panel.border = element_rect(color = "black", 
                                      fill = NA),
          text = element_text(size = 15),
          legend.position = "right") +
    scale_y_continuous(limits = c(0.15, 0.45)) +
    scale_color_manual(values = custom_colors("colitis")) +
    scale_fill_manual(values = custom_colors("colitis")) +
    labs(title = "CD8 T cells: Luoma", 
         y = "PD1 gene signature", 
         x = NULL,
         color = NULL,
         fill = NULL) +
    add_pvalue(stat.test,
             label = "p.adj.signif",
             tip.length = 0,
             label.size = 8)

## CORRIDONI
stat.test <- data.frame(corri@meta.data %>%
                            group_by(treatment,sample) %>%
                            summarize(CD8_PD1_signature1 = mean(CD8_PD1_signature1)) %>%
                            ungroup() %>%
                            t_test(CD8_PD1_signature1 ~ treatment) %>% 
                            add_xy_position() %>%
                            add_significance())

p_corri <- corri@meta.data %>%
    group_by(treatment, sample) %>%
    summarize(CD8_PD1_signature1 = mean(CD8_PD1_signature1)) %>%
    ggplot(aes(x = treatment,
               y = CD8_PD1_signature1)) +
    geom_violin(aes(
               color = treatment,
               fill = treatment),
               scale = "width",
                adjust = 1,
                trim = F,
                alpha = 0.5,
                draw_quantiles = c(0.25,0.5,0.75),
                lwd = 1) +
    geom_point(aes(color = treatment,
               fill = treatment),
               size = 5,
               position = position_jitter(width = 0.4)) +
    theme_minimal() +
    theme(panel.grid = element_blank(),
          panel.border = element_rect(color = "black", 
                                      fill = NA),
          text = element_text(size = 15),
          legend.position = "none") +
    scale_y_continuous(limits = c(0.15, 0.45)) +
    scale_color_manual(values = custom_colors("colitis")) +
    scale_fill_manual(values = custom_colors("colitis")) +
    labs(title = "CD8 T cells: Corridoni", 
         y = "PD1 gene signature", 
         x = NULL,
         color = NULL,
         fill = NULL) +
    add_pvalue(stat.test,
             label = "p.signif", #t test is not adjusted because there are only two groups
             tip.length = 0,
             label.size = 8)

## BOLAND
stat.test <- data.frame(boland_cd8@meta.data %>%
                            group_by(treatment,sample) %>%
                            summarize(CD8_PD1_signature1 = mean(CD8_PD1_signature1)) %>%
                            ungroup() %>%
                            t_test(CD8_PD1_signature1 ~ treatment) %>% 
                            add_xy_position() %>%
                            add_significance())

p_boland <- boland_cd8@meta.data %>%
    group_by(treatment, sample) %>%
    summarize(CD8_PD1_signature1 = mean(CD8_PD1_signature1)) %>%
    ggplot(aes(x = treatment,
               y = CD8_PD1_signature1)) +
    geom_violin(aes(
               color = treatment,
               fill = treatment),
               scale = "width",
                adjust = 1,
                trim = F,
                alpha = 0.5,
                draw_quantiles = c(0.25,0.5,0.75),
                lwd = 1) +
    geom_point(aes(color = treatment,
               fill = treatment),
               size = 5,
               position = position_jitter(width = 0.4)) +
    theme_minimal() +
    theme(panel.grid = element_blank(),
          panel.border = element_rect(color = "black", 
                                      fill = NA),
          text = element_text(size = 15),
          legend.position = "none") +
    scale_y_continuous(limits = c(0.15, 0.45)) +
    scale_color_manual(values = custom_colors("colitis")) +
    scale_fill_manual(values = custom_colors("colitis")) +
    labs(title = "CD8 T cells: Boland", 
         y = "PD1 gene signature", 
         x = NULL,
         color = NULL,
         fill = NULL) +
    add_pvalue(stat.test,
             label = "p.signif",
             tip.length = 0,
             label.size = 8)

grid.arrange(p_corri, p_boland, p_luoma, nrow = 1, widths = c(1,1,2))
```

```{r cd8 signature violins}
#temporary fix to adjust naming scheme
luoma_cd8$treatment <- factor(recode_factor(luoma_cd8$treatment, "CT" = "healthy", "NC" = "ICI", "C" = "ICI+colitis"), 
                              levels = c("healthy","ICI","ICI+colitis"))
corri$treatment <- factor(recode_factor(corri$treatment, "HC" = "healthy", "UC" = "colitis"), levels = c("healthy","colitis"))
boland_cd8$treatment <- factor(recode_factor(boland_cd8$treatment, "HC" = "healthy", "UC" = "colitis"), levels = c("healthy","colitis"))

stat.test <- data.frame(luoma_cd8@meta.data %>%
                            group_by(treatment,sample) %>%
                            summarize(CD8_Spt_signature1 = mean(CD8_Spt_signature1)) %>%
                            ungroup() %>%
                            t_test(CD8_Spt_signature1 ~ treatment) %>% 
                            add_xy_position())

p_luoma <- luoma_cd8@meta.data %>%
    group_by(treatment, sample) %>%
    summarize(CD8_Spt_signature1 = mean(CD8_Spt_signature1)) %>%
    ggplot(aes(x = treatment,
               y = CD8_Spt_signature1)) +
    geom_violin(aes(
               color = treatment,
               fill = treatment),
               scale = "width",
                adjust = 1,
                trim = F,
                alpha = 0.5,
                draw_quantiles = c(0.25,0.5,0.75),
                lwd = 1) +
    geom_point(aes(color = treatment,
               fill = treatment),
               size = 5,
               position = position_jitter(width = 0.4)) +
    theme_minimal() +
    theme(panel.grid = element_blank(),
          panel.border = element_rect(color = "black", 
                                      fill = NA),
          text = element_text(size = 15),
          legend.position = "right") +
    scale_y_continuous(limits = c(0.15, 0.45)) +
    scale_color_manual(values = custom_colors("colitis")) +
    scale_fill_manual(values = custom_colors("colitis")) +
    labs(title = "CD8 T cells: Luoma", 
         y = "PD1 gene signature", 
         x = NULL,
         color = NULL,
         fill = NULL) +
    add_pvalue(stat.test,
             label = "p.adj.signif",
             tip.length = 0,
             label.size = 8)

## CORRIDONI
stat.test <- data.frame(corri@meta.data %>%
                            group_by(treatment,sample) %>%
                            summarize(CD8_Spt_signature1 = mean(CD8_Spt_signature1)) %>%
                            ungroup() %>%
                            t_test(CD8_Spt_signature1 ~ treatment) %>% 
                            add_xy_position() %>%
                            add_significance())

p_corri <- corri@meta.data %>%
    group_by(treatment, sample) %>%
    summarize(CD8_Spt_signature1 = mean(CD8_Spt_signature1)) %>%
    ggplot(aes(x = treatment,
               y = CD8_Spt_signature1)) +
    geom_violin(aes(
               color = treatment,
               fill = treatment),
               scale = "width",
                adjust = 1,
                trim = F,
                alpha = 0.5,
                draw_quantiles = c(0.25,0.5,0.75),
                lwd = 1) +
    geom_point(aes(color = treatment,
               fill = treatment),
               size = 5,
               position = position_jitter(width = 0.4)) +
    theme_minimal() +
    theme(panel.grid = element_blank(),
          panel.border = element_rect(color = "black", 
                                      fill = NA),
          text = element_text(size = 15),
          legend.position = "none") +
    scale_y_continuous(limits = c(0.15, 0.45)) +
    scale_color_manual(values = custom_colors("colitis")) +
    scale_fill_manual(values = custom_colors("colitis")) +
    labs(title = "CD8 T cells: Corridoni", 
         y = "PD1 gene signature", 
         x = NULL,
         color = NULL,
         fill = NULL) +
    add_pvalue(stat.test,
             label = "p.signif", #t test is not adjusted because there are only two groups
             tip.length = 0,
             label.size = 8)

## BOLAND
stat.test <- data.frame(boland_cd8@meta.data %>%
                            group_by(treatment,sample) %>%
                            summarize(CD8_Spt_signature1 = mean(CD8_Spt_signature1)) %>%
                            ungroup() %>%
                            t_test(CD8_Spt_signature1 ~ treatment) %>% 
                            add_xy_position() %>%
                            add_significance())

p_boland <- boland_cd8@meta.data %>%
    group_by(treatment, sample) %>%
    summarize(CD8_Spt_signature1 = mean(CD8_Spt_signature1)) %>%
    ggplot(aes(x = treatment,
               y = CD8_Spt_signature1)) +
    geom_violin(aes(
               color = treatment,
               fill = treatment),
               scale = "width",
                adjust = 1,
                trim = F,
                alpha = 0.5,
                draw_quantiles = c(0.25,0.5,0.75),
                lwd = 1) +
    geom_point(aes(color = treatment,
               fill = treatment),
               size = 5,
               position = position_jitter(width = 0.4)) +
    theme_minimal() +
    theme(panel.grid = element_blank(),
          panel.border = element_rect(color = "black", 
                                      fill = NA),
          text = element_text(size = 15),
          legend.position = "none") +
    scale_y_continuous(limits = c(0.15, 0.45)) +
    scale_color_manual(values = custom_colors("colitis")) +
    scale_fill_manual(values = custom_colors("colitis")) +
    labs(title = "CD8 T cells: Boland", 
         y = "PD1 gene signature", 
         x = NULL,
         color = NULL,
         fill = NULL) +
    add_pvalue(stat.test,
             label = "p.signif",
             tip.length = 0,
             label.size = 8)

grid.arrange(p_corri, p_boland, p_luoma, nrow = 1, widths = c(1,1,2))
```


```{r cd4 Tfh signature violins}
tfh_markers <- FindMarkers(pan4, 
                           ident.1 = "Tcon_Tfh", 
                           group.by = "seurat_clusters") %>% 
    filter(p_val_adj<0.05) %>%
    arrange(desc(avg_log2FC))
tfh_signature <- convertMouseGeneList(tfh_markers %>% filter(avg_log2FC > 0) %>% rownames()) #convert the positive logFC genes to human
luoma_tcon <- AddModuleScore(luoma_tcon, features = list(tfh_signature), name = "tfh_signature")
boland_tcon <- AddModuleScore(boland_tcon, features = list(tfh_signature), name = "tfh_signature")

#temporary fix to adjust naming scheme
luoma_tcon$treatment <- factor(recode_factor(luoma_tcon$treatment, "CT" = "healthy", "NC" = "ICI", "C" = "ICI+colitis"), 
                              levels = c("healthy","ICI","ICI+colitis"))
boland_tcon$treatment <- factor(recode_factor(boland_tcon$treatment, "HC" = "healthy", "UC" = "colitis"), levels = c("healthy","colitis"))

cd4_merge <- merge(luoma_tcon, boland_tcon)
cd4_merge$treatment <- factor(cd4_merge$treatment, 
                              levels = c("healthy","colitis","ICI","ICI+colitis"))

stat.test <- data.frame(cd4_merge@meta.data %>%
                            group_by(treatment,sample,paper) %>%
                            summarize(tfh_signature1 = mean(tfh_signature1)) %>%
                            group_by(paper) %>%
                            t_test(tfh_signature1 ~ treatment) %>% 
                            add_xy_position())

p_luoma <- cd4_merge@meta.data %>%
    group_by(treatment, sample, paper) %>%
    summarize(tfh_signature1 = mean(tfh_signature1)) %>%
    ggplot(aes(x = treatment,
               y = tfh_signature1)) +
    geom_violin(aes(
               color = treatment,
               fill = treatment),
               scale = "width",
                adjust = 1,
                trim = F,
                alpha = 0.5,
                draw_quantiles = c(0.25,0.5,0.75),
                lwd = 1) +
    geom_point(aes(color = treatment,
               fill = treatment),
               size = 5,
               position = position_jitter(width = 0.4)) +
    theme_minimal() +
    theme(panel.grid = element_blank(),
          panel.spacing = unit(0, "cm"),
          panel.border = element_rect(color = "black", 
                                      fill = NA),
          text = element_text(size = 15),
          legend.position = "none") +
    scale_y_continuous(limits = c(0, 0.125)) +
    scale_color_manual(values = custom_colors("colitis")) +
    scale_fill_manual(values = custom_colors("colitis")) +
    labs(title = "tcon T cells: Luoma", 
         y = "PD1 gene signature", 
         x = NULL,
         color = NULL,
         fill = NULL) +
    add_pvalue(stat.test,
             label = "p.adj.signif",
             tip.length = 0,
             label.size = 8) +
  facet_wrap(~paper,  scales = "free_x")

## BOLAND
stat.test <- data.frame(boland_tcon@meta.data %>%
                            group_by(treatment,sample) %>%
                            summarize(tfh_signature1 = mean(tfh_signature1)) %>%
                            ungroup() %>%
                            t_test(tfh_signature1 ~ treatment) %>% 
                            add_xy_position() %>%
                            add_significance())

p_boland <- boland_tcon@meta.data %>%
    group_by(treatment, sample) %>%
    summarize(tfh_signature1 = mean(tfh_signature1)) %>%
    ggplot(aes(x = treatment,
               y = tfh_signature1)) +
    geom_violin(aes(
               color = treatment,
               fill = treatment),
               scale = "width",
                adjust = 1,
                trim = F,
                alpha = 0.5,
                draw_quantiles = c(0.25,0.5,0.75),
                lwd = 1) +
    geom_point(aes(color = treatment,
               fill = treatment),
               size = 5,
               position = position_jitter(width = 0.4)) +
    theme_minimal() +
    theme(panel.grid = element_blank(),
          panel.border = element_rect(color = "black", 
                                      fill = NA),
          text = element_text(size = 15),
          legend.position = "none") +
    #scale_y_continuous(limits = c(0.15, 0.45)) +
    scale_color_manual(values = custom_colors("colitis")) +
    scale_fill_manual(values = custom_colors("colitis")) +
    labs(title = "tcon T cells: Boland", 
         y = "PD1 gene signature", 
         x = NULL,
         color = NULL,
         fill = NULL) +
    add_pvalue(stat.test,
             label = "p.signif",
             tip.length = 0,
             label.size = 8)

grid.arrange(p_boland, p_luoma, nrow = 1, widths = c(2,3))
```

```{r cd4 Spt signature violins}

stat.test <- data.frame(luoma_tcon@meta.data %>%
                            group_by(treatment,sample) %>%
                            summarize(CD4_Spt_signature1 = mean(CD4_Spt_signature1)) %>%
                            ungroup() %>%
                            t_test(CD4_Spt_signature1 ~ treatment) %>% 
                            add_xy_position())

p_luoma <- luoma_tcon@meta.data %>%
    group_by(treatment, sample) %>%
    summarize(CD4_Spt_signature1 = mean(CD4_Spt_signature1)) %>%
    ggplot(aes(x = treatment,
               y = CD4_Spt_signature1)) +
    geom_violin(aes(
               color = treatment,
               fill = treatment),
               scale = "width",
                adjust = 1,
                trim = F,
                alpha = 0.5,
                draw_quantiles = c(0.25,0.5,0.75),
                lwd = 1) +
    geom_point(aes(color = treatment,
               fill = treatment),
               size = 5,
               position = position_jitter(width = 0.4)) +
    theme_minimal() +
    theme(panel.grid = element_blank(),
          panel.border = element_rect(color = "black", 
                                      fill = NA),
          text = element_text(size = 15),
          legend.position = "right") +
    #scale_y_continuous(limits = c(0.15, 0.45)) +
    scale_color_manual(values = custom_colors("colitis")) +
    scale_fill_manual(values = custom_colors("colitis")) +
    labs(title = "tcon T cells: Luoma", 
         y = "Spt gene signature", 
         x = NULL,
         color = NULL,
         fill = NULL) +
    add_pvalue(stat.test,
             label = "p.adj.signif",
             tip.length = 0,
             label.size = 8)

## BOLAND
stat.test <- data.frame(boland_tcon@meta.data %>%
                            group_by(treatment,sample) %>%
                            summarize(CD4_Spt_signature1 = mean(CD4_Spt_signature1)) %>%
                            ungroup() %>%
                            t_test(CD4_Spt_signature1 ~ treatment) %>% 
                            add_xy_position() %>%
                            add_significance())

p_boland <- boland_tcon@meta.data %>%
    group_by(treatment, sample) %>%
    summarize(CD4_Spt_signature1 = mean(CD4_Spt_signature1)) %>%
    ggplot(aes(x = treatment,
               y = CD4_Spt_signature1)) +
    geom_violin(aes(
               color = treatment,
               fill = treatment),
               scale = "width",
                adjust = 1,
                trim = F,
                alpha = 0.5,
                draw_quantiles = c(0.25,0.5,0.75),
                lwd = 1) +
    geom_point(aes(color = treatment,
               fill = treatment),
               size = 5,
               position = position_jitter(width = 0.4)) +
    theme_minimal() +
    theme(panel.grid = element_blank(),
          panel.border = element_rect(color = "black", 
                                      fill = NA),
          text = element_text(size = 15),
          legend.position = "none") +
    #scale_y_continuous(limits = c(0.15, 0.45)) +
    scale_color_manual(values = custom_colors("colitis")) +
    scale_fill_manual(values = custom_colors("colitis")) +
    labs(title = "tcon T cells: Boland", 
         y = "Spt gene signature", 
         x = NULL,
         color = NULL,
         fill = NULL) +
    add_pvalue(stat.test,
             label = "p.signif",
             tip.length = 0,
             label.size = 8)

grid.arrange(p_boland, p_luoma, nrow = 1, widths = c(1,2))
```

Create 2D plots showing the different gene signatures across the treatment groups:
```{r generated 2D plots}
medians_2d <- boland_cd8@meta.data %>% 
  group_by(celltype,paper,pathology,sample) %>% 
  summarize(PD1_score_median = median(CD8_PD1_signature1), Spt_score_median = median(CD8_Spt_signature1))

ggplot(boland_cd8@meta.data,
       aes(x=PD1_signature1, y=Spt_signature1, group = pathology)) +
    stat_density_2d(geom = "polygon", 
                    aes(alpha = ..level.., 
                        fill = pathology), 
                    bins = 6) +
    theme(panel.background = element_blank(),
          strip.background = element_blank(),
          panel.spacing.x = unit(0, "lines"),
          text = element_text(size=20),
          legend.title = element_blank(),
          panel.border = element_rect(colour = "black", fill=NA, size=1),
          legend.position = "top") +
    scale_fill_manual(values = custom_colors("colitis"), name = NULL) +
    scale_color_manual(values = custom_colors("colitis"), name = NULL) +
    scale_alpha_binned(range = c(0.2,0.9)) +
    labs(y = "Spt ModuleScore", 
         x = "PD1 ModuleScore") + 
    facet_grid(cols = vars(paper), 
               scales="free") +
    guides(alpha = "none")  +
    geom_point(data = medians_2d %>% filter(celltype == "CD8"), aes(x=PD1_score_median,y=Spt_score_median, color=pathology), size=4, alpha = 1)
```

Association with clonal expansion:
```{r clonal expansion}


texh <- FindMarkers(pan8, 
                       ident.1 = "CD8_texh",
                      group.by = "seurat_clusters") %>% 
            filter(p_val_adj < 0.05 & avg_log2FC > 0) %>%
            arrange(desc(avg_log2FC)) %>%
            top_n(150, wt = avg_log2FC)
ifn <- FindMarkers(pan8, 
                       ident.1 = "Interferon_sensing",
                      group.by = "seurat_clusters") %>% 
            filter(p_val_adj < 0.05 & avg_log2FC > 0) %>%
            arrange(desc(avg_log2FC)) %>%
            top_n(150, wt = avg_log2FC)
mem <- FindMarkers(pan8, 
                       ident.1 = "CD8_mem",
                      group.by = "seurat_clusters") %>% 
            filter(p_val_adj < 0.05 & avg_log2FC > 0) %>%
            arrange(desc(avg_log2FC)) %>%
            top_n(150, wt = avg_log2FC)
effmem <- FindMarkers(pan8, 
                       ident.1 = "CD8_effmem",
                      group.by = "seurat_clusters") %>% 
            filter(p_val_adj < 0.05 & avg_log2FC > 0) %>%
            arrange(desc(avg_log2FC)) %>%
            top_n(150, wt = avg_log2FC)
eff <- FindMarkers(pan8, 
                       ident.1 = "CD8_eff",
                      group.by = "seurat_clusters") %>% 
            filter(p_val_adj < 0.05 & avg_log2FC > 0) %>%
            arrange(desc(avg_log2FC)) %>%
            top_n(150, wt = avg_log2FC)
cm <- FindMarkers(pan8, 
                       ident.1 = "CD8_cm",
                      group.by = "seurat_clusters") %>% 
            filter(p_val_adj < 0.05 & avg_log2FC > 0) %>%
            arrange(desc(avg_log2FC)) %>%
            top_n(150, wt = avg_log2FC)
slec <- FindMarkers(pan8, 
                       ident.1 = "CD8_slec",
                      group.by = "seurat_clusters") %>% 
            filter(p_val_adj < 0.05 & avg_log2FC > 0) %>%
            arrange(desc(avg_log2FC)) %>%
            top_n(150, wt = avg_log2FC)
prolif <- FindMarkers(pan8, 
                       ident.1 = "Proliferating",
                      group.by = "seurat_clusters") %>% 
            filter(p_val_adj < 0.05 & avg_log2FC > 0) %>%
            arrange(desc(avg_log2FC)) %>%
            top_n(150, wt = avg_log2FC)

texh_signature <- convertMouseGeneList(rownames(texh)) #convert the positive logFC genes to human
cm_signature <- convertMouseGeneList( rownames(cm)) #convert the positive logFC genes to human
ifn_signature <- convertMouseGeneList(rownames(ifn)) #convert the positive logFC genes to human
mem_signature <- convertMouseGeneList(rownames(mem)) #convert the positive logFC genes to human
prolif_signature <- convertMouseGeneList(rownames(prolif)) #convert the positive logFC genes to human
eff_signature <- convertMouseGeneList(rownames(eff)) #convert the positive logFC genes to human
slec_signature <- convertMouseGeneList(rownames(slec)) #convert the positive logFC genes to human

seurat_merge <- merge(boland_cd8, list(corri, luoma_cd8))
seurat_merge$treatment <- factor(recode_factor(seurat_merge$treatment, "CT" = "healthy", "NC" = "ICI", "C" = "ICI+colitis", "HC" = "healthy", "UC" = "colitis"), 
                              levels = c("healthy","ICI","ICI+colitis", "colitis"))

seurat_merge <- AddModuleScore(seurat_merge, features = list(cm_signature), name = "cm_signature", assay = "RNA")
seurat_merge <- AddModuleScore(seurat_merge, features = list(texh_signature), name = "texh_signature", assay = "RNA")
seurat_merge <- AddModuleScore(seurat_merge, features = list(ifn_signature), name = "ifn_signature", assay = "RNA")
seurat_merge <- AddModuleScore(seurat_merge, features = list(prolif_signature), name = "prolif_signature", assay = "RNA")
seurat_merge <- AddModuleScore(seurat_merge, features = list(mem_signature), name = "mem_signature", assay = "RNA")
seurat_merge <- AddModuleScore(seurat_merge, features = list(eff_signature), name = "eff_signature", assay = "RNA")
seurat_merge <- AddModuleScore(seurat_merge, features = list(slec_signature), name = "slec_signature", assay = "RNA")

pan8 <- AddModuleScore(pan8, features = list(rownames(cm)), name = "cm_signature", assay = "RNA")
pan8 <- AddModuleScore(pan8, features = list(rownames(texh)), name = "texh_signature", assay = "RNA")
pan8 <- AddModuleScore(pan8, features = list(rownames(ifn)), name = "ifn_signature", assay = "RNA")
pan8 <- AddModuleScore(pan8, features = list(rownames(prolif)), name = "prolif_signature", assay = "RNA")
pan8 <- AddModuleScore(pan8, features = list(rownames(mem)), name = "mem_signature", assay = "RNA")
pan8 <- AddModuleScore(pan8, features = list(rownames(eff)), name = "eff_signature", assay = "RNA")
pan8 <- AddModuleScore(pan8, features = list(rownames(slec)), name = "slec_signature", assay = "RNA")

new_metadata <- seurat_merge@meta.data %>% 
    rownames_to_column('barcode') %>%
    filter(!is.na(raw_clonotype_id)) %>%
    group_by(orig.ident,raw_clonotype_id) %>% 
    mutate(clonotype_count = n()) %>% 
    arrange(desc(clonotype_count)) %>% 
    ungroup() %>% 
    arrange(desc(clonotype_count)) %>% 
    mutate(clonotype_id = paste0('clonotype',match(paste(orig.ident,raw_clonotype_id),unique(paste(orig.ident,raw_clonotype_id))))) %>%
    mutate(clonotype_count_bin = cut(clonotype_count, 
                                     breaks=c(0,1,3,20,170), labels=c("1","2-3","4-20","21-170"))) %>%
    column_to_rownames('barcode') %>%
    dplyr::select(clonotype_id,clonotype_count,clonotype_count_bin)

seurat_merge <- AddMetaData(seurat_merge, metadata = new_metadata)

medians_2d <- seurat_merge@meta.data %>% 
    filter(!is.na(clonotype_count_bin)) %>%
    group_by(celltype, paper, treatment, sample, clonotype_count_bin) %>% 
    summarize(PD1_score_median = mean(texh_signature1), mem_score_median = mean(eff_signature1))

ggplot(seurat_merge@meta.data %>% 
           filter(!is.na(clonotype_id)),
       aes(x=texh_signature1, 
           y=eff_signature1, 
           group = treatment)) +
    stat_density_2d(geom = "polygon", 
                    aes(alpha = ..level.., 
                        fill = treatment), 
                    bins = 6) +
    theme(panel.background = element_blank(),
          panel.border = element_rect(color = "black", fill = NA),
          panel.spacing = unit(0, "lines"),
          strip.background = element_blank()) +
    scale_fill_manual(values = custom_colors("colitis")) +
    scale_color_manual(values = custom_colors("colitis")) +
    labs(y = "Memory ModuleScore", 
         x = "Texh/PD1 ModuleScore") + 
    facet_wrap(paper ~ clonotype_count_bin) +
    guides(alpha = "none") +
    geom_point(data = medians_2d, aes(x = PD1_score_median, y = mem_score_median, color = treatment),
               size = 5)

#mouse next
medians_2d <- pan8@meta.data %>% 
   filter(!is.na(pan_clonotype_count) & celltype == "CD8") %>%
  group_by(mouse_id, treatment, pan_clonotype_count_bin) %>% 
  summarize(PD1_score_median = median(texh_signature1), mem_score_median = median(cm_signature1))

ggplot() +
  stat_density_2d(data = pan8@meta.data %>% 
                  filter(!is.na(pan_clonotype_count) & celltype == "CD8" & treatment == "IgG") %>%
                  dplyr::select(-treatment),
                    geom = "polygon", 
                    aes(alpha = ..level..,
                        x=texh_signature1, 
                        y=cm_signature1), 
                    bins = 6, fill = "black") +
    geom_point(data = medians_2d %>% filter(treatment == "IgG") %>% ungroup() %>% dplyr::select(-treatment), 
               aes(x = PD1_score_median, y = mem_score_median),
               size = 5) +
    stat_density_2d(data = pan8@meta.data %>% 
         filter(!is.na(pan_clonotype_count) & celltype == "CD8" & treatment != "IgG"),
                    geom = "polygon", 
                    aes(alpha = ..level.., 
                        fill = treatment,
                        x=texh_signature1, 
                        y=cm_signature1), 
                    bins = 6) +
    theme(panel.background = element_blank(),
          panel.border = element_rect(color = "black", fill = NA),
          panel.spacing = unit(0, "lines"),
          strip.background = element_blank()) +
    scale_fill_manual(values = custom_colors("treatments")) +
  scale_color_manual(values = custom_colors("treatments")) +
    labs(y = "Memory ModuleScore", 
         x = "Texh/PD1 ModuleScore") + 
    facet_wrap(treatment ~ pan_clonotype_count_bin, ncol = 5) +
    guides(alpha = "none") +
  scale_alpha_continuous(range = c(0.1,0.5)) +
  geom_point(data = medians_2d %>% filter(treatment != "IgG"), aes(x = PD1_score_median, y = mem_score_median, color = treatment))
```

CT = healthy controls; NC = CPI + no colitis; C = CPI + colitis
```{r Luoma: Initialization of Seurat object}
experiments <- list.files("data/colitis/luoma/CD3") 
experiments

data <- list() # List containing GEX and FB data
luomas <- list() # List containing all the Seurat objects

# Loop through each experiment in experiments
for (experiment in experiments) {
  # Read 10X data
  data[[experiment]] <- Read10X(data.dir = paste0("data/colitis/luoma/CD3/", experiment), strip.suffix = F) 
  # Create Seurat object
  luomas[[experiment]] <- CreateSeuratObject(counts = data[[experiment]], 
                                       project = experiment, min.cells = 50, min.features = 200) 
  
  ## Add meta data to Seurat object
  luomas[[experiment]]@meta.data$sample <- unlist(strsplit(experiment,"-"))[1] # All experiment information
  luomas[[experiment]]@meta.data$treatment <- gsub("[0-9]+", "", luomas[[experiment]]@meta.data$sample)
  luomas[[experiment]]@meta.data$paper <- "luoma"
  
  # Rename cells to add suffix corresponding to sample name
  luomas[[experiment]] <- RenameCells(luomas[[experiment]], 
                                              new.names = sapply(rownames(luomas[[experiment]]@meta.data),  function(x) sub(pattern="1", replacement = unlist(strsplit(experiment,"-"))[1], x)))
}

# Merge all Seurat objects generating unique cell IDs
luoma <- merge(x = luomas[[experiments[1]]], y = luomas[2:length(luomas)])
luoma

# Add pathology depending on treatment group
new_metadata <- luoma@meta.data %>% dplyr::select(treatment)
new_metadata$pathology <- "healthy"
new_metadata[which(new_metadata$treatment == "C"), "pathology"] <- "ICI Colitis"
new_metadata[which(new_metadata$treatment == "NC"), "pathology"] <- "ICI No Colitis"
new_metadata <- new_metadata %>% dplyr::select(pathology)

luoma <- AddMetaData(luoma, metadata = new_metadata)

```
